### 2022-03-19 #4253 @editorialbot
[PRE REVIEW]: A Modularised Java library for 3D Euclidean geometry
**Submitting author:** @agdturner (<a href="http://orcid.org/0000-0002-6098-6313">Andy Turner</a>)
**Repository:** https://github.com/agdturner/ccg-v3d
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
