### 2022-03-21 19d6860
Update paper.md
Fixed some typos, made dependencies clearer and some other general language improvements and minor clarifications.
