+++ paper.md
@@ -19,21 +19,21 @@ bibliography: paper.bib
 
 # Summary
 
-A three-dimensional ([3D](https://en.wikipedia.org/wiki/Euclidean_space)) [Euclidean geometry](https://en.wikipedia.org/wiki/Euclidean_geometry) Java library for modelling 3D changes in geometries over time. The library is modularised, based on [openJDK 17](https://openjdk.java.net/projects/jdk/17/), and the only additional dependency is [ccg-math](https://github.com/agdturner/ccg-math) - a stand alone modularised mathematics library.
+A three-dimensional ([3D](https://en.wikipedia.org/wiki/Euclidean_space)) [Euclidean geometry](https://en.wikipedia.org/wiki/Euclidean_geometry) Java library for modelling 3D changes in geometries over time. The library is modularised, based on [openJDK 17](https://openjdk.java.net/projects/jdk/17/), and dependends on [ccg-math](https://mvnrepository.com/artifact/io.github.agdturner/ccg-math) which in turn depends on both [big-math](https://mvnrepository.com/artifact/ch.obermuhlner/big-math) and [ccg-io](https://mvnrepository.com/artifact/io.github.agdturner/ccg-io).
 
-Point locations in space are defined using 3D [cartesian](https://en.wikipedia.org/wiki/Cartesian_coordinate_system) coordinates with [orthogonal](https://en.wikipedia.org/wiki/Orthogonality) X, Y and Z axes that meet at the origin - a point <x,y,z> where x=y=z=0. The coordinate system is "right handed", so if X increases to the right of this page and Y increases towards the top of this page, then Z increases out from the page, (see [orientation](https://en.wikipedia.org/wiki/Orientation_(vector_space)) for details of handedness, essentially there are two choices - a "left handed" system would have the Z direction of increase reversed). Right handed is more commonly used in both physics and geography.
+Point locations in space are defined using 3D [cartesian](https://en.wikipedia.org/wiki/Cartesian_coordinate_system) coordinates with [orthogonal](https://en.wikipedia.org/wiki/Orthogonality) X, Y and Z axes that meet at the origin - a point <x,y,z> where x=y=z=0. The coordinate system is "right handed": so if X increases to the right of this page, and Y increases towards the top of this page, then Z increases out from the page, (see [orientation](https://en.wikipedia.org/wiki/Orientation_(vector_space)) for details of handedness, essentially there are two choices - a "left handed" system would have the Z direction of increase reversed). Right handed is thought to be more commonly used in both physics and geography.
 
 Coordinates are either stored as [Math_BigRational](https://github.com/agdturner/ccg-math/blob/master/src/main/java/uk/ac/leeds/ccg/math/number/Math_BigRational.java) numbers - a subset of [rational numbers](https://en.wikipedia.org/wiki/Rational_number) with java.math.BigDecimal numerators and denominators, or [Math_BigRationalSqrt](https://github.com/agdturner/ccg-math/blob/master/src/main/java/uk/ac/leeds/ccg/math/number/Math_BigRationalSqrt.java) - numbers that also represent square roots (calculated to a given [Order of Magnitude](https://en.wikipedia.org/wiki/Order_of_magnitude) precision as required).
 
[…]
