### 2022-05-12 #4283 @agdturner
Hi @VivianePons. @behollister has identified an issue to do with getting started with the library using Maven. There was a slight confusion, but I believe I have addressed the issue. The library is organised along Maven guidelines which place all the test classes in a similar but separate directory structure to the source code. @behollister had suggested to reorganise this along the lines of a standard Eclipse project which I don't think is necessary. Hopefully loading the project as a Maven project in Eclipse has made things easier. I have not used Eclipse for a number of years preferring Apache NetBeans. I assume it is just as easy to open a Maven project in Eclipse as it is in Netbeans so that all the unit tests are easy to run. Since submitting to JOSS I have found a further Use Case scenario for the library. It is a physics problem which I hope to work on over the summer. This could help strengthen the statement of need - especially if I succeed in using the library to solve the problem that Joe Barker of @stonerlab set me which is along the following lines: Toon Verstraelen (https://scicomp.stackexchange.com/users/156/toon-verstraelen), Minimum image convention for triclinic unit cell, URL (version: 2012-09-01): https://scicomp.stackexchange.com/q/3107. I can provide further details of this if wanted, but essentially there should be a general solution that is more efficient than a brute force search in order to find the minimum distance from one point to a regular set of other points in some kind of parallelepiped tessellation as is common in crystalline/metallic materials... Thanks everyone for your efforts with this review.
