### 2022-05-25 agdturner/ccg-v3d#2 @behollister
Build Errors w/ Eclipse IDE (Maven Project)
Proceeding with build instructions for a Maven Project. Getting the build errors shown here in Eclipse IDE:
![errors](https://user-images.githubusercontent.com/20388050/170152978-88971cb9-9189-4cb2-a9f1-d333e15994fd.png)
Related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4283
