### 2022-05-25 #4283 @behollister
Build issue using Maven project via Eclipse IDE: https://github.com/agdturner/ccg-v3d/issues/2
