### 2022-05-25 agdturner/ccg-v3d#2 @agdturner
I am not familiar with Eclipse recently. I presume you have loaded the
project as a Maven project. What are the issues at the top level? - Why is
there a red cross? Have you tried compiling in src/main/java. I am not sure
why the src folder is added in the top level, did Eclipse do this, or did
you decide that was what should be done?

On Wed, May 25, 2022 at 1:26 AM Dr. Brad Hollister ***@***.***>
wrote:

> Proceeding with build instructions for a Maven Project. Getting the build
> errors shown here in Eclipse IDE:
> [image: errors]
> <https://user-images.githubusercontent.com/20388050/170152978-88971cb9-9189-4cb2-a9f1-d333e15994fd.png>
> Related to JOSS review: openjournals/joss-reviews#4283
> <https://github.com/openjournals/joss-reviews/issues/4283>
>
