### 2022-05-25 agdturner/ccg-v3d#2 @agdturner
The project opens fine for me in Eclipse:
![image](https://user-images.githubusercontent.com/5878594/170243220-3212ac1b-443f-4fc6-81a3-ee6cd5245cf0.png)
