### 2022-05-25 agdturner/ccg-v3d#2 @agdturner
I used the latest Eclipse download for this and the following procedure from https://vaadin.com/learn/tutorials/import-maven-project-eclipse: Import the project:
- In Eclipse, select File > Import… ​
- In the import window, expand Maven, select Existing Maven Projects, and click Next
- Click Browse and select the directory that contains the pom.xml file of the project you want to import
- Click Finish
