### 2022-05-25 agdturner/ccg-v3d#2 @behollister
The "[ERROR] symbol: class Math_Quaternion_BigRational" is likely a path issue for finding the library definition of Math_Quanternion_BigRational.

Which library / package is that from exactly? I'm not that familiar with Maven, but it appears it is downloading and installing dependencies.

Is it possible that the configuration (POM.xml) is missing a dependency, but that your dev platform has it?

I'm using the process you mention. After invoking Run As -> Maven Install (from the project explorer), I get the following. See COMPILATION ERROR below.

[INFO] Scanning for projects...
[INFO]
[INFO] --------------------< io.github.agdturner:ccg-v3d >---------------------
[INFO] Building ccg-v3d 0.15-SNAPSHOT
[INFO] --------------------------------[ jar ]---------------------------------
[INFO] Downloading from : https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-surefire-plugin/3.0.0-M4/maven-surefire-plugin-3.0.0-M4.pom
[INFO] Downloaded from : https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-surefire-plugin/3.0.0-M4/maven-surefire-plugin-3.0.0-M4.pom (8.1 kB at 4.4 kB/s)
[INFO] Downloading from : https://repo.maven.apache.org/maven2/org/apache/maven/surefire/surefire/3.0.0-M4/surefire-3.0.0-M4.pom
[INFO] Downloaded from : https://repo.maven.apache.org/maven2/org/apache/maven/surefire/surefire/3.0.0-M4/surefire-3.0.0-M4.pom (27 kB at 130 kB/s)
[INFO] Downloading from : https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-jar-plugin/2.4/maven-jar-plugin-2.4.pom
[INFO] Downloaded from : https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-jar-plugin/2.4/maven-jar-plugin-2.4.pom (5.8 kB at 35 kB/s)
[INFO] Downloading from : https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/22/maven-plugins-22.pom
[INFO] Downloaded from : https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-plugins/22/maven-plugins-22.pom (13 kB at 58 kB/s)
[INFO] Downloading from : https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-install-plugin/2.4/maven-install-plugin-2.4.pom
[INFO] Downloaded from : https://repo.maven.apache.org/maven2/org/apache/maven/plugins/maven-install-plugin/2.4/maven-install-plugin-2.4.pom (6.4 kB at 24 kB/s)
[INFO]
[INFO] --- maven-resources-plugin:2.6:resources (default-resources) @ ccg-v3d ---
[INFO] Using 'UTF-8' encoding to copy filtered resources.
[INFO] skip non existing resourceDirectory /home/brad/develop/JOSS-review/ccg-v3d/src/main/resources
[INFO]
[INFO] --- maven-compiler-plugin:3.8.1:compile (default-compile) @ ccg-v3d ---
[INFO] Changes detected - recompiling the module!
[INFO] Compiling 27 source files to /home/brad/develop/JOSS-review/ccg-v3d/target/classes
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/V3D_VGeometry.java]]
[parsing completed 25ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Points.java]]
[parsing completed 14ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedron.java]]
[parsing completed 43ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_IntersectionAndDistanceCalculations.java]]
[parsing completed 5ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Point.java]]
[parsing completed 47ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_LineSegment.java]]
[parsing completed 53ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/module-info.java]]
[parsing completed 3ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/V3D_VTriangle.java]]
[parsing completed 2ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Envelope.java]]
[parsing completed 68ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/V3D_VTetrahedron.java]]
[parsing completed 2ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Face.java]]
[parsing completed 1ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Ray.java]]
[parsing completed 14ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Line.java]]
[parsing completed 24ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/V3D_VLine.java]]
[parsing completed 2ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_TrianglesCoplanar.java]]
[parsing completed 6ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/V3D_V.java]]
[parsing completed 3ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Plane.java]]
[parsing completed 19ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/V3D_VPoint.java]]
[parsing completed 3ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Rectangle.java]]
[parsing completed 10ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_FiniteGeometry.java]]
[parsing completed 0ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java]]
[parsing completed 7ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedrons.java]]
[parsing completed 4ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_LineSegmentsCollinear.java]]
[parsing completed 7ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Geometry.java]]
[parsing completed 2ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/core/V3D_Environment.java]]
[parsing completed 2ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java]]
[parsing completed 38ms]
[parsing started SimpleFileObject[/home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Volume.java]]
[parsing completed 1ms]
[loading /home/brad/.m2/repository/io/github/agdturner/ccg-math/2.0.10/ccg-math-2.0.10.jar(/module-info.class)]
[loading /home/brad/.m2/repository/io/github/agdturner/ccg-io/1.1.0/ccg-io-1.1.0.jar(/module-info.class)]
[loading /modules/java.base/module-info.class]
[loading /modules/java.desktop/module-info.class]
[loading /modules/java.xml/module-info.class]
[loading /modules/java.datatransfer/module-info.class]
[loading /modules/java.prefs/module-info.class]
[loading /modules/java.logging/module-info.class]
[loading /modules/java.base/java/io/Serializable.class]
[loading /modules/java.base/java/lang/Object.class]
[loading /modules/java.base/java/util/Objects.class]
[loading /home/brad/.m2/repository/io/github/agdturner/ccg-math/2.0.10/ccg-math-2.0.10.jar(/uk/ac/leeds/ccg/math/number/Math_BigRational.class)]
[loading /home/brad/.m2/repository/io/github/agdturner/ccg-math/2.0.10/ccg-math-2.0.10.jar(/uk/ac/leeds/ccg/math/number/Math_BigRationalSqrt.class)]
[loading /modules/java.base/java/math/BigDecimal.class]
[loading /modules/java.base/java/lang/Comparable.class]
[loading /home/brad/.m2/repository/ch/obermuhlner/big-math/2.3.0/big-math-2.3.0.jar(/ch/obermuhlner/math/big/BigDecimalMath.class)]
[loading /modules/java.base/java/math/MathContext.class]
[loading /home/brad/.m2/repository/io/github/agdturner/ccg-math/2.0.10/ccg-math-2.0.10.jar(/uk/ac/leeds/ccg/math/arithmetic/Math_BigInteger.class)]
[loading /modules/java.base/java/lang/String.class]
[loading /home/brad/.m2/repository/io/github/agdturner/ccg-math/2.0.10/ccg-math-2.0.10.jar(/uk/ac/leeds/ccg/math/arithmetic/Math_BigDecimal.class)]
[loading /home/brad/.m2/repository/io/github/agdturner/ccg-math/2.0.10/ccg-math-2.0.10.jar(/uk/ac/leeds/ccg/math/matrices/Math_Matrix_BR.class)]
[loading /modules/java.base/java/util/HashSet.class]
[loading /modules/java.base/java/util/Iterator.class]
[loading /modules/java.base/java/util/Arrays.class]
[loading /modules/java.base/java/util/List.class]
[loading /modules/java.base/java/util/Set.class]
[loading /modules/java.base/java/lang/Deprecated.class]
[loading /modules/java.base/java/lang/Override.class]
[loading /modules/java.base/java/lang/annotation/Annotation.class]
[loading /modules/java.base/java/lang/annotation/Retention.class]
[loading /modules/java.base/java/lang/annotation/RetentionPolicy.class]
[loading /modules/java.base/java/lang/annotation/Target.class]
[loading /modules/java.base/java/lang/annotation/ElementType.class]
[checking <anonymous >]
[checking uk.ac.leeds.ccg.v3d.geometry.light.V3D_VGeometry]
[loading /modules/java.base/java/lang/AutoCloseable.class]
[checking uk.ac.leeds.ccg.v3d.geometry.light.V3D_V]
[loading /modules/java.base/java/math/BigInteger.class]
[loading /modules/java.base/java/lang/Number.class]
[loading /modules/java.base/java/lang/Class.class]
[loading /modules/java.base/java/lang/reflect/GenericDeclaration.class]
[loading /modules/java.base/java/lang/reflect/AnnotatedElement.class]
[loading /modules/java.base/java/lang/reflect/Type.class]
[loading /modules/java.base/java/lang/invoke/TypeDescriptor.class]
[loading /modules/java.base/java/lang/invoke/TypeDescriptor$OfField.class]
[loading /modules/java.base/java/lang/constant/Constable.class]
[loading /modules/java.base/java/lang/Byte.class]
[loading /modules/java.base/java/lang/Character.class]
[loading /modules/java.base/java/lang/Short.class]
[loading /modules/java.base/java/lang/Long.class]
[loading /modules/java.base/java/lang/Float.class]
[loading /modules/java.base/java/lang/Integer.class]
[loading /modules/java.base/java/lang/Double.class]
[loading /modules/java.base/java/lang/Boolean.class]
[loading /modules/java.base/java/lang/Void.class]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Point]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Geometry]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_IntersectionAndDistanceCalculations]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_FiniteGeometry]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Vector]
[checking uk.ac.leeds.ccg.v3d.geometry.light.V3D_VPoint]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Envelope]
[loading /modules/java.base/java/lang/RuntimeException.class]
[loading /modules/java.base/java/lang/Throwable.class]
[loading /modules/java.base/java/lang/Exception.class]
[loading /modules/java.base/java/lang/UnsupportedOperationException.class]
[checking uk.ac.leeds.ccg.v3d.core.V3D_Environment]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Line]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_LineSegment]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Ray]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Plane]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Triangle]
[loading /modules/java.base/java/util/Collection.class]
[loading /modules/java.base/java/lang/Iterable.class]
[loading /modules/java.base/java/lang/Cloneable.class]
[loading /modules/java.base/java/util/AbstractSet.class]
[loading /modules/java.base/java/util/AbstractCollection.class]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Face]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Tetrahedron]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Volume]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Points]
[checking uk.ac.leeds.ccg.v3d.geometry.light.V3D_VTriangle]
[checking uk.ac.leeds.ccg.v3d.geometry.light.V3D_VTetrahedron]
[checking uk.ac.leeds.ccg.v3d.geometry.light.V3D_VLine]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_TrianglesCoplanar]
[loading /modules/java.base/java/util/function/IntFunction.class]
[loading /modules/java.base/java/util/stream/Stream.class]
[loading /modules/java.base/java/util/stream/BaseStream.class]
[loading /modules/java.base/java/util/function/Predicate.class]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Rectangle]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_Tetrahedrons]
[loading /modules/java.base/java/util/Optional.class]
[loading /modules/java.base/java/lang/StringBuilder.class]
[loading /modules/java.base/java/lang/CharSequence.class]
[loading /modules/java.base/java/lang/AbstractStringBuilder.class]
[loading /modules/java.base/java/lang/Appendable.class]
[loading /modules/java.base/java/lang/StringBuffer.class]
[loading /modules/java.base/java/lang/constant/ConstantDesc.class]
[loading /modules/java.base/java/util/function/Consumer.class]
[checking uk.ac.leeds.ccg.v3d.geometry.V3D_LineSegmentsCollinear]
[total 2203ms]
[INFO] -------------------------------------------------------------
**[ERROR] COMPILATION ERROR :
[INFO] -------------------------------------------------------------
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[25,35] cannot find symbol
  symbol:   class Math_Quaternion_BigRational
  location: package uk.ac.leeds.ccg.math.number
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[427,39] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[428,39] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[429,39] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[707,50] cannot find symbol
  symbol:   method sin(uk.ac.leeds.ccg.math.arithmetic.Math_BigInteger,int)
  location: variable thetaDiv2 of type uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[708,39] cannot find symbol
  symbol:   method cos(uk.ac.leeds.ccg.math.arithmetic.Math_BigInteger,int)
  location: variable thetaDiv2 of type uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[712,9] cannot find symbol
  symbol:   class Math_Quaternion_BigRational
  location: class uk.ac.leeds.ccg.v3d.geometry.V3D_Vector
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[712,45] cannot find symbol
  symbol:   class Math_Quaternion_BigRational
  location: class uk.ac.leeds.ccg.v3d.geometry.V3D_Vector
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[715,9] cannot find symbol
  symbol:   class Math_Quaternion_BigRational
  location: class uk.ac.leeds.ccg.v3d.geometry.V3D_Vector
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[715,46] cannot find symbol
  symbol:   class Math_Quaternion_BigRational
  location: class uk.ac.leeds.ccg.v3d.geometry.V3D_Vector
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[717,9] cannot find symbol
  symbol:   class Math_Quaternion_BigRational
  location: class uk.ac.leeds.ccg.v3d.geometry.V3D_Vector
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[717,45] cannot find symbol
  symbol:   class Math_Quaternion_BigRational
  location: class uk.ac.leeds.ccg.v3d.geometry.V3D_Vector
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[721,9] cannot find symbol
  symbol:   class Math_Quaternion_BigRational
  location: class uk.ac.leeds.ccg.v3d.geometry.V3D_Vector
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java:[882,48] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java:[884,48] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java:[886,48] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedron.java:[333,68] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedron.java:[335,68] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedron.java:[337,68] cannot find symbol
  symbol:   method round(int)
  location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[INFO] 19 errors
[INFO] -------------------------------------------------------------
[INFO] ------------------------------------------------------------------------
[INFO] BUILD FAILURE
[INFO] ------------------------------------------------------------------------
[INFO] Total time:  6.734 s
[INFO] Finished at: 2022-05-25T11:58:00-07:00
[INFO] ------------------------------------------------------------------------
[ERROR] Failed to execute goal org.apache.maven.plugins:maven-compiler-plugin:3.8.1:compile (default-compile) on project ccg-v3d: Compilation failure: Compilation failure:
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[25,35] cannot find symbol
[ERROR]   symbol:   class Math_Quaternion_BigRational
[ERROR]   location: package uk.ac.leeds.ccg.math.number
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[427,39] cannot find symbol
[ERROR]   symbol:   method round(int)
[ERROR]   location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[428,39] cannot find symbol
[ERROR]   symbol:   method round(int)
[ERROR]   location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[429,39] cannot find symbol
[ERROR]   symbol:   method round(int)
[ERROR]   location: class uk.ac.leeds.ccg.math.number.Math_BigRational
[ERROR] /home/brad/develop/JOSS-review/ccg-v3d/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java:[707,50] cannot find symbol
[ERROR]   symbol:   method sin(uk.ac.leeds.ccg.math.arithmetic.Math_BigInteger,int)
[ERROR]   location: variable thetaDiv2 of type uk.ac.leeds.ccg.math.number.Math_BigRational
[…]
