### 2022-05-25 agdturner/ccg-v3d#2 @agdturner
Math_Quaternion_BigRational is in ccg-math in the path uk.ac.leeds.ccg.math.number
The version of ccg-math is 2.0.10 and this contains that class:
![image](https://user-images.githubusercontent.com/5878594/170367129-8d89b81f-c74c-492a-ad37-ab12197fedf5.png)

What version of Eclipse are you using? Same as me: Version: 2022-03 (4.23.0) Build id: 20220310-1457?
