### 2022-05-25 agdturner/ccg-v3d#2 @agdturner
I ran the tests in Eclipse and it seems like everything is fine:
![image](https://user-images.githubusercontent.com/5878594/170371108-f0bca7cb-05ad-46eb-820e-e466b00ae452.png)
