### 2022-05-26 agdturner/ccg-v3d#2 @behollister
Okay, missing "uk.ac.leeds.ccg.math.number.Math_Quaternion_BigRational." Doesn't look like my local copy has Math_Quaternion_BigRational.class under the uk.ac.leeds.ccg.math.number package (see attached screen pic).

Shouldn't that be provided by the Maven build system, as added in the POM? Is perhaps an incorrect version of the math.number package missing the Quaternion class as fetched by Maven (different from your local environment)?

![missingQuatLib](https://user-images.githubusercontent.com/20388050/170542930-f7cb31d6-fae1-4d86-89cd-597192ce6dc2.png)
