### 2022-05-28 agdturner/ccg-v3d#2 @agdturner
Sorry @behollister you are quite correct, this is not in the ccg-math version 2.0.10 Maven artefact. I have bumped the version of this dependency to 2.1 and changed the ccg-v3d pom.xml to refer to this version and updated the version of ccg-v3d to version 0.15. Can you use these new versions? Is there anything I need to do to adjust things for the JOSS review/paper? Once again I do apologise for this and thanks for your work and patients with this review. Andy
