### 2022-06-03 agdturner/ccg-v3d#2 @behollister
Hi Andy,

Got busy the last few days with a new book proposal. Will continue with the review tomorrow. If all goes well, will try to finish up then.

Note that there's still the other reviewer that hasn't started yet, I don't believe.
