### 2022-06-03 agdturner/ccg-v3d#3 @behollister
JOSS Review: Needs Usage Example
Per https://github.com/openjournals/joss-papers/blob/joss.04283/joss.04283/10.21105.joss.04283.pdf,
@agdturner states there are application cases for the library. However, only the unit tests are provided at present.

According to the review checklist:
Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

Related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4283
