### 2022-06-03 #4283 @behollister
@agdturner please see remaining items on review checklist for @behollister.

For documentation of the API, please run Doxygen and provide linked html (hosted on github) for:
Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Currently, only code comments and brief README.md outline API.
