### 2022-06-03 agdturner/ccg-v3d#4 @behollister
JOSS Review: Software Paper References
There is only a single reference. It does not fully support the claim (from the software paper) for the need of the library under review. Only big-math https://github.com/eobermuhlner/big-math is referenced currently.

"References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?"

Please include papers and/or repositories providing further evidence for claims.

Refers to JOSS Review: https://github.com/openjournals/joss-reviews/issues/4283
