### 2022-06-03 agdturner/ccg-v3d#5 @behollister
JOSS Review: API Documentation
@agdturner please see remaining items on review checklist for @behollister.

For documentation of the API, please run Doxygen and provide linked html (hosted on github) for:
Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

Currently, only code comments and brief README.md outline API.

Related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4283
