### 2022-06-03 #4283 @behollister
@VivianePons, @agdturner, according to https://github.com/openjournals/joss-reviews/issues/4283#issuecomment-1124573068, it might be prudent for @agdturner to expand on unchecked items (@behollister 'checklist') as of Jun 3, before a possible acceptance of the submission.

@agdturner, you mention additional work this summer that will strengthen the submission in terms of "statement of need."
