### 2022-06-05 agdturner/ccg-v3d#5 @agdturner
Javadocs are compiled as part of the Maven release process, these are in the target/apidocs directory. Are you sure that running Doxygen is necessary? I'm not convinced given the answers to the following linked StackOverflow question: https://stackoverflow.com/questions/225447/doxygen-vs-javadoc
