### 2022-06-05 agdturner/ccg-v3d#3 @agdturner
I hope to find time to work on use case scenarios soon. From a practical sense, there are many use case scenarios in the spatial domain in terms of Digital Twins and georeferencing. There are practical applications for much of the functionality, for instance finding the minimum distance between any two line segments to an order of magnitude precision, or the minimum distance between two triangles - something which is key for collision detection/avoidance. I am planning to work on a physics use case to do with parallelepiped minimum distance calculations for a physics use case scenario in the summer. I am not sure I yet fully understand the problem, but the idea is to figure out a way of doing this that is quicker than a brute force search for all the other parallelepipeds within a given radius. I am informed that it is easy for some radii, but not for others... My hope is that the arbitrary precision nature of this library will provide some advantage over floating point...
