### 2022-06-05 agdturner/ccg-v3d#4 @agdturner
Would it be best to withdraw the submission, develop the Use Case (some Use Case Scenarios), and then resubmit to JOSS?
