### 2022-06-09 #4283 @behollister
Hi @VivianePons. Can you please answer @agdturner 's question here:
https://github.com/agdturner/ccg-v3d/issues/4#issuecomment-1146891572

Also, before any decision can be made, I believe we need feedback from @abhishekvp as well.
