### 2022-06-10 #4283 @VivianePons
Dear @agdturner and @behollister

If I understand correctly, @behollister has raised an issue and @agdturner is asking if it's better to withdraw and resubmit. This really depends on how much time is needed to "fix" the issue in a way that would be sufficient for the the reviewer to be satisfied. From what I read, it does not concern directly the software but more the paper itself and in particular the use cases.

I'd say, it's really up to you @agdturner . If you believe you can make this change in the coming days / weeks, then we can continue with this process. On the other hand, if you prefer to have more time and to prepare a better, more complete, submission, this is also possible.

About @abhishekvp : he hasn't answered my emails. If we continue to the process, I will contact him again and if still no response, I'll look for another reviewer. For the paper to get published, we indeed need 2 completed reviews
