### 2022-06-11 #4283 @agdturner
I don't have time in the next two months to develop a scientific use case that proves the utility of the software in providing a new capability. There are many general applications for processing tetrahedra (https://en.wikipedia.org/wiki/Tetrahedron#Applications) and this is one thing that the library supports, for example, calculating the minimum distance between two tetrahedrons, which could be useful for physical object collision avoidance...
