### 2022-06-11 #4283 @VivianePons
Yes I perfectly understand.

From what I read in the comment, the original question was about the "References" check box but somehow, it drifted and now it seems to me we are discussing the "Substancial scholarly effort"

So my questions would be for @behollister : in the current state of the paper and with the explanation from @agdturner , do you believe you can check those two boxes? If not, can you be a bit more specific about what is missing? For example, what part of the paper would require more citations in your opinion?
