### 2022-06-14 #4283 @behollister
Hi @VivianePons and  @agdturner,

Currently there are only unit tests and one reference.

An application (short) need not be an original use case, but something complementary to existing literature -- @abhishekvp may have less/more requirements. Select an existing publication suggesting a calculation using the submission library for completeness. Include the application (and reference paper) with the JOSS submission. Relates to: https://github.com/agdturner/ccg-v3d/issues/4.

I don't believe this should require a withdraw and resubmit.
