### 2022-08-03 #4283 @abhishekvp
Apologies for the late review. Please find my review comments below:

The paper proposes and implements the [ccg-v3d](https://github.com/agdturner/ccg-v3d) - an extensible Java library for working with 3D Euclidian Geometry.

Thanks to @agdturner for considering to submit to the Journal of Open Source Software. The author's contribution to the open source community are appreciated.

Although the code contributions in the repository seem complete and useful, the research paper has significant room for improvement. As per the reviewer checklist - specific comments to each of the checklist items are as follows:

- Summary: Although fairly articulated, it is recommended that the summary convey the functionality of the library in brief. In its current state, the summary is a 8 paragraph long section with significant technical details. Specific technical details are absolutely relevant but perhaps can be attenuated considering that the summary section is aimed for a diverse, non-specialist audience. The specific technical details can be considered to be moved to another section ahead in the paper.
- A Statement of Need: The author has attempted to convey the need by stating that the precision for floating point arithmetic varies and could possible result in uncertainties, thus using heavy-weight numeric representations could be useful. It is recommended that author elaborate on the "uncertainties" and convincingly convey the need for this library in context to the existing libraries for this problem and domain. It is also suggested that author try to avoid using first-person pronouns ("I") in a research report. A passive voice could be adopted instead.
-   State of the field - The author has not described how this particular library compares with other commonly-used packages.
- Quality of Writing - The quality of writing is fair.
- Reference - It is strongly recommended that the author consider adding more reference pertaining to existing software libraries/packages that work with 3D geometries and work towards establishing the relevance of [ccg-v3d](https://github.com/agdturner/ccg-v3d) in their context as related works.

On the documentation aspect of the code-repository -

- Example Usage: It is important for a library to clearly specify example usages. The author is strongly recommended to add those in the repository documentation.

Overall the code repository is mature, but the paper has room for improvement. It is recommended that author refer the guidelines for submitting to JOSS [here](https://joss.readthedocs.io/en/latest/submitting.html). Sincerely appreciate the author's efforts towards sharing this work with the open-source community.
