### 2022-09-02 24b898d
Added untested simplify method to V3D_LineSegmentsCollinear.
