### 2022-09-03 f72730e
Updated simplify method for V3D_LineSegmentsCollinear added stub methods for getConvexHull.
