### 2022-09-03 48e99af
Updated simplify method for V3D_LineSegmentsCollinear added stub methods for getConvexHull.
