### 2022-09-05 237e780
Updated simplify method for V3D_LineSegmentsCollinear added stub methods for getConvexHull.
