+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_LineSegmentsCollinearTest.java
@@ -1,0 +1,485 @@
+/*
+ * Copyright 2022 Centre for Computational Geography.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.ac.leeds.ccg.v3d.geometry;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+import uk.ac.leeds.ccg.math.number.Math_BigRational;
+import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
+import uk.ac.leeds.ccg.v3d.V3D_Test;
+
+/**
+ *
+ * @author Andy Turner
+ */
+public class V3D_LineSegmentsCollinearTest extends V3D_Test {
+
+    public V3D_LineSegmentsCollinearTest() {
+    }
+
+    @BeforeAll
+    public static void setUpClass() {
+    }
+
+    @AfterAll
+    public static void tearDownClass() {
+    }
+
+    @BeforeEach
+    public void setUp() {
+    }
+
+    @AfterEach
+    public void tearDown() {
+    }
+
+    /**
+     * Test of toString method, of class V3D_LineSegmentsCollinear.
+     */
+    @Test
+    public void testToString() {
+        System.out.println("toString");
+        V3D_LineSegmentsCollinear instance = new V3D_LineSegmentsCollinear(
+                new V3D_LineSegment(pP0P0P0, pP1P0P0),
+                new V3D_LineSegment(pP0P1P0, pP1P1P0));
+        String expResult = "uk.ac.leeds.ccg.v3d.geometry.V3D_LineSegmentsCollinear(V3D_LineSegment\n"
+                + "(\n"
+                + " oom=-3\n"
+                + " ,\n"
+                + " offset=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dy=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dz=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0)\n"
+                + " )\n"
+                + " ,\n"
+                + " p=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dy=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dz=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0)\n"
+                + " )\n"
+                + " ,\n"
+                + " q=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=1, sqrtx=1, oom=0),\n"
+                + "  dy=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
[…]
