### 2022-09-07 8fba9f5
First test case for getConvexHull.
