+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_TrianglesCoplanar.java
@@ -23,7 +23,9 @@ import java.util.Iterator;
 import java.util.List;
 //import java.util.List;
 import java.util.Objects;
+import java.util.TreeSet;
 import uk.ac.leeds.ccg.math.number.Math_BigRational;
+import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
 //import java.util.ArrayList;
 
 /**
@@ -75,7 +77,12 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
      * For storing the convex hull.
      */
     protected ArrayList<V3D_Point> convexHull;
-            
+
+    /**
+     * For storing the points.
+     */
+    protected HashSet<V3D_Point> points;
+
     /**
      * Create a new instance.
      *
@@ -85,7 +92,11 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
         super(triangles[0].e, triangles[0].offset, triangles[0].p,
                 triangles[0].q, triangles[0].r);
         this.triangles = new ArrayList<>();
-        this.triangles.addAll(Arrays.asList(triangles));
+        // Ensure all triangles have the same offset as the first.
+        for (V3D_Triangle t : triangles) {
+            t.setOffset(triangles[0].offset);
+            this.triangles.add(t);
+        }
     }
 
     @Override
@@ -222,7 +233,9 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
      * then a simplification takes place and this is done iteratively until all
      * such simplification is done. This however does not simplify the following
      * case where it can be seen that the 4 triangles represent an area that can
-     * be given as a single triangle: null null null null null     {@code
+     * be given as a single triangle: null null null null null null null null
+     * null null null null null null null null null null null null null null
+     * null null null null null null     {@code
      * *------*------*
      * |     /|     /
      * |    / |    /
@@ -432,7 +445,6 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
 //    public boolean isEnvelopeIntersectedBy(V3D_Line l, int oom) {
 //        return getEnvelope().isIntersectedBy(l, oom);
 //    }
-
     /**
      * For this to work as expected, all triangles should have the same offset
      * (point for the rotation).
@@ -446,35 +458,314 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
             t.rotate(axisOfRotation, theta);
         }
     }
-    
+
     /**
-     * @return The convex hull. 
+     * @return Get the relPoints.
+     */
+    public HashSet<V3D_Point> getPoints() {
+        if (points == null) {
+            points = new HashSet<>();
+            for (V3D_Triangle t : triangles) {
+                t.setOffset(offset);
+                points.add(t.getP());
+                points.add(t.getQ());
+                points.add(t.getR());
+            }
+        }
+        return points;
+    }
+
+    /**
+     * Return the convex hull calculating it first if it has not already been
+     * calculated. Below is the basic algorithm:
+     * <ol>
+     * <li>Partition the points:
+     * <ul>
+     * <li>Calculate the distances between the points with the minimum and
+     * maximum x, the minimum and maximum y, and the minimum and maximum z
+     * values.</li>
+     * <li>Choose the points that have the largest distance between them to
+     * define the dividing plane that is orthogonal to the plane of the
+     * polygon.</li>
+     * <li>Let the points on one side of the dividing plane be one group and
+     * those on the other be the another group.</li>
+     * </ul></li>
[…]
