### 2022-09-07 f05883a
Ordered convex hull output.
