+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_TrianglesCoplanarTest.java
@@ -32,22 +32,22 @@ import uk.ac.leeds.ccg.v3d.V3D_Test;
  * @author Andy Turner
  */
 public class V3D_TrianglesCoplanarTest extends V3D_Test {
-    
+
     public V3D_TrianglesCoplanarTest() {
     }
-    
+
     @BeforeAll
     public static void setUpClass() {
     }
-    
+
     @AfterAll
     public static void tearDownClass() {
     }
-    
+
     @BeforeEach
     public void setUp() {
     }
-    
+
     @AfterEach
     public void tearDown() {
     }
@@ -62,11 +62,103 @@ public class V3D_TrianglesCoplanarTest extends V3D_Test {
                 new V3D_Triangle(pP0P0P0, pP0P1P0, pP1P1P0),
                 new V3D_Triangle(pP0P0P0, pP2P2P0, pP0P2P0),
                 new V3D_Triangle(pP0P0P0, pP0P2P0, pN1P2P0));
-        String expResult = "";
+        String expResult = "uk.ac.leeds.ccg.v3d.geometry.V3D_TrianglesCoplanar(V3D_Triangle\n"
+                + "(\n"
+                + " oom=-3\n"
+                + " ,\n"
+                + " offset=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dy=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dz=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0)\n"
+                + " )\n"
+                + " ,\n"
+                + " p=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dy=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dz=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0)\n"
+                + " )\n"
+                + " ,\n"
+                + " q=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dy=Math_BigRationalSqrt(x=1, sqrtx=1, oom=0),\n"
+                + "  dz=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0)\n"
+                + " )\n"
+                + " ,\n"
+                + " r=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=1, sqrtx=1, oom=0),\n"
+                + "  dy=Math_BigRationalSqrt(x=1, sqrtx=1, oom=0),\n"
+                + "  dz=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0)\n"
+                + " )\n"
+                + "), V3D_Triangle\n"
+                + "(\n"
+                + " oom=-3\n"
+                + " ,\n"
+                + " offset=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dy=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
+                + "  dz=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0)\n"
+                + " )\n"
+                + " ,\n"
+                + " p=V3D_Vector\n"
+                + " (\n"
+                + "  dx=Math_BigRationalSqrt(x=0, sqrtx=0, oom=0),\n"
[…]
