+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_TrianglesCoplanar.java
@@ -508,6 +508,8 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
      * <li>Repeat the process dealing with each group in turn (Steps 3 to 9) in
      * a depth first manner.</li>
      * </ol>
+     * The convex hull is returned in an order effectively tracing out a 
+     * polygon.
      *
      * @return Get the convex hull.
      */
@@ -575,11 +577,11 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
                 if (ydl2.compareTo(zdl2) == 1) {
                     convexHull.add(yminp);
                     convexHull.add(ymaxp);
-                    getConvexHull0(pts, yminp, ymaxp, n);
+                    getConvexHull0(pts, yminp, ymaxp, n, 1);
                 } else {
                     convexHull.add(zminp);
                     convexHull.add(zmaxp);
-                    getConvexHull0(pts, zminp, zmaxp, n);
+                    getConvexHull0(pts, zminp, zmaxp, n, 1);
                 }
             } else if (yminIndex == ymaxIndex) {
                 V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp);
@@ -589,11 +591,11 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
                 if (xdl2.compareTo(zdl2) == 1) {
                     convexHull.add(xminp);
                     convexHull.add(xmaxp);
-                    getConvexHull0(pts, xminp, xmaxp, n);
+                    getConvexHull0(pts, xminp, xmaxp, n, 1);
                 } else {
                     convexHull.add(zminp);
                     convexHull.add(zmaxp);
-                    getConvexHull0(pts, zminp, zmaxp, n);
+                    getConvexHull0(pts, zminp, zmaxp, n, 1);
                 }
             } else if (zminIndex == zmaxIndex) {
                 V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp);
@@ -603,11 +605,11 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
                 if (xdl2.compareTo(ydl2) == 1) {
                     convexHull.add(xminp);
                     convexHull.add(xmaxp);
-                    getConvexHull0(pts, xminp, xmaxp, n);
+                    getConvexHull0(pts, xminp, xmaxp, n, 1);
                 } else {
                     convexHull.add(yminp);
                     convexHull.add(ymaxp);
-                    getConvexHull0(pts, yminp, ymaxp, n);
+                    getConvexHull0(pts, yminp, ymaxp, n, 1);
                 }
             } else {
                 V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp);
@@ -620,21 +622,21 @@ public class V3D_TrianglesCoplanar extends V3D_Plane implements V3D_Face {
                     if (xdl2.compareTo(zdl2) == 1) {
                         convexHull.add(xminp);
                         convexHull.add(xmaxp);
-                        getConvexHull0(pts, xminp, xmaxp, n);
+                        getConvexHull0(pts, xminp, xmaxp, n, 1);
                     } else {
                         convexHull.add(zminp);
                         convexHull.add(zmaxp);
-                        getConvexHull0(pts, zminp, zmaxp, n);
+                        getConvexHull0(pts, zminp, zmaxp, n, 1);
                     }
                 } else {
                     if (ydl2.compareTo(zdl2) == 1) {
                         convexHull.add(yminp);
                         convexHull.add(ymaxp);
-                        getConvexHull0(pts, yminp, ymaxp, n);
[…]
