### 2022-09-09 2493d8b
Update paper.bib
Added entry for clipper 2 and updated agdt-java to ccg.
