+++ paper.bib
@@ -16,20 +16,29 @@
   url = {https://github.com/eobermuhlner/big-math}
 }
 
-@misc{agdt-java-math,
+@misc{ccg-math,
   author = {A. Turner},
-  title = {agdt-java-math: A modularised Java library with functionality: for fixed point and arbitrary precision arithmetic; matrices; getting prime numbers; and, generating pseudorandom numbers.},
+  title = {ccg-math: A modularised Java library with functionality: for fixed point and arbitrary precision arithmetic; matrices; getting prime numbers; and, generating pseudorandom numbers.},
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/agdturner/agdt-java-math}
+  url = {https://github.com/agdturner/ccg-math}
 }
 
-@misc{agdt-java-generic,
+@misc{ccg-generic,
   author = {A. Turner},
-  title = {agdt-java-generic: A modularised Java library with generic functionality.},
+  title = {ccg-generic: A modularised Java library with generic functionality.},
   year = {2021},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://github.com/agdturner/agdt-java-generic}
+  url = {https://github.com/agdturner/ccg-generic}
+}
+
+@misc{clipper2,
+  author = {A. Johnson},
+  title = {Clipper2 - Polygon Clipping and Offsetting.},
+  year = {2022},
+  publisher = {A. Johnson},
+  journal = {Website},
+  url = {http://www.angusj.com/clipper2}
 }
