### 2022-09-09 6d0a03a
Bug fix for intersection code. Added a new class for polygons.
