+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Polygon.java
@@ -1,0 +1,376 @@
+/*
+ * Copyright 2020 Andy Turner, University of Leeds.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.ac.leeds.ccg.v3d.geometry;
+
+import java.math.BigDecimal;
+import java.util.ArrayList;
+import java.util.Iterator;
+import uk.ac.leeds.ccg.math.number.Math_BigRational;
+//import java.util.ArrayList;
+
+/**
+ * Comprising two collections of V3D_ConvexHullCoplanar one representing parts
+ * and the other representing holes. Parts may intersect. Holes may intersect.
+ * So in OGC terms, this is pore like a multi-polygon as holes may effectively
+ * split parts and they may collectively cover all parts. Simplification of a
+ * V3D_Polygon may be possible and may involve creating new geometries to
+ * represent parts individually which may involve some rounding.
+ *
+ * @author Andy Turner
+ * @version 1.0
+ */
+public class V3D_Polygon extends V3D_Plane implements V3D_Face {
+
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * For storing the envelope.
+     */
+    protected V3D_Envelope en;
+
+    /**
+     * The collection of non-holes.
+     */
+    protected final ArrayList<V3D_ConvexHullCoplanar> parts;
+
+    /**
+     * The collection of holes.
+     */
+    protected final ArrayList<V3D_ConvexHullCoplanar> holes;
+
+    /**
+     * For storing the convex hull of the parts.
+     */
+    protected V3D_ConvexHullCoplanar convexHull;
+
+    /**
+     * Create a new instance.
+     *
+     * @param parts A non-empty list of V3D_ConvexHullCoplanar parts.
+     * @param holes A potentially empty list of V3D_ConvexHullCoplanar holes.
+     */
+    public V3D_Polygon(ArrayList<V3D_ConvexHullCoplanar> parts,
+            ArrayList<V3D_ConvexHullCoplanar> holes) {
+        super(parts.get(0).triangles.get(0));
+        this.parts = parts;
+        this.holes = holes;
+    }
+
+    /**
+     * Create a new instance.
+     *
+     * @param triangles A non-empty list of coplanar triangles.
+     */
+    public V3D_Polygon(V3D_Triangle... parts) {
+        super(parts[0]);
+        this.parts = new ArrayList<>();
+        this.holes = null;
+    }
+
+    @Override
+    public String toString() {
+        String s = this.getClass().getName() + "(";
+        {
+            s += "\nParts(\n";
+            Iterator<V3D_ConvexHullCoplanar> ite = parts.iterator();
+            s += ite.next().toString();
+            while (ite.hasNext()) {
+                s += ", " + ite.next();
+            }
+            s += "\n)\n";
+        }
+        {
+            if (holes != null) {
+                s += "\nHoles(\n";
+                Iterator<V3D_ConvexHullCoplanar> ite = holes.iterator();
+                s += ite.next().toString();
+                while (ite.hasNext()) {
+                    s += ", " + ite.next();
+                }
+                s += "\n)\n";
+            }
+        }
+        s += "\n)";
+        return s;
+    }
+
[…]
