### 2022-09-09 a8309a2
Line segment - line segment intersection code deals with more special cases.
