+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_LineSegment.java
@@ -411,127 +411,181 @@ public class V3D_LineSegment extends V3D_Line implements V3D_FiniteGeometry,
             return null;
         }
         // Get intersection of infinite lines. 
-        V3D_Geometry li = new V3D_Line(this).getIntersection(new V3D_Line(l),
+        V3D_Geometry r = new V3D_Line(this).getIntersection(new V3D_Line(l), 
                 oom);
-        if (li == null) {
-            return null;
+        if (r == null) {
+            return r;
         }
-        if (li instanceof V3D_Point) {
-            return li;
+        if (r instanceof V3D_Point) {
+            return r;
         }
         /**
          * Check the type of intersection. {@code
-         * 1)   p ---------- q
-         *         l.p ---------- l.q
-         * 2)   p ------------------------ q
-         *         l.p ---------- l.q
-         * 3)        p ---------- q
-         *    l.p ------------------------ l.q
-         * 4)        p ---------- q
-         *    l.p ---------- l.q
-         * 5)   q ---------- p
-         *         l.p ---------- l.q
-         * 6)   q ------------------------ p
-         *         l.p ---------- l.q
-         * 7)        q ---------- p
-         *    l.p ------------------------ l.q
-         * 8)        q ---------- p
-         *    l.p ---------- l.q
-         * 9)   p ---------- q
-         *         l.q ---------- l.p
-         * 10)   p ------------------------ q
-         *         l.q ---------- l.p
-         * 11)       p ---------- q
-         *    l.q ------------------------ l.p
-         * 12)       p ---------- q
-         *    l.q ---------- l.p
-         * 13)  q ---------- p
-         *         l.q ---------- l.p
-         * 14)  q ------------------------ p
-         *         l.q ---------- l.p
-         * 15)       q ---------- p
-         *    l.q ------------------------ l.p
-         * 16)       q ---------- p
-         *    l.q ---------- l.p
-         *
-         * 17)             p ---------- q
-         *    l.p ---------- l.q
-         * 18)  q ---------- p
-         *              l.q ---------- l.p
+         * A) p_q l.p_l.q
+         * 1)  p +----------+ q
+         *              l.p +----------+ l.q
+         * 2)  p +----------+ q
+         *         l.p +----------+ l.q
+         * 3)  p +------------------------+ q
+         *         l.p +----------+ l.q
+         * 4)        p +----------+ q
+         *     l.p +----------------------+ l.q
+         * 5)    p +----------+ q
+         *     l.p +----------+ l.q
+         * 6)         p +----------+ q
+         *     l.p +----------+ l.q
+         * 7)              p +----------+ q
+         *     l.p +---------+ l.q
+         * B) q_p l.p_l.q
+         * 8)  q +----------+ p
+         *              l.p +----------+ l.q
+         * 9)  q +----------+ p
+         *         l.p +----------+ l.q
+         * 10) q +------------------------+ p
+         *         l.p +----------+ l.q
+         * 11)       q +----------+ p
+         *     l.p +----------------------+ l.q
+         * 12)       q +----------+ p
+         *         l.p +----------+ l.q
+         * 13)       q +----------+ p
+         *     l.p +----------+ l.q
+         * 14)              q +----------+ p
+         *     l.p +----------+ l.q
+         * C) p_q l.q_l.p
+         * 15) p +----------+ q
+         *         l.q +----------+ l.p
+         * 16) p +----------+ q
+         *              l.q +----------+ l.p
+         * 17) p +------------------------+ q
+         *         l.q +----------+ l.p
+         * 18)       p +----------+ q
+         *     l.q +------------------------+ l.p
+         * 19)       p +----------+ q
[…]
