### 2022-09-13 28b3d82
Fixed bug with line triangle intersection.
