### 2022-09-15 6626f4c
Fixed bug in intersection code.
