### 2022-09-25 4a3da01
Rather than extending line, line segments now hold a reference to a line. Similarly triangles hold a reference to a plane.
