+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedron.java
@@ -60,14 +60,9 @@ import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
  * @author Andy Turner
  * @version 1.0
  */
-public class V3D_Tetrahedron extends V3D_Geometry implements V3D_Volume {
+public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
 
     private static final long serialVersionUID = 1L;
-
-    /**
-     * For storing the envelope.
-     */
-    protected V3D_Envelope en;
 
     /**
      * For defining one of the points that defines the tetrahedron.
@@ -112,7 +107,7 @@ public class V3D_Tetrahedron extends V3D_Geometry implements V3D_Volume {
     /**
      * Create a new instance. {@code p}, {@code q}, {@code r} and {@code s} must
      * all be different, not the zero vector and collectively they must be three
-     * dimensional. This is generally the fastest way to construct a 
+     * dimensional. This is generally the fastest way to construct a
      * tetrahedron.
      *
      * @param e What {@link #e} is set to.
@@ -155,14 +150,14 @@ public class V3D_Tetrahedron extends V3D_Geometry implements V3D_Volume {
     }
 
     /**
-     * Create a new instance. {@code p}, must not be coplanar to t. No test is 
+     * Create a new instance. {@code p}, must not be coplanar to t. No test is
      * done to check this is the case.
-     * 
+     *
      * @param p Used to set {@link #p}, {@link #e} and {@link #offset}.
      * @param t Used to set {@link #q}, {@link #r} and {@link #s}.
      */
     public V3D_Tetrahedron(V3D_Point p, V3D_Triangle t) {
-        this(p, t.getP(), t.getQ(), t.getR());
+        this(p, t.p.getP(), t.p.getQ(), t.p.getR());
     }
 
     @Override
@@ -203,19 +198,18 @@ public class V3D_Tetrahedron extends V3D_Geometry implements V3D_Volume {
     public V3D_Envelope getEnvelope() {
         if (en == null) {
 //            en = pqr.getEnvelope(oom).union(qsr.getEnvelope(oom));
-            en = getP(e.oom).getEnvelope()
-                    .union(getQ(e.oom).getEnvelope())
-                    .union(getR(e.oom).getEnvelope())
-                    .union(getS(e.oom).getEnvelope());
+            en = getP().getEnvelope()
+                    .union(getQ().getEnvelope())
+                    .union(getR().getEnvelope())
+                    .union(getS().getEnvelope());
         }
         return en;
     }
 
     /**
-     * @param oom The Order of Magnitude for the precision of the calculation.
      * @return {@link #p} with {@link #offset} applied.
      */
-    public V3D_Point getP(int oom) {
+    public V3D_Point getP() {
         return new V3D_Point(e, offset, p);
     }
 
@@ -223,23 +217,21 @@ public class V3D_Tetrahedron extends V3D_Geometry implements V3D_Volume {
      * @param oom The Order of Magnitude for the precision of the calculation.
      * @return {@link #p} with {@link #offset} applied.
      */
-    public V3D_Point getQ(int oom) {
+    public V3D_Point getQ() {
         return new V3D_Point(e, offset, q);
     }
 
     /**
-     * @param oom The Order of Magnitude for the precision of the calculation.
      * @return {@link #p} with {@link #offset} applied.
      */
-    public V3D_Point getR(int oom) {
+    public V3D_Point getR() {
         return new V3D_Point(e, offset, r);
     }
 
     /**
-     * @param oom The Order of Magnitude for the precision of the calculation.
      * @return {@link #p} with {@link #offset} applied.
      */
[…]
