### 2022-09-26 2ea7501
Rather than extending line, line segments now hold a reference to a line. Similarly triangles hold a reference to a plane.
