+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java
@@ -65,7 +65,7 @@ import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
 public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
 
     private static final long serialVersionUID = 1L;
-    
+
     public V3D_Plane p;
 
 //    /**
@@ -206,8 +206,8 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
         }
         return en;
     }
-    
-@Override
+
+    @Override
     public V3D_Point[] getPoints() {
         V3D_Point[] re = new V3D_Point[3];
         re[0] = p.getP();
@@ -215,7 +215,7 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
         re[2] = p.getR();
         return re;
     }
-    
+
 //    /**
 //     * @param v The vector to apply.
 //     * @param oom The Order of Magnitude for the precision of the calculation.
@@ -332,79 +332,72 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
         if (p.isIntersectedBy(l, oom)) {
             V3D_Geometry g = p.getIntersection(l, oom);
             if (g instanceof V3D_Point pt) {
-                /**
-                 * Logically, if the point is on the triangle than it
-                 * intersects. However, the point is only given accurately to
-                 * the given precision, so testing if the point is on the
-                 * triangle does not work. The option is to use the direction of
-                 * the line vector to define planes using each edge of the
-                 * triangle. We can then test that the point is on the same side
-                 * of these planes to know it is on the triangle. The commented
-                 * out line below seems logical, but does not work. It is left
-                 * in the code just in case.
-                 */
                 if (isIntersectedBy(pt, oom)) {
                     return true;
+                    /**
+                     * Logically, if the point of the line plane intersection is
+                     * on the triangle then it intersects. However, the point is
+                     * only given accurately to the given precision, so testing
+                     * if the point is on the triangle does not work. The
+                     * commented out else clause and this comment is in the code
+                     * on purpose.
+                     */
 //                } else {
 //                    return false; // This seems logical, but does not work for all cases in practice.
                 }
-                // Find the direction of the line towards the triangle and use:
-                // https://en.wikipedia.org/wiki/M%C3%B6ller%E2%80%93Trumbore_intersection_algorithm
-//                V3D_Point p = getP();
-//                V3D_Point q = getQ();
-//                V3D_Point r = getR();
-//                V3D_Point qp = new V3D_Point(e, q.offset, getQV().add(l.v, oom));
-//                V3D_Plane a = new V3D_Plane(p, q, qp);
-//                V3D_Point rp = new V3D_Point(e, r.offset, getRV().add(l.v, oom));
-//                V3D_Plane b = new V3D_Plane(q, r, rp);
-//  //              V3D_Point pp = new V3D_Point(e, p.offset, getPV().add(l.v, oom));
-//                V3D_Plane c = new V3D_Plane(r, p, rp);
-//                Math_BigRational apt = a.getSideOfPlane(pt);
-//                Math_BigRational bpt = b.getSideOfPlane(pt);
-//                Math_BigRational cpt = c.getSideOfPlane(pt);
-//                if (apt.compareTo(Math_BigRational.ZERO) == 1) {
-//                    if (bpt.compareTo(Math_BigRational.ZERO) == 1) {
-//                        if (cpt.compareTo(Math_BigRational.ZERO) == 1) {
-//                            return true;
-//                        }
[…]
