+++ README.md
@@ -26,7 +26,7 @@ Point positions in space are defined using [cartesian](https://en.wikipedia.org/
 [Sonatype SNAPSHOT Directory](https://oss.sonatype.org/content/repositories/snapshots/io/github/agdturner/ccg-v3d/0.15-SNAPSHOT/)
 
 ## Getting Started
-It is advised that you load this as a [Maven project](https://maven.apache.org/) in your favourite IDE. This should make it easy to download the dependencies and run the tests. The tests provide some indication of how you might used the library for example to determine the intersection of two geometries. Another entry point is to use the [ccg-r3d](https://github.com/agdturner/ccg-r3d) - a rendering library that can visualise [V3D_Geometry](https://github.com/agdturner/ccg-v3d/blob/master/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Geometry) objects.
+It is advised that you load this as a [Maven project](https://maven.apache.org/) in your favourite IDE. This should make it easy to download the dependencies and run the tests. The tests provide some indication of how you might used the library for example to determine the intersection of two geometries. Another entry point is to use the [ccg-r3d](https://github.com/agdturner/ccg-r3d) - a rendering library that can visualise [V3D_Geometry](https://github.com/agdturner/ccg-v3d/blob/master/src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Geometry.java) objects.
 
 ## Details
 Slightly more lightweight geometries are in the [uk.ac.leeds.ccg.v3d.geometry.light](https://github.com/agdturner/ccg-v3d/blob/master/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/) package. These use [V3D_V](https://github.com/agdturner/ccg-v3d/blob/master/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/V3D_V.java) which have three Math_BigRational components related to each respective axis. Most classes extend [V3D_VGeometry](https://github.com/agdturner/ccg-v3d/blob/master/src/main/java/uk/ac/leeds/ccg/v3d/geometry/light/V3D_VGeometry.java) - an abstract class that holds a V3D_V "offset" which can be thought of as the offset of the centroid of the geometry:
