### 2022-09-29 5c7c390
Fixed some issues with triangle-triangle and triangle tetrahedron intersection.
