+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedron.java
@@ -884,41 +884,76 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
         }
     }
 
+    /**
+     * The intersection will be null, a point, a line segment, a triangle or a
+     * quadrilateral.
+     */
     @Override
     public V3D_FiniteGeometry getIntersection(V3D_Plane p, int oom) {
-        // The intersection will be null, a point, a line segment or a triangle...
         V3D_FiniteGeometry pqri = getPqr().getIntersection(p, oom);
         if (pqri == null) {
             V3D_FiniteGeometry psqi = getPsq().getIntersection(p, oom);
             if (psqi == null) {
                 return null;
+            } else if (psqi instanceof V3D_Point psqip) {
+                // psqip must be the point s!
+                return psqip;
+            } else if (psqi instanceof V3D_LineSegment psqil) {
+                /**
+                 * There will also be a line segment of qsr and spr that
+                 * together form a triangle.
+                 */
+                V3D_LineSegment qsril = (V3D_LineSegment) getQsr().getIntersection(p, oom);
+                V3D_LineSegment spril = (V3D_LineSegment) getSpr().getIntersection(p, oom);
+                return V3D_Triangle.getGeometry(psqil, qsril, spril, oom);
             } else {
-                return psqi;
+                // Triangle
+                return (V3D_Triangle) psqi;
             }
-        } else {
-            if (pqri instanceof V3D_Point) {
-                return pqri;
-            } else if (pqri instanceof V3D_LineSegment pqril) {
-                V3D_FiniteGeometry psqi = getPsq().getIntersection(p, oom);
-                if (psqi == null) {
-                    V3D_FiniteGeometry qsri = getQsr().getIntersection(p, oom);
-                    if (qsri instanceof V3D_Point qsrip) {
-                        return new V3D_Triangle(pqril, qsrip);
-                    } else if (qsri instanceof V3D_LineSegment qsril) {
-                        return V3D_Triangle.getGeometry(pqril, qsril);
-                    } else {
-                        return qsri;
-                    }
-                } else if (psqi instanceof V3D_Point psqip) {
-                    return new V3D_Triangle(pqril, psqip);
-                } else if (psqi instanceof V3D_LineSegment psqil) {
-                    return V3D_Triangle.getGeometry(pqril, psqil);
+        } else if (pqri instanceof V3D_Point pqrip) {
+            V3D_FiniteGeometry psqi = getPsq().getIntersection(p, oom);
+            if (psqi == null) {
+                return pqrip;
+            } else if (psqi instanceof V3D_Point psqip) {
+                V3D_FiniteGeometry qsri = getQsr().getIntersection(p, oom);
+                if (qsri == null) {
+                    return psqip;
                 } else {
-                    return psqi;
+                    return V3D_Triangle.getGeometry((V3D_LineSegment) qsri, psqip);
+                }
+            } else if (psqi instanceof V3D_LineSegment psqil) {
+                V3D_FiniteGeometry qsri = getQsr().getIntersection(p, oom);
+                if (qsri instanceof V3D_Point qsrip) {
+                    return psqil;
+                } else {
+                    //V3D_LineSegment spri = (V3D_LineSegment) getSpr().getIntersection(p, oom);
+                    return V3D_Triangle.getGeometry(psqil, (V3D_LineSegment) qsri);
                 }
             } else {
-                return pqri;
+                // Triangle
+                return (V3D_Triangle) psqi;
             }
+        } else if (pqri instanceof V3D_LineSegment pqril) {
+            V3D_FiniteGeometry psqi = getPsq().getIntersection(p, oom);
[…]
