+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java
@@ -16,6 +16,7 @@
 package uk.ac.leeds.ccg.v3d.geometry;
 
 import java.math.BigDecimal;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Objects;
@@ -319,6 +320,27 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
 //                return true;
 //            }
 //        }
+        return false;
+    }
+
+    /**
+     * @param p The point to check if it is in alignment.
+     * @param oom The order of magnitude for the precision.
+     * @return {
+     * @ceode true} iff p is aligned with this.
+     */
+    public boolean isAligned(V3D_Point p, int oom) {
+        V3D_Vector n = this.p.getN(oom);
+        V3D_Plane lp = new V3D_Plane(e, offset, this.p.p, this.p.q, this.p.p.add(n, oom), false);
+        if (lp.isOnSameSide(p, this.p.getR(), oom)) {
+            V3D_Plane lq = new V3D_Plane(e, offset, this.p.q, this.p.r, this.p.q.add(n, oom), false);
+            if (lq.isOnSameSide(p, this.p.getP(), oom)) {
+                V3D_Plane lr = new V3D_Plane(e, offset, this.p.r, this.p.p, this.p.r.add(n, oom), false);
+                if (lr.isOnSameSide(p, this.p.getQ(), oom)) {
+                    return true;
+                }
+            }
+        }
         return false;
     }
 
@@ -783,8 +805,8 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
 
     /**
      * Computes and returns the intersection between {@code this} and {@code t}.
-     * The intersection could be: null, a point, a line segment, a triangle, a
-     * quadrangle, a convex pentagon, or a convex hexagon.
+     * The intersection could be: null, a point, a line segment, a triangle, or
+     * a V3D_ConvexHullCoplanar (with 4, 5, or 6 sides).
      *
      * @param t The triangle to test for intersection with this.
      * @param oom The Order of Magnitude for the precision.
@@ -1132,7 +1154,7 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
      * Used in intersecting two triangles to give the overall intersection. If
      * l1, l2 and l3 are equal then the line segment is returned. If there are 3
      * unique points then a triangle is returned. If there are 4 or more unique
-     * points, then contiguous coplanar triangles are returned.
+     * points, then a V3D_ConvexHullCoplanar is returned.
      *
      *
      * @param l1 A line segment.
@@ -1163,182 +1185,233 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
         } else if (n == 3) {
             Iterator<V3D_Point> ite = points.iterator();
             return getGeometry(ite.next(), ite.next(), ite.next());
-        } else if (n == 4) {
-            V3D_Triangle t1;
-            V3D_Triangle t2;
-            // Case: quadrangle (closed polygon with 4 sides)
-            V3D_Point illl2 = (V3D_Point) l1.getIntersection(l2, oom);
-            V3D_Point illl3 = (V3D_Point) l1.getIntersection(l3, oom);
-            V3D_Point il2l3 = (V3D_Point) l2.getIntersection(l3, oom);
-            if (illl2 == null) {
-                V3D_Point op1 = l1.l.getOtherPoint(illl3);
-                V3D_Point op2 = l2.l.getOtherPoint(il2l3);
-                t1 = new V3D_Triangle(op1, op2, l3p);
-                t2 = new V3D_Triangle(l3p, l3q, op2);
[…]
