### 2022-09-30 0f52cdf
Fixed some issues with triangle-triangle and triangle tetrahedron intersection.
