### 2022-10-03 4cd057b
Added controls for setting precision and rounding in intersection calculations.
