+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedrons.java
@@ -1,6 +1,8 @@
 package uk.ac.leeds.ccg.v3d.geometry;
 
+import java.io.Serializable;
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Iterator;
@@ -13,7 +15,8 @@ import uk.ac.leeds.ccg.math.number.Math_BigRational;
  * @author Andy Turner
  * @version 1.0
  */
-public class V3D_Tetrahedrons extends V3D_FiniteGeometry implements V3D_Volume {
+public class V3D_Tetrahedrons extends V3D_FiniteGeometry implements V3D_Volume,
+        Serializable {
 
     private static final long serialVersionUID = 1L;
 
@@ -90,11 +93,11 @@ public class V3D_Tetrahedrons extends V3D_FiniteGeometry implements V3D_Volume {
      * @return The area of the triangle (rounded).
      */
     @Override
-    public BigDecimal getArea(int oom) {
+    public BigDecimal getArea(int oom, RoundingMode rm) {
         BigDecimal r = BigDecimal.ZERO;
         Iterator<V3D_Tetrahedron> ite = tetrahedrons.iterator();
         while (ite.hasNext()) {
-            r = r.add(ite.next().getArea(oom));
+            r = r.add(ite.next().getArea(oom, rm));
         }
         return r;
     }
@@ -103,11 +106,11 @@ public class V3D_Tetrahedrons extends V3D_FiniteGeometry implements V3D_Volume {
      * @param oom The Order of Magnitude for the precision of the calculation.
      */
     @Override
-    public BigDecimal getVolume(int oom) {
+    public BigDecimal getVolume(int oom, RoundingMode rm) {
         BigDecimal r = BigDecimal.ZERO;
         Iterator<V3D_Tetrahedron> ite = tetrahedrons.iterator();
         while (ite.hasNext()) {
-            r = r.add(ite.next().getVolume(oom));
+            r = r.add(ite.next().getVolume(oom, rm));
         }
         return r;
     }
@@ -118,47 +121,47 @@ public class V3D_Tetrahedrons extends V3D_FiniteGeometry implements V3D_Volume {
 //    }
 
     @Override
-    public BigDecimal getDistance(V3D_Point p, int oom) {
+    public BigDecimal getDistance(V3D_Point p, int oom, RoundingMode rm) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public Math_BigRational getDistanceSquared(V3D_Point p, int oom) {
+    public Math_BigRational getDistanceSquared(V3D_Point p, int oom, RoundingMode rm) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public BigDecimal getDistance(V3D_Line l, int oom) {
+    public BigDecimal getDistance(V3D_Line l, int oom, RoundingMode rm) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public BigDecimal getDistance(V3D_LineSegment l, int oom) {
+    public BigDecimal getDistance(V3D_LineSegment l, int oom, RoundingMode rm) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
-    public boolean isIntersectedBy(V3D_Point p, int oom) {
+    public boolean isIntersectedBy(V3D_Point p, int oom, RoundingMode rm) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 
     @Override
[…]
