### 2022-10-05 147d282
Change ray and line segment to not inherit from line. Provided simplified toString methods.
