+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Plane.java
@@ -178,9 +178,9 @@ public class V3D_Plane extends V3D_Geometry
         for (V3D_Point p : points) {
             if (!V3D_Line.isCollinear(e, l, p)) {
                 //return new V3D_Plane(l.getP(oom), l.getQ(oom), p, oom);
-                return new V3D_Plane(e, 
-                        l.getP().getVector(e.oom, e.rm), 
-                        l.getQ().getVector(e.oom, e.rm), 
+                return new V3D_Plane(e,
+                        l.getP().getVector(e.oom, e.rm),
+                        l.getQ().getVector(e.oom, e.rm),
                         p.getVector(e.oom, e.rm));
             }
         }
@@ -199,8 +199,8 @@ public class V3D_Plane extends V3D_Geometry
         for (V3D_Point p : points) {
             if (!V3D_Line.isCollinear(e, l, p)) {
                 //return new V3D_Plane(l.getP(oom), l.getQ(oom), p, oom);
-                return new V3D_Plane(e, V3D_Vector.ZERO, 
-                        l.getP().getVector(e.oom, e.rm), 
+                return new V3D_Plane(e, V3D_Vector.ZERO,
+                        l.getP().getVector(e.oom, e.rm),
                         l.getQ().getVector(e.oom, e.rm),
                         p.getVector(e.oom, e.rm));
             }
@@ -419,7 +419,8 @@ public class V3D_Plane extends V3D_Geometry
 
     @Override
     public String toString() {
-        return toString("");
+        //return toString("");
+        return toStringSimple("");
     }
 
     /**
@@ -435,6 +436,15 @@ public class V3D_Plane extends V3D_Geometry
 
     /**
      * @param pad A padding of spaces.
+     * @return A description of this.
+     */
+    public String toStringSimple(String pad) {
+        return pad + this.getClass().getSimpleName() + "("
+                + toStringFieldsSimple("") + ")";
+    }
+
+    /**
+     * @param pad A padding of spaces.
      * @return A description of the fields.
      */
     @Override
@@ -446,6 +456,18 @@ public class V3D_Plane extends V3D_Geometry
                 + pad + "q=" + q.toString(pad) + "\n"
                 + pad + ",\n"
                 + pad + "r=" + r.toString(pad);
+    }
+    
+    /**
+     * @param pad A padding of spaces.
+     * @return A description of the fields.
+     */
+    @Override
+    protected String toStringFieldsSimple(String pad) {
+        return super.toStringFieldsSimple(pad) + ",\n"
+                + pad + "p=" + p.toStringSimple(pad) + ",\n"
+                + pad + "q=" + q.toStringSimple(pad) + ",\n"
+                + pad + "r=" + r.toStringSimple(pad);
     }
 
 //    /**
@@ -887,30 +909,32 @@ public class V3D_Plane extends V3D_Geometry
                 lp.getX(oomN2, rm).add(lv.getDX(oomN2, rm).multiply(t)),
                 lp.getY(oomN2, rm).add(lv.getDY(oomN2, rm).multiply(t)),
                 lp.getZ(oomN2, rm).add(lv.getDZ(oomN2, rm).multiply(t)));
-        // Check if r is on the line.
-        if (!l.isIntersectedBy(res, oom, rm)) {
-            System.out.println("Not on line! - l.isIntersectedBy(r, oom, rm)");
-        }
-        if (!isIntersectedBy(res, oom, rm)) {
-            System.out.println("Not on plane! - !isIntersectedBy(r, oom, rm)");
-            // Check side of line
-            if (isOnSameSide(res, lp, oom, rm)) {
-                System.out.println("isOnSameSide(res, lp, oom, rm)");
-            } else {
-                System.out.println("!isOnSameSide(res, lp, oom, rm)");
[…]
