+++ src/test/java/uk/ac/leeds/ccg/v3d/V3D_Test.java
@@ -54,6 +54,27 @@ public abstract class V3D_Test {
     public static final Math_BigRational P9999 = Math_BigRational.valueOf(9999);
     public static final Math_BigRational P10000 = Math_BigRational.valueOf(10000);
     public static final Math_BigRational P10001 = Math_BigRational.valueOf(10001);
+    public static final Math_BigRational P1E5 = P10.multiply(P10000);
+    public static final Math_BigRational P1E6 = P10.multiply(P1E5);
+    public static final Math_BigRational P1E7 = P10.multiply(P1E6);
+    public static final Math_BigRational P1E8 = P10.multiply(P1E7);
+    public static final Math_BigRational P1E9 = P10.multiply(P1E8);
+    public static final Math_BigRational P1E10 = P10.multiply(P1E9);
+    public static final Math_BigRational P1E11 = P10.multiply(P1E10);
+    public static final Math_BigRational P1E12 = P10.multiply(P1E11);
+    public static final Math_BigRational P1E13 = P10.multiply(P1E12);
+    public static final Math_BigRational P0_1 = Math_BigRational.ONE.divide(Math_BigRational.TEN);
+    public static final Math_BigRational P0_1E2 = P0_1.multiply(P0_1);
+    public static final Math_BigRational P0_1E3 = P0_1.multiply(P0_1E2);
+    public static final Math_BigRational P0_1E4 = P0_1.multiply(P0_1E3);
+    public static final Math_BigRational P0_1E5 = P0_1.multiply(P0_1E4);
+    public static final Math_BigRational P0_1E6 = P0_1.multiply(P0_1E5);
+    public static final Math_BigRational P0_1E7 = P0_1.multiply(P0_1E6);
+    public static final Math_BigRational P0_1E8 = P0_1.multiply(P0_1E7);
+    public static final Math_BigRational P0_1E9 = P0_1.multiply(P0_1E8);
+    public static final Math_BigRational P0_1E10 = P0_1.multiply(P0_1E9);
+    public static final Math_BigRational P0_1E11 = P0_1.multiply(P0_1E10);
+    public static final Math_BigRational P0_1E12 = P0_1.multiply(P0_1E11);
     public static final Math_BigRational N0 = P0.negate();
     public static final Math_BigRational N1 = P1.negate();
     public static final Math_BigRational N2 = P2.negate();
@@ -74,6 +95,26 @@ public abstract class V3D_Test {
     public static final Math_BigRational N9999 = P9999.negate();
     public static final Math_BigRational N10000 = P10000.negate();
     public static final Math_BigRational N10001 = P10001.negate();
+    public static final Math_BigRational N1E5 = P1E5.negate();
+    public static final Math_BigRational N1E6 = P1E6.negate();
+    public static final Math_BigRational N1E7 = P1E7.negate();
+    public static final Math_BigRational N1E8 = P1E8.negate();
+    public static final Math_BigRational N1E9 = P1E9.negate();
+    public static final Math_BigRational N1E10 = P1E10.negate();
+    public static final Math_BigRational N1E11 = P1E11.negate();
+    public static final Math_BigRational N1E12 = P1E12.negate();
+    public static final Math_BigRational N0_1 = P0_1.negate();
+    public static final Math_BigRational N0_1E2 = P0_1E2.negate();
+    public static final Math_BigRational N0_1E3 = P0_1E3.negate();
+    public static final Math_BigRational N0_1E4 = P0_1E4.negate();
+    public static final Math_BigRational N0_1E5 = P0_1E5.negate();
+    public static final Math_BigRational N0_1E6 = P0_1E6.negate();
[…]
