+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Rectangle.java
@@ -17,7 +17,6 @@ package uk.ac.leeds.ccg.v3d.geometry;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.util.Objects;
 import uk.ac.leeds.ccg.math.number.Math_BigRational;
 import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
 import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
@@ -192,6 +191,13 @@ public class V3D_Rectangle extends V3D_Triangle implements V3D_Face {
         this.s = s2.rel;
     }
 
+    @Override
+    public String toString() {
+        //return toString("");
+        return toStringSimple("");
+    }
+
+    @Override
     public String toString(String pad) {
         return this.getClass().getSimpleName() + "\n"
                 + pad + "(\n"
@@ -199,15 +205,26 @@ public class V3D_Rectangle extends V3D_Triangle implements V3D_Face {
                 + pad + ")";
     }
 
-    @Override
-    protected String toStringFields(String pad) {
-        return super.toStringFields(pad) + "\n"
-                + pad + ",\n"
-                + pad + "s=" + s.toString(pad);
+    public String toStringSimple(String pad) {
+        return pad + this.getClass().getSimpleName() + "("
+                + toStringFieldsSimple("") + ")";
     }
 
     @Override
-    public V3D_Point[] getPoints() {
+    protected String toStringFields(String pad) {
+        return pad + "p=" + this.p.toString(pad) + "\n"
+                + pad + ",\n"
+                + pad + "s=" + this.s.toString(pad) + "\n";
+    }
+
+    @Override
+    protected String toStringFieldsSimple(String pad) {
+        return pad + "p=" + this.p.toStringSimple(pad) + ",\n"
+                + pad + "s=" + s.toStringSimple(pad);
+    }
+
+    @Override
+    public V3D_Point[] getPoints(int oom, RoundingMode rm) {
         V3D_Point[] re = new V3D_Point[4];
         re[0] = this.p.getP();
         re[1] = this.p.getQ();
@@ -231,7 +248,7 @@ public class V3D_Rectangle extends V3D_Triangle implements V3D_Face {
     }
 
     @Override
-    public V3D_Envelope getEnvelope() {
+    public V3D_Envelope getEnvelope(int oom, RoundingMode rm) {
         if (en == null) {
             en = new V3D_Envelope(e, p.getP(), p.getQ(), p.getR(), getS());
         }
@@ -426,7 +443,7 @@ public class V3D_Rectangle extends V3D_Triangle implements V3D_Face {
      */
     @Override
     public boolean isIntersectedBy(V3D_LineSegment l, int oom, RoundingMode rm) {
-        if (getEnvelope().isIntersectedBy(l.getEnvelope())) {
+        if (getEnvelope(oom, rm).isIntersectedBy(l.getEnvelope(oom, rm))) {
             if (p.isIntersectedBy(l, oom, rm)) {
                 V3D_Geometry pli = p.getIntersection(l, oom, rm);
                 if (pli instanceof V3D_Point plip) {
@@ -481,7 +498,7 @@ public class V3D_Rectangle extends V3D_Triangle implements V3D_Face {
                 if (i2 == null) {
                     return i1;
                 } else {
-                    return join(i1, i2);
+                    return join(i1, i2, oom, rm);
                 }
             }
         } else {
@@ -665,7 +682,7 @@ public class V3D_Rectangle extends V3D_Triangle implements V3D_Face {
      * @return A point or line segment.
      */
     protected V3D_FiniteGeometry join(V3D_FiniteGeometry pointOrLineSegment1,
[…]

[…]
