+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedron.java
@@ -136,17 +136,18 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
      * @param r Used to set {@link #r}.
      * @param s Used to set {@link #s}.
      */
-    public V3D_Tetrahedron(V3D_Point p, V3D_Point q, V3D_Point r, V3D_Point s) {
+    public V3D_Tetrahedron(V3D_Point p, V3D_Point q, V3D_Point r, V3D_Point s,
+            int oom, RoundingMode rm) {
         super(p.e, p.offset);
         this.p = new V3D_Vector(p.rel);
         V3D_Point qp = new V3D_Point(q);
-        qp.setOffset(offset);
+        qp.setOffset(offset, oom, rm);
         this.q = qp.rel;
         V3D_Point rp = new V3D_Point(r);
-        rp.setOffset(offset);
+        rp.setOffset(offset, oom, rm);
         this.r = rp.rel;
         V3D_Point sp = new V3D_Point(s);
-        sp.setOffset(offset);
+        sp.setOffset(offset, oom, rm);
         this.s = sp.rel;
     }
 
@@ -157,8 +158,8 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
      * @param p Used to set {@link #p}, {@link #e} and {@link #offset}.
      * @param t Used to set {@link #q}, {@link #r} and {@link #s}.
      */
-    public V3D_Tetrahedron(V3D_Point p, V3D_Triangle t) {
-        this(p, t.p.getP(), t.p.getQ(), t.p.getR());
+    public V3D_Tetrahedron(V3D_Point p, V3D_Triangle t, int oom, RoundingMode rm) {
+        this(p, t.p.getP(), t.p.getQ(), t.p.getR(), oom, rm);
     }
 
     @Override
@@ -200,9 +201,9 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
         if (en == null) {
 //            en = pqr.getEnvelope(oom).union(qsr.getEnvelope(oom));
             en = getP().getEnvelope(oom, rm)
-                    .union(getQ().getEnvelope(oom, rm))
-                    .union(getR().getEnvelope(oom, rm))
-                    .union(getS().getEnvelope(oom, rm));
+                    .union(getQ().getEnvelope(oom, rm), oom, rm)
+                    .union(getR().getEnvelope(oom, rm), oom, rm)
+                    .union(getS().getEnvelope(oom, rm), oom, rm);
         }
         return en;
     }
@@ -435,7 +436,7 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
                 } else if (qsri instanceof V3D_Point qsrip) {
                     V3D_FiniteGeometry spri = getSpr().getIntersection(l, oom, rm);
                     if (spri == null) {
-                        return V3D_LineSegment.getGeometry(qsrip, pqrip);
+                        return V3D_LineSegment.getGeometry(qsrip, pqrip, oom, rm);
                     } else if (spri instanceof V3D_Point sprip) {
                         return V3D_Triangle.getGeometry(qsrip, pqrip, sprip, oom, rm);
                     } else {
@@ -449,7 +450,7 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
                 if (qsri == null) {
                     V3D_FiniteGeometry spri = getSpr().getIntersection(l, oom, rm);
                     if (spri == null) {
-                        return V3D_LineSegment.getGeometry(psqip, pqrip);
+                        return V3D_LineSegment.getGeometry(psqip, pqrip, oom, rm);
                     } else if (spri instanceof V3D_Point sprip) {
[…]
