+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_TetrahedronTest.java
@@ -59,8 +59,10 @@ public class V3D_TetrahedronTest extends V3D_Test {
     @Test
     public void testToString_0args() {
         System.out.println("toString");
+        int oom = -3;
+        RoundingMode rm = RoundingMode.HALF_UP;
         V3D_Tetrahedron instance = new V3D_Tetrahedron(pP0P0P0, pP1P0P0,
-                pP0P1P0, pP0P1P1);
+                pP0P1P0, pP0P1P1, oom, rm);
         String expResult = "V3D_Tetrahedron\n"
                 + "(\n"
                 + " oom=-3\n"
@@ -110,9 +112,11 @@ public class V3D_TetrahedronTest extends V3D_Test {
     @Test
     public void testToString_String() {
         System.out.println("toString");
+        int oom = -3;
+        RoundingMode rm = RoundingMode.HALF_UP;
         String pad = "";
         V3D_Tetrahedron instance = new V3D_Tetrahedron(pP0P0P0, pP1P0P0,
-                pP0P1P0, pP0P1P1);
+                pP0P1P0, pP0P1P1, oom, rm);
         String expResult = "V3D_Tetrahedron\n"
                 + "(\n"
                 + " oom=-3\n"
@@ -162,9 +166,11 @@ public class V3D_TetrahedronTest extends V3D_Test {
     @Test
     public void testToStringFields() {
         System.out.println("toStringFields");
+        int oom = -3;
+        RoundingMode rm = RoundingMode.HALF_UP;
         String pad = "";
         V3D_Tetrahedron instance = new V3D_Tetrahedron(pP0P0P0, pP1P0P0,
-                pP0P1P0, pP0P1P1);
+                pP0P1P0, pP0P1P1, oom, rm);
         String expResult = "oom=-3\n"
                 + ",\n"
                 + "offset=V3D_Vector\n"
@@ -214,8 +220,8 @@ public class V3D_TetrahedronTest extends V3D_Test {
         int oom = -3;
         RoundingMode rm = RoundingMode.HALF_UP;
         V3D_Tetrahedron instance = new V3D_Tetrahedron(pP0P0P0, pP1P0P0,
-                pP0P1P0, pP0P1P1);
-        V3D_Envelope expResult = new V3D_Envelope(e, pP0P0P0, pP1P0P0,
+                pP0P1P0, pP0P1P1, oom, rm);
+        V3D_Envelope expResult = new V3D_Envelope(e, oom, rm, pP0P0P0, pP1P0P0,
                 pP0P1P0, pP0P1P1);
         V3D_Envelope result = instance.getEnvelope(oom, rm);
         assertEquals(expResult, result);
@@ -227,8 +233,10 @@ public class V3D_TetrahedronTest extends V3D_Test {
     @Test
     public void testGetP() {
         System.out.println("getP");
+        int oom = -3;
+        RoundingMode rm = RoundingMode.HALF_UP;
         V3D_Tetrahedron instance = new V3D_Tetrahedron(pP0P0P0, pP1P0P0,
-                pP0P1P0, pP0P1P1);
+                pP0P1P0, pP0P1P1, oom, rm);
         V3D_Point expResult = pP0P0P0;
         V3D_Point result = instance.getP();
         assertEquals(expResult, result);
@@ -240,8 +248,10 @@ public class V3D_TetrahedronTest extends V3D_Test {
[…]
