### 2022-10-06 2a94afc
Added controls for precision and rounding. Most tests now pass, some are commented out or disabled.
