### 2022-10-07 c348c72
Increased precision of dot and cross product calculations.
