### 2022-10-08 f200503
Improved ray intersection.
