+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Plane.java
@@ -625,11 +625,20 @@ public class V3D_Plane extends V3D_Geometry
      * @return The normal vector.
      */
     public V3D_Vector getN(int oom, RoundingMode rm) {
+//        if (n == null) {
+//            if (isQAtOrigin(oom, rm)) {
+//                n = getQRV(oom, rm).getCrossProduct(getRPV(oom, rm), oom, rm);
+//            } else {
+//                n = getPQV(oom, rm).getCrossProduct(getQRV(oom, rm), oom, rm);
+//            }
+//        }
+//        //return rotate(n, bI, theta);
+//        return n;
         if (n == null) {
             if (isQAtOrigin(oom, rm)) {
-                n = getQRV(oom, rm).getCrossProduct(getRPV(oom, rm), oom, rm);
+                return getQRV(oom, rm).getCrossProduct(getRPV(oom, rm), oom, rm);
             } else {
-                n = getPQV(oom, rm).getCrossProduct(getQRV(oom, rm), oom, rm);
+                return getPQV(oom, rm).getCrossProduct(getQRV(oom, rm), oom, rm);
             }
         }
         //return rotate(n, bI, theta);
@@ -854,8 +863,8 @@ public class V3D_Plane extends V3D_Geometry
         if (this.isIntersectedBy(lq, oom, rm)) {
             return lq;
         }
-        int oomN3 = oom - 3;
-        V3D_Vector lv = l.getV(oomN3, rm);
+        int oomN6 = oom - 6;
+        V3D_Vector lv = l.getV(oomN6, rm);
         Math_BigRational[][] m = new Math_BigRational[4][4];
         m[0][0] = Math_BigRational.ONE;
         m[0][1] = Math_BigRational.ONE;
@@ -876,18 +885,18 @@ public class V3D_Plane extends V3D_Geometry
         V3D_Point tp = getP();
         V3D_Point tq = getQ();
         V3D_Point tr = getR();
-        m[1][0] = tp.getX(oomN3, rm);
-        m[1][1] = tq.getX(oomN3, rm);
-        m[1][2] = tr.getX(oomN3, rm);
-        m[1][3] = lp.getX(oomN3, rm);
-        m[2][0] = tp.getY(oomN3, rm);
-        m[2][1] = tq.getY(oomN3, rm);
-        m[2][2] = tr.getY(oomN3, rm);
-        m[2][3] = lp.getY(oomN3, rm);
-        m[3][0] = tp.getZ(oomN3, rm);
-        m[3][1] = tq.getZ(oomN3, rm);
-        m[3][2] = tr.getZ(oomN3, rm);
-        m[3][3] = lp.getZ(oomN3, rm);
+        m[1][0] = tp.getX(oomN6, rm);
+        m[1][1] = tq.getX(oomN6, rm);
+        m[1][2] = tr.getX(oomN6, rm);
+        m[1][3] = lp.getX(oomN6, rm);
+        m[2][0] = tp.getY(oomN6, rm);
+        m[2][1] = tq.getY(oomN6, rm);
+        m[2][2] = tr.getY(oomN6, rm);
+        m[2][3] = lp.getY(oomN6, rm);
+        m[3][0] = tp.getZ(oomN6, rm);
+        m[3][1] = tq.getZ(oomN6, rm);
+        m[3][2] = tr.getZ(oomN6, rm);
+        m[3][3] = lp.getZ(oomN6, rm);
         Math_Matrix_BR numm = new Math_Matrix_BR(m);
         m[0][0] = Math_BigRational.ONE;
         m[0][1] = Math_BigRational.ONE;
@@ -905,24 +914,35 @@ public class V3D_Plane extends V3D_Geometry
 //        m[3][1] = q.getZ(oom);
 //        m[3][2] = r.getZ(oom);
[…]
