### 2022-10-09 42fa679
Changed to use rays instead of lines as appropriate.
