### 2022-10-17 4dd2b63
Major simplification of V3D_Envelope.
