+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java
@@ -29,33 +29,33 @@ import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
  * For representing and processing triangles in 3D. A triangle has a non-zero
  * area. The corner points are {@link #pl}, {@link #q} and {@link #r}. The
  * following depicts a generic triangle {@code
-                          pq
-  pl *- - - - - - - - - - - + - - - - - - - - - - -* q
-     \~                   mpq                   ~/
-      \  ~                 |                 ~  /
-       \    ~              |              ~    /
-        \      ~           |           ~      /
-         \        ~        |        ~        /
-          \          ~     |     ~          /
-           \            ~  |  ~            /
-       -n  -n  -n  -n  -n  c  n  n  n  n  n  normal heading out from the page.
-             \          ~  |  ~          /
-              \      ~     |     ~      /
-               \  ~        |        ~  /
-                + mrp      |      mqr +
-             rp  \         |         /  qr
-                  \        |        /
-                   \       |       /
-                    \      |      /
-                     \     |     /
-                      \    |    /
-                       \   |   /
-                        \  |  /
-                         \ | /
-                          \|/
-                           *
-                           r
- }
+ * pq
+ * pl *- - - - - - - - - - - + - - - - - - - - - - -* q
+ * \~                   mpq                   ~/
+ * \  ~                 |                 ~  /
+ * \    ~              |              ~    /
+ * \      ~           |           ~      /
+ * \        ~        |        ~        /
+ * \          ~     |     ~          /
+ * \            ~  |  ~            /
+ * -n  -n  -n  -n  -n  c  n  n  n  n  n  normal heading out from the page.
+ * \          ~  |  ~          /
+ * \      ~     |     ~      /
+ * \  ~        |        ~  /
+ * + mrp      |      mqr +
+ * rp  \         |         /  qr
+ * \        |        /
+ * \       |       /
+ * \      |      /
+ * \     |     /
+ * \    |    /
+ * \   |   /
+ * \  |  /
+ * \ | /
+ * \|/
+ *
+ * r
+ * }
  * The mid points of the triangle edges may or may not be possible to pin point.
  * Likewise the centroid which is at the intersection of the line segments from
  * these midpoints to the opposite corner.
@@ -69,9 +69,9 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
 
     public V3D_Plane pl;
 
-    public final V3D_Vector p;
-    public final  V3D_Vector q;
-    public final  V3D_Vector r;
+    public V3D_Vector p;
+    public V3D_Vector q;
+    public V3D_Vector r;
 //    /**
 //     * For storing the line segment from {@link #pl} to {@link #q}.
 //     */
@@ -106,6 +106,7 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
 //     * The centroid.
 //     */
 //    private V3D_Point c;
+
     /**
      * Creates a new triangle.
      *
@@ -196,7 +197,6 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
 //                lpq.getQ(oom, rm).getVector(oom, rm),
 //                lqr.getQ(oom, rm).getVector(oom, rm));
 //    }
-
     /**
      * Creates a new instance.
      *
@@ -218,54 +218,54 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
     public final V3D_Point getP() {
         return new V3D_Point(e, offset, p);
     }
-    
+
     /**
      * @return A new point based on {@link #q} and {@link #offset}.
      */
     public final V3D_Point getQ() {
         return new V3D_Point(e, offset, q);
     }
-    
+
     /**
[…]
