### 2022-10-17 #4283 @agdturner
Hi @arfon, everyone. Thanks for the message and for the reviews which seem fair. I am making slow (and hopefully steady progress) in addressing reviews and refactoring the code to make the development more sustainable. I developed some rendering code in a very similarly named repository ([ccg-r3d](https://github.com/agdturner/ccg-r3d)) and had some success in visualising a Utah Teapot (a sort of "Hello World" program equivalent for 3D model graphics rendering). I'm currently working on rendering models from different camera angles and distances. There have been some structural changes in the refactoring work, so I want to update the paper to reflect these once I have this part working. To try to detail those things that you can do with the code I've been developing that are not easy or are not possible with other 3D Modelling code I am going to spend some time exploring other open source software as suggested whilst developing science use case scenarios. It may take me some time to complete the work that the reviewers have asked for...
