+++ README.md
@@ -1,9 +1,9 @@
 # [ccg-v3d](https://github.com/agdturner/ccg-v3d)
 
 ## Description
-A three-dimensional ([3D](https://en.wikipedia.org/wiki/Euclidean_space)) [Euclidean geometry](https://en.wikipedia.org/wiki/Euclidean_geometry) Java library. The library is modularised, based on [openJDK 17](https://openjdk.java.net/projects/jdk/17/), and only depends on [ccg-math](https://github.com/agdturner/ccg-math) - a modularised mathematics library.
+A three-dimensional ([3D](https://en.wikipedia.org/wiki/Euclidean_space)) [Euclidean geometry](https://en.wikipedia.org/wiki/Euclidean_geometry) Java library. The library is modularised, based on [openJDK 17](https://openjdk.java.net/projects/jdk/17/), and has very few lightweight dependencies.
 
-Point positions in space are defined using [cartesian](https://en.wikipedia.org/wiki/Cartesian_coordinate_system) coordinates with [orthogonal](https://en.wikipedia.org/wiki/Orthogonality) X, Y and Z axes that meet at the origin - a point <x,y,z> where x=y=z=0. Coordinates are either stored as [Math_BigRational](https://github.com/agdturner/ccg-math/blob/master/src/main/java/uk/ac/leeds/ccg/math/number/Math_BigRational.java) numbers - a subset of [rational numbers](https://en.wikipedia.org/wiki/Rational_number), or [Math_BigRationalSqrt](https://github.com/agdturner/ccg-math/blob/master/src/main/java/uk/ac/leeds/ccg/math/number/Math_BigRationalSqrt.java) - that also support non-rational square roots. Whilst there are still limits on precision, calculations can be done to a user given [Order of Magnitude](https://en.wikipedia.org/wiki/Order_of_magnitude) (OOM) precision. The coordinate system is "right handed", so if X increases to the right of this page, Y increases towards the top of this page, then Z increases out from the page, (see [Orientation](https://en.wikipedia.org/wiki/Orientation_(vector_space)) and the [note on the choice of right over left handedness](#handedness) for details of handedness. This library originally began developement with a view to supporting 3D Geography and Earth Science applications, but it may support other scientific and non-scientific applications. Notes about the [origin](#origin), and [geography](#geography) are also provided below.
[…]
