### 2022-10-19 0c90e0b
Removed many isIntersected by methods as often to know if something intersects it is necessary to compute the intersection.
