### 2022-10-20 2b6cdb0
Removed many isIntersected by methods as often to know if something intersects it is necessary to compute the intersection.
