### 2022-10-20 7cc1fad
Simplified interfaces. Reworked some intersection code.
