+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_PointsCollinear.java
@@ -26,8 +26,7 @@ import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
  * @author Andy Turner
  * @version 1.0
  */
-public class V3D_PointsCollinear extends V3D_FiniteGeometry implements
-        V3D_IntersectionAndDistanceCalculations {
+public class V3D_PointsCollinear extends V3D_FiniteGeometry {
 
     private static final long serialVersionUID = 1L;
 
@@ -211,16 +210,6 @@ public class V3D_PointsCollinear extends V3D_FiniteGeometry implements
 
     @Override
     public V3D_Geometry getIntersection(V3D_Ray r, int oom, RoundingMode rm) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public BigDecimal getDistance(V3D_Ray r, int oom, RoundingMode rm) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    @Override
-    public Math_BigRational getDistanceSquared(V3D_Ray l, int oom, RoundingMode rm) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
 }
