+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_IntersectionAndDistanceCalculations.java
@@ -1,328 +1,0 @@
-/*
- * Copyright 2020 Andy Turner, University of Leeds.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package uk.ac.leeds.ccg.v3d.geometry;
-
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import uk.ac.leeds.ccg.math.number.Math_BigRational;
-
-/**
- * V3D_FiniteGeometry
- *
- * @author Andy Turner
- * @version 1.0
- */
-public interface V3D_IntersectionAndDistanceCalculations {
-    
-    /**
-     * Identify if the geometry is intersected by point {@code p}.
-     *
-     * @param p The point to test for intersection with.
-     * @param oom The Order of Magnitude for the precision.
-     * @param rm The RoundingMode if rounding is needed.
-     * @return {@code true} iff the geometry is intersected by {@code p}.
-     */
-    public abstract boolean isIntersectedBy(V3D_Point p, int oom, 
-            RoundingMode rm);
-
-//    /**
-//     * Identify if the geometry is intersected by line {@code l}.
-//     *
-//     * @param l The line to test for intersection with.
-//     * @param oom The Order of Magnitude for the precision.
-//     * @param rm The RoundingMode if rounding is needed.
-//     * @return {@code true} iff the geometry is intersected by {@code l}.
-//     */
-//    public abstract boolean isIntersectedBy(V3D_Line l, int oom, 
-//            RoundingMode rm);
-
-//    /**
-//     * Identify if the geometry is intersected by ray {@code r}.
-//     *
-//     * @param r The ray to test for intersection with.
-//     * @param oom The Order of Magnitude for the precision.
-//     * @param rm The RoundingMode if rounding is needed.
-//     * @return {@code true} iff the geometry is intersected by {@code r}.
-//     */
-//    public abstract boolean isIntersectedBy(V3D_Ray r, int oom, 
-//            RoundingMode rm);
-//
-//    /**
-//     * Identify if the geometry is intersected by line segment {@code l}.
-//     *
-//     * @param l The line segment to test for intersection with.
-//     * @param oom The Order of Magnitude for the precision.
-//     * @param rm The RoundingMode if rounding is needed.
-//     * @return {@code true} iff the geometry is intersected by {@code l}.
-//     */
-//    public abstract boolean isIntersectedBy(V3D_LineSegment l, int oom, 
-//            RoundingMode rm);
-//
-//    /**
-//     * Identify if the geometry is intersected by plane {@code p}.
-//     *
-//     * @param p The plane to test for intersection with.
-//     * @param oom The Order of Magnitude for the precision.
-//     * @param rm The RoundingMode if rounding is needed.
-//     * @return {@code true} iff the geometry is intersected by {@code p}.
-//     */
-//    public abstract boolean isIntersectedBy(V3D_Plane p, int oom, 
-//            RoundingMode rm);
-//
-//    /**
-//     * Identify if the geometry is intersected by plane {@code p}.
-//     *
-//     * @param t The triangle to test for intersection with.
-//     * @param oom The Order of Magnitude for the precision.
-//     * @param rm The RoundingMode if rounding is needed.
[…]
