+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Ray.java
@@ -20,7 +20,6 @@ import java.math.RoundingMode;
 import uk.ac.leeds.ccg.math.number.Math_BigRational;
 import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
 import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
-import static uk.ac.leeds.ccg.v3d.geometry.V3D_LineSegment.getGeometry;
 
 /**
  * 3D representation of a ray - like a line, but one that starts at a point
@@ -28,38 +27,37 @@ import static uk.ac.leeds.ccg.v3d.geometry.V3D_LineSegment.getGeometry;
  * {@link #p} and goes through the point {@link #q}. The "*" denotes a point in
  * 3D and the ray is shown as a linear feature of "e" symbols in the following
  * depiction: {@code
-                                       z
-                          y           -
-                          +          /                * pl=<x0,y0,z0>
-                          |         /                e
-                          |    z0 -/                e
-                          |       /                e
-                          |      /                e
-                          |     /               e
-                          |    /               e
-                          |   /               e
-                       y0-|  /               e
-                          | /               e
-                          |/         x1    e
-  - ----------------------|-----------|---e---|---- + x
-                         /|              e   x0
-                        / |-y1          e
-                       /  |           e
-                      /   |          e
-                 z1 -/    |         e
-                    /     |        e
-                   /      |       * q=<x1,y1,z1>
-                  /       |      e
-                 /        |     e
-                +         -    e
-               z          y   e
- }
+ *                                     z
+ *                         y           -
+ *                         +          /                * pl=<x0,y0,z0>
+ *                         |         /                e
+ *                         |    z0 -/                e
+ *                         |       /                e
+ *                         |      /                e
+ *                         |     /               e
+ *                         |    /               e
+ *                         |   /               e
+ *                      y0-|  /               e
+ *                         | /               e
+ *                         |/         x1    e
+ * - ----------------------|-----------|---e---|---- + x
+ *                        /|              e   x0
+ *                       / |-y1          e
+ *                      /  |           e
+ *                     /   |          e
+ *                z1 -/    |         e
+ *                   /     |        e
+ *                  /      |       * q=<x1,y1,z1>
+ *                 /       |      e
+ *                /        |     e
+ *               +         -    e
+ *              z          y   e
+ * }
  *
  * @author Andy Turner
  * @version 1.0
  */
-public class V3D_Ray extends V3D_Geometry
-        implements V3D_IntersectionAndDistanceCalculations {
+public class V3D_Ray extends V3D_Geometry {
 
     private static final long serialVersionUID = 1L;
 
@@ -140,7 +138,7 @@ public class V3D_Ray extends V3D_Geometry
      * @return {@code true} iff {@code r} is the same as {@code this}.
      */
     public boolean equals(V3D_Ray r, int oom, RoundingMode rm) {
-        return l.getP().equals(r.l.getP(), oom, rm) 
+        return l.getP().equals(r.l.getP(), oom, rm)
                 && l.v.isScalarMultiple(r.l.v, oom, rm);
     }
 
[…]
