### 2022-10-20 60eacca
Simplified interfaces. Reworked some intersection code.
