+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_RayTest.java
@@ -280,123 +280,6 @@ public class V3D_RayTest extends V3D_Test {
     }
 
     /**
-     * Test of getLineOfIntersection method, of class V3D_Ray.
-     */
-    @Test
-    public void testGetLineOfIntersection_V3D_Line_int() {
-        System.out.println("getLineOfIntersection");
-        int oom = -3;
-        RoundingMode rm = RoundingMode.HALF_UP;
-        V3D_Ray instance = new V3D_Ray(pP0P0P0, pP0P0P1, oom, rm);
-        V3D_Line l = new V3D_Line(pP1P0P0, pP1P1P0, oom, rm);
-        V3D_Geometry expResult = new V3D_LineSegment(pP0P0P0, pP1P0P0, oom, rm);
-        V3D_Geometry result = instance.getLineOfIntersection(l, oom, rm);
-        assertTrue(((V3D_LineSegment) expResult).equals((V3D_LineSegment) result, oom, rm));
-        // Test 2
-        instance = new V3D_Ray(pP0P0P0, pP0P1P0, oom, rm);
-        result = instance.getLineOfIntersection(l, oom, rm);
-        assertNull(result);
-    }
-
-    /**
-     * Test of getPointOfIntersection method, of class V3D_Ray.
-     */
-    @Test
-    public void testGetPointOfIntersection() {
-        System.out.println("getPointOfIntersection");
-        int oom = -3;
-        RoundingMode rm = RoundingMode.HALF_UP;
-        V3D_Point pt;
-        V3D_Ray instance;
-        V3D_Point expResult;
-        V3D_Geometry result;
-        // Test 1
-        pt = pP0P0P0;
-        instance = new V3D_Ray(pP1P0P0, pP1P1P0, oom, rm);
-        expResult = pP1P0P0;
-        result = instance.getPointOfIntersection(pt, oom, rm);
-        assertTrue(expResult.equals((V3D_Point) result, oom, rm));
-        // Test 2
-        instance = new V3D_Ray(pP1N1P0, pP1P1P0, oom, rm);
-        expResult = pP1P0P0;
-        result = instance.getPointOfIntersection(pt, oom, rm);
-        assertTrue(expResult.equals((V3D_Point) result, oom, rm));
-    }
-
-    /**
-     * Test of getDistance method, of class V3D_Ray.
-     */
-    @Test
-    public void testGetDistance_V3D_Point_int() {
-        System.out.println("getDistance");
-        int oom = -3;
-        RoundingMode rm = RoundingMode.HALF_UP;
-        V3D_Point p;
-        V3D_Ray instance;
-        BigDecimal expResult;
-        BigDecimal result;
-        // Test 1
-        p = pP0P0P0;
-        instance = new V3D_Ray(pP1P0P0, pP1P1P0, oom, rm);
-        expResult = BigDecimal.ONE;
-        result = instance.getDistance(p, oom, rm);
-        assertTrue(expResult.compareTo(result) == 0);
-    }
-
-    /**
-     * Test of getDistance method, of class V3D_Ray.
-     */
-    @Test
-    public void testGetDistance_V3D_Ray_int() {
-        System.out.println("getDistance");
-        int oom = -3;
-        RoundingMode rm = RoundingMode.HALF_UP;
-        V3D_Ray p;
-        V3D_Ray instance;
-        BigDecimal expResult;
-        BigDecimal result;
-        // Test 1
-        p = new V3D_Ray(pP0P0P0, pP1P1P0, oom, rm);
-        instance = new V3D_Ray(pP1P0P0, pP1P1P0, oom, rm);
-        expResult = new Math_BigRationalSqrt(2L, oom, rm).getSqrt(oom, rm)
[…]
