### 2022-10-22 8cc5b03
Simplified interfaces. Reworked some intersection code.
