### 2022-10-23 9bda6a1
Simplified interfaces. Reworked some intersection code.
