+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Envelope.java
@@ -16,12 +16,9 @@
 package uk.ac.leeds.ccg.v3d.geometry;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
-import java.math.MathContext;
 import java.math.RoundingMode;
 import java.util.ArrayList;
 import uk.ac.leeds.ccg.math.number.Math_BigRational;
-import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
 import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
 
 /**
@@ -93,7 +90,7 @@ public class V3D_Envelope implements Serializable {
     private static final long serialVersionUID = 1L;
 
     private V3D_Vector offset;
-    V3D_Environment e;
+    private final V3D_Environment e;
     /**
      * The minimum x-coordinate.
      */
@@ -142,6 +139,7 @@ public class V3D_Envelope implements Serializable {
      * @param points The points used to form the envelop.
      */
     public V3D_Envelope(V3D_Environment e, int oom, RoundingMode rm, V3D_Point... points) {
+        this.e = e;
         //offset = points[0].offset;
         offset = V3D_Vector.ZERO;
         int len = points.length;
@@ -214,12 +212,11 @@ public class V3D_Envelope implements Serializable {
                 new V3D_Point(e, xMax, yMax, zMax));
     }
 
-    @Override
-    public String toString() {
+    public String toString(int oom, RoundingMode rm) {
         return this.getClass().getSimpleName()
-                + "(xMin=" + xMin + ", xMax=" + xMax
-                + ", yMin=" + yMin + ", yMax=" + yMax
-                + ", zMin=" + zMin + ", zMax=" + zMax + ")";
+                + "(xMin=" + getXMin(oom, rm) + ", xMax=" + getXMax(oom, rm)
+                + ", yMin=" + getYMin(oom, rm) + ", yMax=" + getYMax(oom, rm)
+                + ", zMin=" + getZMin(oom, rm) + ", zMax=" + getZMax(oom, rm) + ")";
     }
 
     public void translate(V3D_Vector v, int oom, RoundingMode rm) {
@@ -544,88 +541,123 @@ public class V3D_Envelope implements Serializable {
 //        pts[7] = new V3D_Point(rtf);
         V3D_Point c = getCentroid(oom, rm);
         
-        V3D_Vector v2 = new V3D_Vector(pt, c, oom, rm).getCrossProduct(v, oom, rm);
+        V3D_Vector cv = new V3D_Vector(pt, c, oom, rm);
+        
+        V3D_Vector v2 = cv.getCrossProduct(v, oom, rm);
         
         /**
          * Calculate the plane touching this in the direction given from pt to
          * c.
          */
-        // Find a closest point.
+        // Find a closest point (there could be up to 4).
         Math_BigRational d2min = pt.getDistanceSquared(pts[0], oom, rm);
         V3D_Point closest = pts[0];
         for (int i = 1; i < pts.length; i++) {
             Math_BigRational d2 = pt.getDistanceSquared(pts[i], oom, rm);
             if (d2.compareTo(d2min) == -1) {
+                d2min = d2;
                 closest = pts[i];
             }
         }
-        V3D_Plane pl0 = new V3D_Plane(closest, new V3D_Vector(pt, c, oom, rm));
+        // Use the closest point to define the plane of the screen.
+        V3D_Plane pl0 = new V3D_Plane(closest, cv);
+        // Intersect the rays from pts to each point with the screen.
+        V3D_Point[] ipts = new V3D_Point[pts.length];
[…]
