### 2022-10-23 3fd273c
Simplification and javadoc for release.
