+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java
@@ -18,7 +18,6 @@ package uk.ac.leeds.ccg.v3d.geometry;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Iterator;
 import java.util.List;
 import uk.ac.leeds.ccg.math.arithmetic.Math_BigDecimal;
@@ -64,16 +63,30 @@ import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
  * @author Andy Turner
  * @version 1.0
  */
-public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face,
-        V3D_Intersection {
+public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
 
     private static final long serialVersionUID = 1L;
 
+    /**
+     * The plane of the triangle
+     */
     public V3D_Plane pl;
 
+    /**
+     * Defines one of the corners of the triangle.
+     */
     public V3D_Vector p;
+
+    /**
+     * Defines one of the corners of the triangle.
+     */
     public V3D_Vector q;
+
+    /**
+     * Defines one of the corners of the triangle.
+     */
     public V3D_Vector r;
+    
 //    /**
 //     * For storing the line segment from {@link #pl} to {@link #q}.
 //     */
@@ -130,6 +143,8 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face,
      * @param p What {@link #pl} is set to.
      * @param q What {@link #q} is set to.
      * @param r What {@link #r} is set to.
+     * @param oom The Order of Magnitude for the precision.
+     * @param rm The RoundingMode if rounding is needed.
      */
     public V3D_Triangle(V3D_Environment e, V3D_Vector p, V3D_Vector q,
             V3D_Vector r, int oom, RoundingMode rm) {
@@ -144,6 +159,8 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face,
      * @param p What {@link #pl} is set to.
      * @param q What {@link #q} is set to.
      * @param r What {@link #r} is set to.
+     * @param oom The Order of Magnitude for the precision.
+     * @param rm The RoundingMode if rounding is needed.
      */
     public V3D_Triangle(V3D_Environment e, V3D_Vector offset, V3D_Vector p,
             V3D_Vector q, V3D_Vector r, int oom, RoundingMode rm) {
@@ -161,6 +178,8 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face,
      * triangle.
      * @param r Defines the other point relative to l.offset that defines the
      * triangle.
+     * @param oom The Order of Magnitude for the precision.
+     * @param rm The RoundingMode if rounding is needed.
      */
     public V3D_Triangle(V3D_LineSegment l, V3D_Vector r, int oom, RoundingMode rm) {
         this(l.e, l.offset, l.l.p, l.q, r, oom, rm);
@@ -173,38 +192,21 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face,
      * triangle.
      * @param r Defines the other point relative to l.offset that defines the
      * triangle.
+     * @param oom The Order of Magnitude for the precision.
+     * @param rm The RoundingMode if rounding is needed.
      */
     public V3D_Triangle(V3D_LineSegment l, V3D_Point r, int oom, RoundingMode rm) {
         this(l.e, l.offset, l.l.p, l.q, r.getVector(oom, rm).subtract(l.offset, oom, rm), oom, rm);
-//, oom, rm);
-//        super(l.e);
-//        this.pl = new V3D_Plane(l.getP(), l.getQ(oom, rm), r, oom, rm);
     }
-
-//    /**
[…]
