+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_EnvelopeTest.java
@@ -796,132 +796,132 @@ public class V3D_EnvelopeTest extends V3D_Test {
         pt = pP0P0N2;
         // Test 1
         v = V3D_Vector.I;
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pN1N1N1, pN1P1N1, pP1P1N1, pP1N1N1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test 2
         v = V3D_Vector.I.reverse();
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pP1P1N1, pP1N1N1, pN1N1N1, pN1P1N1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test 3
         v = V3D_Vector.J;
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pP1N1N1, pN1N1N1, pN1P1N1, pP1P1N1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test 4
         v = V3D_Vector.J.reverse();
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pN1P1N1, pP1P1N1, pP1N1N1, pN1N1N1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test left face square on.
         pt = pN2P0P0;
         // Test 5
         v = V3D_Vector.K;
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pN1P1N1, pN1N1N1, pN1N1P1, pN1P1P1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test 6
         v = V3D_Vector.K.reverse();
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pN1N1P1, pN1P1P1, pN1P1N1, pN1N1N1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test 7
         v = V3D_Vector.J;
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pN1N1N1, pN1N1P1, pN1P1P1, pN1P1N1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test 8
         v = V3D_Vector.J.reverse();
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pN1P1N1, pN1N1N1, pN1N1P1, pN1P1P1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test bottom face square on.
         pt = pP0N2P0;
         // Test 9
         v = V3D_Vector.I;
-        result = instance.getViewport(oom, rm, pt, v);
+        result = instance.getViewport(pt, v, oom, rm);
         expResult = new V3D_Rectangle(pN1N1N1, pN1N1P1, pP1N1P1, pP1N1N1, oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
         // Test 10
[…]
