+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Envelope.java
@@ -89,8 +89,16 @@ public class V3D_Envelope implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
+    /**
+     * A copy of the V3D_Environment.
+     */
+    public final V3D_Environment e;
+    
+    /**
+     * For storing the offset of this.
+     */
     private V3D_Vector offset;
-    private final V3D_Environment e;
+    
     /**
      * The minimum x-coordinate.
      */
@@ -556,17 +564,17 @@ public class V3D_Envelope implements Serializable {
      * that face. In some cases the viewport may intersect (touch) along an edge
      * between two faces. In all other cases, the viewport will intersect
      * (touch) just one corner of this. The right edge of the viewport is in the
-     * direction given by the vector v from te point pt.
+     * direction given by the vector v from the point pt.
      *
-     * @param oom The Order of Magnitude for the precision.
-     * @param rm The RoundingMode used in the calculation.
      * @param pt The point from which observation of this is occurring.
      * @param v The vector pointing to the right of the viewport.
+     * @param oom The Order of Magnitude for the precision.
+     * @param rm The RoundingMode used in the calculation.
      * @return A viewport - a rectangle between pt and this such that all of
      * this is contained in the planes from the point through the viewport.
      */
-    public V3D_Rectangle getViewport(int oom, RoundingMode rm, V3D_Point pt,
-            V3D_Vector v) {
+    public V3D_Rectangle getViewport(V3D_Point pt, V3D_Vector v, int oom, 
+            RoundingMode rm) {
         V3D_Rectangle r;
         V3D_Point[] pts = getPoints(oom, rm);
 //        pts[0] = new V3D_Point(lba);
@@ -578,11 +586,8 @@ public class V3D_Envelope implements Serializable {
 //        pts[6] = new V3D_Point(rta);
 //        pts[7] = new V3D_Point(rtf);
         V3D_Point c = getCentroid(oom, rm);
-
         V3D_Vector cv = new V3D_Vector(pt, c, oom, rm);
-
         V3D_Vector v2 = cv.getCrossProduct(v, oom, rm);
-
         /**
          * Calculate the plane touching this in the direction given from pt to
          * c.
@@ -607,7 +612,129 @@ public class V3D_Envelope implements Serializable {
             ipts[i] = (V3D_Point) ray.getIntersection(pl0, oom, rm);
         }
         // Figure out the extremes in relation to v and v2
+        // Find top, bottom, left and right planes
+        V3D_Plane vpl = new V3D_Plane(c, v);
+        V3D_Plane v2pl = new V3D_Plane(c, v2);
+        // Find top and bottom
+        V3D_Point ap = new V3D_Point(c);
+        ap.translate(v2, oom, rm);
+        Math_BigRational aa = Math_BigRational.ZERO;
+        V3D_Plane tpl = null;
+        //V3D_Point pb = new V3D_Point(v2pl.getP());
+        //pb.translate(v2.reverse(), oom, rm);
+        Math_BigRational ba = Math_BigRational.ZERO;
+        V3D_Plane bpl = null;
+        for (var x : ipts) {
+            V3D_Point pp = vpl.getPointOfProjectedIntersection(x, oom, rm);
+            //Math_BigRational a = v2.getAngle(new V3D_Vector(pt, pp, oom, rm), oom, rm).abs();
+            Math_BigRational a = cv.getAngle(new V3D_Vector(pt, pp, oom, rm), oom, rm).abs();
+            if (v2pl.isOnSameSide(ap, x, oom, rm)) {
[…]

[…]
