### 2022-10-27 #4283 @arfon
@agdturner – I'm going to mark this as paused for now and let's revisit in a couple of months?

> Since we started the review, I have stepped back as an editor for JOSS. So whatever we decide, as this is not going to be finish soon, it would be better to assign a new editor for the paper.

Yes, of course. No problem at all. I can take this one on from here. Thank you @VivianePons!
