### 2022-10-27 #4283 @agdturner
>

Hi @arfon
Thanks for taking over the editing role for this. Let me see what I can get around to in the next couple of months. I have improved my 3D rendering code and am working on some science use case scenarios.
