### 2022-10-28 e2557b6
Simplified point.
