+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_PointTest.java
@@ -227,75 +227,6 @@ public class V3D_PointTest extends V3D_Test {
     }
 
     /**
-     * Test of getDistance method, of class V3D_Point.
-     */
-    @Test
-    public void testGetDistance_V3D_Line_int() {
-        System.out.println("getDistance");
-        int oom = -3;
-        RoundingMode rm = RoundingMode.HALF_UP;
-        V3D_Line l = new V3D_Line(pP0P0P0, pP0P0P1, oom, rm);
-        V3D_Point instance = pP0P1P0;
-        BigDecimal expResult = P1.toBigDecimal(oom, rm);
-        BigDecimal result = instance.getDistance(l, oom, rm);
-        assertEquals(expResult, result);
-        // Test 2
-        l = new V3D_Line(pP0P0P0, pP0P0P1, oom, rm);
-        instance = new V3D_Point(e, P3, P4, P0);
-        expResult = P5.toBigDecimal(oom, rm);
-        result = instance.getDistance(l, oom, rm);
-        assertEquals(expResult, result);
-        // Test 3
-        l = new V3D_Line(pP0P0P1, pP0P0P0, oom, rm);
-        instance = new V3D_Point(e, P3, P4, P0);
-        expResult = P5.toBigDecimal(oom, rm);
-        result = instance.getDistance(l, oom, rm);
-        assertEquals(expResult, result);
-        // Test 4
-        l = new V3D_Line(pP0P0P0, pP0P0P1, oom, rm);
-        instance = new V3D_Point(e, P4, P3, P0);
-        expResult = P5.toBigDecimal(oom, rm);
-        result = instance.getDistance(l, oom, rm);
-        assertEquals(expResult, result);
-        // Test 3
-        l = new V3D_Line(pP0P0P0, pP0P0P1, oom, rm);
-        instance = new V3D_Point(e, P4, P3, P10);
-        expResult = P5.toBigDecimal(oom, rm);
-        result = instance.getDistance(l, oom, rm);
-        assertEquals(expResult, result);
-    }
-
-    /**
-     * Test of getDistance method, of class V3D_Point.
-     */
-    @Test
-    public void testGetDistance_V3D_LineSegment_int() {
-        System.out.println("getDistance");
-        int oom = -3;
-        RoundingMode rm = RoundingMode.HALF_UP;
-        V3D_LineSegment l = new V3D_LineSegment(pP0P0P0, pP2P0P0, oom, rm);
-        V3D_Point instance = pP1P1P0;
-        BigDecimal expResult = BigDecimal.ONE;
-        BigDecimal result = instance.getDistance(l, oom, rm);
-        assertTrue(expResult.compareTo(result) == 0);
-        // Test 2
-        instance = pN1N1P0;
-        result = instance.getDistance(l, oom, rm);
-        assertTrue(expResult.compareTo(result) == 0);
-        // Test 3
-        instance = pP2P2P0;
-        expResult = BigDecimal.valueOf(2);
-        result = instance.getDistance(l, oom, rm);
-        assertTrue(expResult.compareTo(result) == 0);
-        // Test 4
-        instance = pP2P2P0;
-        l = new V3D_LineSegment(pP0P0P0, pP1P0P0, oom, rm);
-        expResult = new Math_BigRationalSqrt(5, oom, rm).toBigDecimal(oom, rm);
-        result = instance.getDistance(l, oom, rm);
-        assertTrue(expResult.compareTo(result) == 0);
-    }
-
-    /**
      * Test of toString method, of class V3D_Point.
      */
     @Test
