+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Point.java
@@ -67,7 +67,7 @@ public class V3D_Point extends V3D_FiniteGeometry {
     /**
      * The origin of the Euclidean space.
      */
-    public static final V3D_Point ORIGIN = new V3D_Point(new V3D_Environment(), 0, 0, 0);
+    public static final V3D_Point ORIGIN = new V3D_Point(0, 0, 0);
 
     /**
      * The position relative to the {@link #offset}. Taken together with
@@ -81,7 +81,7 @@ public class V3D_Point extends V3D_FiniteGeometry {
      * @param p The point to clone/duplicate.
      */
     public V3D_Point(V3D_Point p) {
-        super(p.e, new V3D_Vector(p.offset));
+        super(new V3D_Vector(p.offset));
         this.rel = new V3D_Vector(p.rel);
     }
 
@@ -89,36 +89,33 @@ public class V3D_Point extends V3D_FiniteGeometry {
      * Create a new instance with {@link #offset} set to
      * {@link V3D_Vector#ZERO}.
      *
-     * @param e What {@link #e} is set to.
      * @param rel Cloned to initialise {@link #rel}.
      */
-    public V3D_Point(V3D_Environment e, V3D_Vector rel) {
-        this(e, V3D_Vector.ZERO, rel);
+    public V3D_Point(V3D_Vector rel) {
+        this(V3D_Vector.ZERO, rel);
     }
 
     /**
      * Create a new instance.
      *
-     * @param e What {@link #e} is set to.
      * @param offset Cloned to initialise {@link #offset}.
      * @param rel Cloned to initialise {@link #rel}.
      */
-    public V3D_Point(V3D_Environment e, V3D_Vector offset, V3D_Vector rel) {
+    public V3D_Point(V3D_Vector offset, V3D_Vector rel) {
 //        super(new V3D_Vector(offset), Math.min(offset.getMagnitude().getOom(),
 //                rel.getMagnitude().getOom()));
         //super(e, new V3D_Vector(offset));
-        super(e, offset);
+        super(offset);
         this.rel = new V3D_Vector(rel);
     }
 
     /**
      * Create a new instance.
      *
-     * @param e What {@link #e} is set to.
      * @param p The point to clone/duplicate.
      */
-    public V3D_Point(V3D_Environment e, V3D_VPoint p) {
-        super(e, new V3D_Vector(p.offset));
+    public V3D_Point(V3D_VPoint p) {
+        super(new V3D_Vector(p.offset));
         this.rel = new V3D_Vector(p.rel);
     }
 
@@ -137,14 +134,13 @@ public class V3D_Point extends V3D_FiniteGeometry {
      * Create a new instance with {@link #offset} set to
      * {@link V3D_Vector#ZERO}.
      *
-     * @param e What {@link #e} is set to.
      * @param x What {@link #rel} x component is set to.
      * @param y What {@link #rel} y component is set to.
      * @param z What {@link #rel} z component is set to.
      */
-    public V3D_Point(V3D_Environment e, Math_BigRational x, Math_BigRational y,
+    public V3D_Point(Math_BigRational x, Math_BigRational y,
             Math_BigRational z) {
-        super(e, V3D_Vector.ZERO);
+        super(V3D_Vector.ZERO);
         this.rel = new V3D_Vector(x, y, z);
     }
 
@@ -152,14 +148,13 @@ public class V3D_Point extends V3D_FiniteGeometry {
      * Create a new instance with {@link #offset} set to
      * {@link V3D_Vector#ZERO}.
      *
-     * @param e What {@link #e} is set to.
      * @param x What {@link #rel} x component is set to.
      * @param y What {@link #rel} y component is set to.
[…]
