+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_TriangleTest.java
@@ -71,7 +71,7 @@ public class V3D_TriangleTest extends V3D_Test {
         pt = pN1N1P0;
         assertFalse(instance.isIntersectedBy(pt, oom, rm));
         // Test 3
-        pt = new V3D_Point(e, Math_BigRational.ONE.divide(2), P0, P0);
+        pt = new V3D_Point(Math_BigRational.ONE.divide(2), P0, P0);
         assertTrue(instance.isIntersectedBy(pt, oom, rm));
     }
 
@@ -84,7 +84,7 @@ public class V3D_TriangleTest extends V3D_Test {
         int oom = -3;
         RoundingMode rm = RoundingMode.HALF_UP;
         V3D_Triangle instance = new V3D_Triangle(pP0P0P0, pP0P1P0, pP1P0P0, oom, rm);
-        V3D_Envelope expResult = new V3D_Envelope(e, oom, rm, pP0P0P0, pP0P1P0, pP1P0P0);
+        V3D_Envelope expResult = new V3D_Envelope(oom, rm, pP0P0P0, pP0P1P0, pP1P0P0);
         V3D_Envelope result = instance.getEnvelope(oom, rm);
         assertTrue(expResult.equals(result, oom, rm));
     }
@@ -129,14 +129,14 @@ public class V3D_TriangleTest extends V3D_Test {
 //        // line
 //        // x = 0, y = 0, z = t
 //        // points (0, 0, 0), (0, 0, 1) 
-//        l = new V3D_Line(new V3D_Point(e, 0, 0, -120),
-//                new V3D_Point(e, 0, 3, 0), oom, rm);
+//        l = new V3D_Line(new V3D_Point(0, 0, -120),
+//                new V3D_Point(0, 3, 0), oom, rm);
 //        // plane
 //        // points (0, 0, -1), (0, 4, 0), (2, 0, 0)
 //        instance = new V3D_Triangle(
-//                new V3D_Point(e, -40, -30, 6),
-//                new V3D_Point(e, 0, 30, 5),
-//                new V3D_Point(e, 40, 30, 5), oom, rm);
+//                new V3D_Point(-40, -30, 6),
+//                new V3D_Point(0, 30, 5),
+//                new V3D_Point(40, 30, 5), oom, rm);
 //        assertTrue(instance.isIntersectedBy(l, oom, rm));
 //    }
 //
@@ -151,11 +151,11 @@ public class V3D_TriangleTest extends V3D_Test {
 //        V3D_LineSegment l;
 //        V3D_Triangle instance;
 //        // Test 1
-//        instance = new V3D_Triangle(e, P0P0P0, P1P0P0, P1P2P0, P2P0P0, oom, rm);
+//        instance = new V3D_Triangle(P0P0P0, P1P0P0, P1P2P0, P2P0P0, oom, rm);
 //        l = new V3D_LineSegment(pP0P0P0, pP1P0P0, oom, rm);
 //        assertTrue(instance.isIntersectedBy(l, oom, rm));
 //        // Test 2
-//        l = new V3D_LineSegment(e, P0P0P0, P0P0P0,
+//        l = new V3D_LineSegment(P0P0P0, P0P0P0,
 //                new V3D_Vector(Math_BigRational.valueOf(0.5d), P0, P0), oom, rm);
 //        assertFalse(instance.isIntersectedBy(l, oom, rm));
 //        // Test 3
@@ -165,7 +165,7 @@ public class V3D_TriangleTest extends V3D_Test {
 //        l = new V3D_LineSegment(pP1P0P1, pP1P0N1, oom, rm);
 //        assertTrue(instance.isIntersectedBy(l, oom, rm));
 //        // Test 5
[…]
