### 2022-10-29 31f9765
Changed rotations to use line.
