+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_VectorTest.java
@@ -885,16 +885,15 @@ public class V3D_VectorTest extends V3D_Test {
     public void testRotate() {
         System.out.println("rotate");
         V3D_Vector axisOfRotation = new V3D_Vector(0, 1, 0);
-        Math_BigDecimal bd = new Math_BigDecimal();
         int oom = -3;
         int oomt = oom - 2;
         RoundingMode rm = RoundingMode.HALF_UP;
-        Math_BigRational Pi = Math_BigRational.valueOf(bd.getPi(oomt, rm));
+        Math_BigRational Pi = Math_BigRational.valueOf(
+                V3D_Environment.bd.getPi(oomt, rm));
         Math_BigRational theta = Pi.divide(2);
-        Math_BigDecimal bD = new Math_BigDecimal();
         V3D_Vector instance = new V3D_Vector(1, 0, 0);
         V3D_Vector expResult = new V3D_Vector(0, 0, -1);
-        V3D_Vector result = instance.rotate(axisOfRotation, theta, bD, oom, rm);
+        V3D_Vector result = instance.rotate(axisOfRotation, theta, oom, rm);
         assertTrue(expResult.equals(result));
         // Test 2
         // From Example 2: https://graphics.stanford.edu/courses/cs348a-17-winter/Papers/quaternion.pdf
@@ -902,7 +901,7 @@ public class V3D_VectorTest extends V3D_Test {
         theta = Pi.multiply(2).divide(3);
         instance = new V3D_Vector(1, 0, 0);
         expResult = new V3D_Vector(0, 1, 0);
-        result = instance.rotate(axisOfRotation, theta, bD, oom, rm);
+        result = instance.rotate(axisOfRotation, theta, oom, rm);
         assertTrue(expResult.equals(result));
 //        // Test 3 Fails :(
 //        // From Example 1: https://www.imsc.res.in/~knr/131129workshop/writeup_knr.pdf
@@ -926,7 +925,7 @@ public class V3D_VectorTest extends V3D_Test {
         theta = Pi.divide(6).negate();
         instance = new V3D_Vector(0.6, 0.8, 0);
         expResult = new V3D_Vector(0.9196, 0.3928, 0);
-        result = instance.rotate(axisOfRotation, theta, bD, oom, rm);
+        result = instance.rotate(axisOfRotation, theta, oom, rm);
         assertTrue(expResult.equals(result));
 //        // Test 5
 //        // From Case 2 https://www.tobynorris.com/work/prog/csharp/quatview/help/case_1.htm
@@ -944,7 +943,7 @@ public class V3D_VectorTest extends V3D_Test {
         theta = Pi.divide(2);
         instance = new V3D_Vector(1, 0, 0);
         expResult = new V3D_Vector(0, 1, 0);
-        result = instance.rotate(axisOfRotation, theta, bD, oom, rm);
+        result = instance.rotate(axisOfRotation, theta, oom, rm);
         assertTrue(expResult.equals(result));
         // Test 7
         oom = -4;
@@ -952,7 +951,7 @@ public class V3D_VectorTest extends V3D_Test {
         theta = Pi;
         instance = new V3D_Vector(1, 0, 0);
         expResult = new V3D_Vector(0, 1, 0);
-        result = instance.rotate(axisOfRotation, theta, bD, oom, rm);
+        result = instance.rotate(axisOfRotation, theta, oom, rm);
         assertTrue(expResult.equals(result));
         // Test 8
         oom = -4;
@@ -960,7 +959,7 @@ public class V3D_VectorTest extends V3D_Test {
         theta = Pi;
         instance = new V3D_Vector(2, 0, 0);
         expResult = new V3D_Vector(0, 2, 0);
[…]
