### 2022-10-31 4c47ccb
Updated intersection methods.
