+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java
@@ -147,12 +147,9 @@ public class V3D_Vector implements Serializable {
      */
     public V3D_Vector(Math_BigRational dx, Math_BigRational dy,
             Math_BigRational dz) {
-        Math_BigRational dx2 = dx.pow(2);
-        Math_BigRational dy2 = dy.pow(2);
-        Math_BigRational dz2 = dz.pow(2);
-        this.dx = new Math_BigRationalSqrt(dx2, dx);
-        this.dy = new Math_BigRationalSqrt(dy2, dy);
-        this.dz = new Math_BigRationalSqrt(dz2, dz);
+        this.dx = new Math_BigRationalSqrt(dx.pow(2), dx);
+        this.dy = new Math_BigRationalSqrt(dy.pow(2), dy);
+        this.dz = new Math_BigRationalSqrt(dz.pow(2), dz);
     }
 
     /**
@@ -174,11 +171,9 @@ public class V3D_Vector implements Serializable {
      */
     public V3D_Vector(Math_BigRationalSqrt dx, Math_BigRational dy,
             Math_BigRational dz) {
-        Math_BigRational dy2 = dy.pow(2);
-        Math_BigRational dz2 = dz.pow(2);
         this.dx = dx;
-        this.dy = new Math_BigRationalSqrt(dy2, dy);
-        this.dz = new Math_BigRationalSqrt(dz2, dz);
+        this.dy = new Math_BigRationalSqrt(dy.pow(2), dy);
+        this.dz = new Math_BigRationalSqrt(dz.pow(2), dz);
     }
 
     /**
@@ -188,11 +183,9 @@ public class V3D_Vector implements Serializable {
      */
     public V3D_Vector(Math_BigRational dx, Math_BigRationalSqrt dy,
             Math_BigRational dz) {
-        Math_BigRational dx2 = dx.pow(2);
-        Math_BigRational dz2 = dz.pow(2);
-        this.dx = new Math_BigRationalSqrt(dx2, dx);
+        this.dx = new Math_BigRationalSqrt(dx.pow(2), dx);
         this.dy = dy;
-        this.dz = new Math_BigRationalSqrt(dz2, dz);
+        this.dz = new Math_BigRationalSqrt(dz.pow(2), dz);
     }
 
     /**
@@ -202,10 +195,8 @@ public class V3D_Vector implements Serializable {
      */
     public V3D_Vector(Math_BigRational dx, Math_BigRational dy,
             Math_BigRationalSqrt dz) {
-        Math_BigRational dx2 = dx.pow(2);
-        Math_BigRational dy2 = dy.pow(2);
-        this.dx = new Math_BigRationalSqrt(dx2, dx);
-        this.dy = new Math_BigRationalSqrt(dy2, dy);
+        this.dx = new Math_BigRationalSqrt(dx.pow(2), dx);
+        this.dy = new Math_BigRationalSqrt(dy.pow(2), dy);
         this.dz = dz;
     }
 
@@ -216,10 +207,9 @@ public class V3D_Vector implements Serializable {
      */
     public V3D_Vector(Math_BigRationalSqrt dx, Math_BigRationalSqrt dy,
             Math_BigRational dz) {
-        Math_BigRational dz2 = dz.pow(2);
         this.dx = dx;
         this.dy = dy;
-        this.dz = new Math_BigRationalSqrt(dz2, dz);
+        this.dz = new Math_BigRationalSqrt(dz.pow(2), dz);
     }
 
     /**
@@ -229,9 +219,8 @@ public class V3D_Vector implements Serializable {
      */
     public V3D_Vector(Math_BigRationalSqrt dx, Math_BigRational dy,
[…]
