### 2022-11-01 2c16724
Cleaned up test class.
