+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_PointTest.java
@@ -57,39 +57,6 @@ public class V3D_PointTest extends V3D_Test {
     }
 
     /**
-     * Test of getDistance method, of class V3D_Point.
-     */
-    @Test
-    public void testGetDistance() {
-        System.out.println("getDistance");
-        int oom = -3;
-        RoundingMode rm = RoundingMode.HALF_UP;
-        V3D_Point instance = pP0P0P0;
-        BigDecimal expResult = BigDecimal.ZERO;
-        BigDecimal result = instance.getDistance(instance, oom, rm);
-        assertEquals(expResult, result);
-        // Test 2
-        instance = new V3D_Point(P3, P4, P0);
-        expResult = P5.toBigDecimal(oom, rm);
-        result = instance.getDistance(pP0P0P0, oom, rm);
-        assertTrue(expResult.compareTo(result) == 0);
-    }
-
-    /**
-     * Test of getEnvelope method, of class V3D_Point.
-     */
-    @Test
-    public void testGetEnvelope() {
-        System.out.println("getEnvelope");
-        int oom = -3;
-        RoundingMode rm = RoundingMode.HALF_UP;
-        V3D_Envelope expResult = new V3D_Envelope(oom, rm, pP0P0P0, pP1P1P1);
-        V3D_Envelope result = pP0P0P0.getEnvelope(oom, rm);
-        result = result.union(pP1P1P1.getEnvelope(oom, rm), oom, rm);
-        assertTrue(expResult.equals(result, oom, rm));
-    }
-    
-    /**
      * Test of equals method, of class V3D_Point.
      */
     @Test
@@ -126,10 +93,23 @@ public class V3D_PointTest extends V3D_Test {
     }
 
     /**
+     * Test of getEnvelope method, of class V3D_Point.
+     */
+    @Test
+    public void testGetEnvelope() {
+        System.out.println("getEnvelope");
+        int oom = -3;
+        RoundingMode rm = RoundingMode.HALF_UP;
+        V3D_Envelope expResult = new V3D_Envelope(oom, rm, pP1P2N2);
+        V3D_Envelope result = pP1P2N2.getEnvelope(oom, rm);
+        assertTrue(expResult.equals(result, oom, rm));
+    }
+
+    /**
      * Test of getDistance method, of class V3D_Point.
      */
     @Test
-    public void testGetDistance_V3D_Point() {
+    public void testGetDistance_int_RoundingMode_V3D_Point() {
         System.out.println("getDistance");
         int oom = -3;
         RoundingMode rm = RoundingMode.HALF_UP;
@@ -191,6 +171,25 @@ public class V3D_PointTest extends V3D_Test {
      * Test of getDistance method, of class V3D_Point.
      */
     @Test
+    public void testGetDistance_int_RoundingMode() {
+        System.out.println("getDistance");
+        int oom = -3;
+        RoundingMode rm = RoundingMode.HALF_UP;
+        V3D_Point instance = pP0P0P0;
+        BigDecimal expResult = BigDecimal.ZERO;
+        BigDecimal result = instance.getDistance(instance, oom, rm);
+        assertEquals(expResult, result);
+        // Test 2
+        instance = new V3D_Point(P3, P4, P0);
+        expResult = P5.toBigDecimal(oom, rm);
+        result = instance.getDistance(pP0P0P0, oom, rm);
+        assertTrue(expResult.compareTo(result) == 0);
+    }    
+    
+    /**
+     * Test of getDistance method, of class V3D_Point.
+     */
+    @Test
     public void testGetDistance_V3D_Point_int() {
         System.out.println("getDistance");
         int oom = -3;
[…]
