### 2022-11-05 f8af20a
Updates for intersections methods.
