+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Envelope.java
@@ -712,30 +712,33 @@ public class V3D_Envelope implements Serializable {
      */
     public V3D_Rectangle getViewport2(V3D_Point pt, V3D_Vector v, int oom,
             RoundingMode rm) {
+        //int oomn2 = oom - 2;
+        int oomn4 = oom - 4;
         V3D_Rectangle r;
         // Get the plane of the viewport.
-        V3D_Point c = getCentroid(oom, rm);
-        Math_BigRational distance = c.getDistance(getPoints(oom, rm)[0], oom, rm);
+        V3D_Point c = getCentroid(oomn4, rm);
+        Math_BigRational distance = c.getDistance(getPoints(oom, rm)[0], oomn4, rm);
         V3D_Point plpt = new V3D_Point(c);
-        V3D_Vector vo = new V3D_Vector(c, pt, oom, rm);
-        plpt.translate(vo.multiply(distance, oom, rm), oom, rm);
+        V3D_Vector vo = new V3D_Vector(c, pt, oomn4, rm).getUnitVector(oomn4, rm);
+        plpt.translate(vo.multiply(distance, oomn4, rm), oomn4, rm);
         V3D_Vector cv = vo.reverse();
         V3D_Plane pl0 = new V3D_Plane(plpt, cv);
         // Figure out the extremes in relation to v (and v2).
-        V3D_Vector v2 = cv.getCrossProduct(v, oom, rm);
+        V3D_Vector v2 = cv.getCrossProduct(v, oomn4, rm);
         // Intersect the rays from pts to each point with the screen.
+        pts = getPoints(oomn4, rm);
         V3D_Point[] ipts = new V3D_Point[pts.length];
         // Get the intersecting points on the screen plane from pt
         for (int i = 0; i < pts.length; i++) {
-            V3D_Ray ray = new V3D_Ray(pt, pts[i], oom, rm);
-            ipts[i] = (V3D_Point) ray.getIntersection(pl0, oom, rm);
+            V3D_Ray ray = new V3D_Ray(pt, pts[i], oomn4, rm);
+            ipts[i] = (V3D_Point) ray.getIntersection(pl0, oomn4, rm);
         }
         // Find top, bottom, left and right planes
         V3D_Plane vpl = new V3D_Plane(c, v);
         V3D_Plane v2pl = new V3D_Plane(c, v2);
         // Find top and bottom
         V3D_Point ap = new V3D_Point(c);
-        ap.translate(v2, oom, rm);
+        ap.translate(v2, oomn4, rm);
         Math_BigRational aa = Math_BigRational.ZERO;
         V3D_Plane tpl = null;
         //V3D_Point pb = new V3D_Point(v2pl.getP());
@@ -743,29 +746,29 @@ public class V3D_Envelope implements Serializable {
         Math_BigRational ba = Math_BigRational.ZERO;
         V3D_Plane bpl = null;
         for (var x : ipts) {
-            V3D_Point pp = vpl.getPointOfProjectedIntersection(x, oom, rm);
+            V3D_Point pp = vpl.getPointOfProjectedIntersection(x, oomn4, rm);
             //Math_BigRational a = v2.getAngle(new V3D_Vector(pt, pp, oom, rm), oom, rm).abs();
-            Math_BigRational a = cv.getAngle(new V3D_Vector(pt, pp, oom, rm), oom, rm).abs();
-            if (v2pl.isOnSameSide(ap, x, oom, rm)) {
+            Math_BigRational a = cv.getAngle(new V3D_Vector(pt, pp, oomn4, rm), oomn4, rm).abs();
+            if (v2pl.isOnSameSide(ap, x, oomn4, rm)) {
[…]
