### 2022-11-05 0dc9d0a
Added ray intersection.
