### 2022-11-06 b429fa6
Added check about if changing offset will have any effect before doing it.
