### 2022-11-06 476c461
Updated ray triangle intersection.
