### 2022-11-06 217d1d7
Re-implemented rectangle so that it no longer extends triangle.
