+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Plane.java
@@ -151,7 +151,7 @@ public class V3D_Plane extends V3D_Geometry {
         }
         return null;
     }
-    
+
     /**
      * The point that defines the plane.
      */
@@ -162,6 +162,11 @@ public class V3D_Plane extends V3D_Geometry {
      * plane.
      */
     public V3D_Vector n;
+
+    /**
+     * For storing the equation of the plane.
+     */
+    protected transient Equation equation;
 
     /**
      * Create a new instance.
@@ -190,8 +195,8 @@ public class V3D_Plane extends V3D_Geometry {
      * Create a new instance.
      *
      * @param l A line segment in the plane.
-     * @param inplane A vector in the plane that is not a scalar multiple of 
-     * the vector of the line of l.
+     * @param inplane A vector in the plane that is not a scalar multiple of the
+     * vector of the line of l.
      * @param oom The Order of Magnitude for the precision.
      * @param rm The RoundingMode for any rounding.
      */
@@ -238,7 +243,7 @@ public class V3D_Plane extends V3D_Geometry {
             RoundingMode rm) {
         this(V3D_Vector.ZERO, p, q, r, oom, rm);
     }
-    
+
     /**
      * Create a new instance.
      *
@@ -269,7 +274,7 @@ public class V3D_Plane extends V3D_Geometry {
      * @param oom The Order of Magnitude for the precision.
      * @param rm The RoundingMode for any rounding.
      */
-    public V3D_Plane(V3D_Vector ptv, V3D_Vector offset, V3D_Vector p, 
+    public V3D_Plane(V3D_Vector ptv, V3D_Vector offset, V3D_Vector p,
             V3D_Vector q, V3D_Vector r, int oom, RoundingMode rm) {
         super(offset);
         V3D_Vector pq = q.subtract(p, oom, rm);
@@ -282,14 +287,14 @@ public class V3D_Plane extends V3D_Geometry {
         }
         this.p = p;
         this.n = pq.getCrossProduct(qr, oom, rm);
-        
+
         V3D_Vector v;
         if (ptv.isZeroVector()) {
             v = p.add(q, oom, rm).add(r, oom, rm).reverse();
         } else {
             v = ptv;
         }
-        
+
         int direction = (n.getDotProduct(v, oom, rm)
                 .divide(n.getDotProduct(n, oom, rm)))
                 .compareTo(Math_BigRational.ZERO);
@@ -298,7 +303,7 @@ public class V3D_Plane extends V3D_Geometry {
         }
         //this.n = qr.getCrossProduct(pq, oom, rm);
     }
-    
+
     /**
      * Create a new instance.
      *
@@ -312,7 +317,7 @@ public class V3D_Plane extends V3D_Geometry {
      * @param oom The Order of Magnitude for the precision.
      * @param rm The RoundingMode for any rounding.
      */
-    public V3D_Plane(V3D_Point pt, V3D_Vector offset, V3D_Vector p, 
+    public V3D_Plane(V3D_Point pt, V3D_Vector offset, V3D_Vector p,
             V3D_Vector q, V3D_Vector r, int oom, RoundingMode rm) {
         this(pt.getVector(oom, rm), offset, p, q, r, oom, rm);
     }
@@ -373,7 +378,7 @@ public class V3D_Plane extends V3D_Geometry {
     public final V3D_Point getP() {
         return new V3D_Point(offset, p);
     }
-    
+
     /**
      * Find a perpendicular vector using: user65203, How to find perpendicular
      * vector to another vector?, URL (version: 2020-10-01):
@@ -407,7 +412,7 @@ public class V3D_Plane extends V3D_Geometry {
 
     /**
      * Get another point on the plane.
-     * 
+     *
      * @param oom The Order of Magnitude for the precision.
[…]

[…]
