### 2022-11-06 476db2e
The plane of the triangle is now variable and can be readily recalculated to a higher precision if required.
