+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java
@@ -64,7 +64,7 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
     /**
      * The plane of the triangle
      */
-    public V3D_Plane pl;
+    protected V3D_Plane pl;
 
     /**
      * Defines one of the corners of the triangle.
@@ -221,7 +221,6 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
     public V3D_Triangle(V3D_Vector offset, V3D_Vector p,
             V3D_Vector q, V3D_Vector r, int oom, RoundingMode rm) {
         super(offset);
-        this.pl = new V3D_Plane(V3D_Vector.ZERO, p, q, r, oom, rm);
         this.oom = oom;
         this.rm = rm;
         this.p = p;
@@ -271,7 +270,6 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
         this.p = p.rel;
         this.q = q.getVector(oom, rm).subtract(p.offset, oom, rm);
         this.r = r.getVector(oom, rm).subtract(p.offset, oom, rm);
-        this.pl = new V3D_Plane(p, p.offset, p.rel, this.q, this.r, oom, rm);
         this.oom = oom;
         this.rm = rm;
     }
@@ -286,14 +284,51 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
      * @param oom The Order of Magnitude for the precision.
      * @param rm The RoundingMode if rounding is needed.
      */
-    public V3D_Triangle(V3D_Point pt, V3D_Point p, V3D_Point q, V3D_Point r, int oom, RoundingMode rm) {
+    public V3D_Triangle(V3D_Point pt, V3D_Point p, V3D_Point q, V3D_Point r, 
+            int oom, RoundingMode rm) {
         super(p.offset);
         this.p = p.rel;
         this.q = q.getVector(oom, rm).subtract(p.offset, oom, rm);
         this.r = r.getVector(oom, rm).subtract(p.offset, oom, rm);
-        this.pl = new V3D_Plane(pt, p.offset, p.rel, this.q, this.r, oom, rm);
         this.oom = oom;
         this.rm = rm;
+    }
+    
+    /**
+     * @param oom The Order of Magnitude for the precision.
+     * @param rm The RoundingMode if rounding is needed.
+     * @return {@link pl} accurate to at least the oom precision using 
+     * RoundingMode rm.
+     */
+    public final V3D_Plane getPl(int oom, RoundingMode rm) {
+        if (pl == null) {
+            pl = new V3D_Plane(offset, p, q, r, oom, rm);
+        } else if (this.oom < oom) {
+            return pl;
+        } else if (this.oom == oom && this.rm.equals(rm)) {
+            return pl;
+        }
+        pl = new V3D_Plane(offset, p, q, r, oom, rm);
+        return pl;
+    }
+
+    /**
+     * @param pt The normal will point to this side of the plane. 
+     * @param oom The Order of Magnitude for the precision.
+     * @param rm The RoundingMode if rounding is needed.
+     * @return {@link pl} accurate to at least the oom precision using 
+     * RoundingMode rm.
+     */
+    public final V3D_Plane getPl(V3D_Point pt, int oom, RoundingMode rm) {
+        if (pl == null) {
+            pl = new V3D_Plane(pt, offset, p, q, r, oom, rm);
+        } else if (this.oom < oom) {
+            return pl;
+        } else if (this.oom == oom && this.rm.equals(rm)) {
+            return pl;
+        }
+        pl = new V3D_Plane(pt, offset, p, q, r, oom, rm);
+        return pl;
     }
 
     /**
[…]
