### 2022-11-11 84cdd16
Simplified light.
