+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Tetrahedron.java
@@ -29,32 +29,32 @@ import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
  * these are stored independently in that the order of the points and the
  * directions of the vectors might be opposite. The points are shared between
  * all the triangles. {@code
- * pl *- - - - - - - - - - - + - - - - - - - - - - -* q
- *     \ ~                                         /|
- *      \     ~                                   / |
- *       \         ~                             /  |
- *        \             ~                       /   |
- *         \                 ~                 /    |
- *          \                      ~          /     |
- *           \                          ~    /      |
- *            \                             / ~     |
- *             \                           /       ~s
- *              \                         /        :
- *               \                       /
- *                \                     /      :
- *                 \                   /
- *                  \                 /     :
- *                   \               /
- *                    \             /    :
- *                     \           /
- *                      \         /   :
- *                       \       /
- *                        \     /  :
- *                         \   /
- *                          \ /:
- *                           *
- *                           r
- * }
+ pl *- - - - - - - - - - - + - - - - - - - - - - -* qv
+     \ ~                                         /|
+      \     ~                                   / |
+       \         ~                             /  |
+        \             ~                       /   |
+         \                 ~                 /    |
+          \                      ~          /     |
+           \                          ~    /      |
+            \                             / ~     |
+             \                           /       ~s
+              \                         /        :
+               \                       /
+                \                     /      :
+                 \                   /
+                  \                 /     :
+                   \               /
+                    \             /    :
+                     \           /
+                      \         /   :
+                       \       /
+                        \     /  :
+                         \   /
+                          \ /:
+                           *
+                           r
+ }
  *
  * @author Andy Turner
  * @version 1.0
@@ -104,7 +104,7 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
     public V3D_Triangle psq;
 
     /**
-     * Create a new instance. {@code pl}, {@code q}, {@code r} and {@code s}
+     * Create a new instance. {@code pl}, {@code qv}, {@code r} and {@code s}
      * must all be different, not the zero vector and collectively they must be
      * three dimensional. This is generally the fastest way to construct a
      * tetrahedron.
@@ -125,7 +125,7 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
     }
 
     /**
-     * Create a new instance. {@code pl}, {@code q}, {@code r} and {@code s}
+     * Create a new instance. {@code pl}, {@code qv}, {@code r} and {@code s}
      * must all be different and not coplanar. No test is done to check these
      * things.
      *
@@ -344,14 +344,14 @@ public class V3D_Tetrahedron extends V3D_FiniteGeometry implements V3D_Volume {
      */
     @Override
     public Math_BigRational getVolume(int oom, RoundingMode rm) {
-        V3D_Triangle tpqr = getPqr(oom, rm);
-        V3D_Point ts = getS();
         int oomn6 = oom - 6;
+        V3D_Triangle tpqr = getPqr(oomn6, rm);
+        V3D_Point ts = getS();
[…]
