+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Point.java
@@ -15,7 +15,6 @@
  */
 package uk.ac.leeds.ccg.v3d.geometry;
 
-import uk.ac.leeds.ccg.v3d.geometry.light.V3D_VPoint;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.ArrayList;
@@ -24,6 +23,7 @@ import java.util.List;
 import uk.ac.leeds.ccg.math.number.Math_BigRational;
 import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
 import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
+import uk.ac.leeds.ccg.v3d.geometry.light.V3D_V;
 
 /**
  * A point is defined by two vectors: {@link #offset} and {@link #rel}. Adding 
@@ -31,31 +31,31 @@ import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
  * {@link #equals(uk.ac.leeds.ccg.v3d.geometry.V3D_Point, int, java.math.RoundingMode)}
  * if they have the same position. The "*" denotes a point in 3D in the following
  * depiction: {@code
- *
- *                          y           -
- *                          +          /                * p=<x0,y0,z0>
- *                          |         /                 |  =offset+rel
- *                          |        /                  |  =<x1+x2,y1+y2,z1+z2>
- *                          |    z0-/-------------------|
- *               r          |      /                   /
- *       rel=<x2,y2,z2>     |     /                   /
- *                          |    /                   /              
- *                          |   /                   /      offset=<x1,y1,z1>
- *                       y0-|  /                   /                o
- *                          | /                   /
- *                          |/                   /
- *  - ----------------------|-------------------/---- + x
- *                         /|                  x0
- *                        / |
- *                       /  |
- *                      /   |
- *                     /    |
- *                    /     |
- *                   /      |
- *                  /       |
- *                 +        |
- *                z         -
- * }
+
+                          y           -
+                          +          /                * pv=<x0,y0,z0>
+                          |         /                 |  =offset+v
+                          |        /                  |  =<x1+x2,y1+y2,z1+z2>
+                          |    z0-/-------------------|
+               r          |      /                   /
+       v=<x2,y2,z2>     |     /                   /
+                          |    /                   /              
+                          |   /                   /      offset=<x1,y1,z1>
+                       y0-|  /                   /                o
+                          | /                   /
+                          |/                   /
+  - ----------------------|-------------------/---- + x
+                         /|                  x0
+                        / |
+                       /  |
+                      /   |
+                     /    |
+                    /     |
+                   /      |
+                  /       |
+                 +        |
+                z         -
+ }
  *
  * @author Andy Turner
  * @version 1.0
@@ -76,7 +76,7 @@ public class V3D_Point extends V3D_FiniteGeometry {
     public V3D_Vector rel;
 
     /**
-     * Create a new instance which is completely independent of {@code p}.
+     * Create a new instance which is completely independent of {@code pv}.
      *
      * @param p The point to clone/duplicate.
      */
@@ -103,7 +103,7 @@ public class V3D_Point extends V3D_FiniteGeometry {
      */
     public V3D_Point(V3D_Vector offset, V3D_Vector rel) {
[…]

[…]
