### 2022-11-11 0a06259
Rename of Zero vector.
