+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_VectorTest.java
@@ -879,7 +879,7 @@ public class V3D_VectorTest extends V3D_Test {
         V3D_Vector instance = new V3D_Vector(1, 0, 0);
         V3D_Vector expResult = new V3D_Vector(0, 0, -1);
         V3D_Vector result = instance.rotate(axisOfRotation, theta, oom, rm);
-        assertTrue(expResult.equals(result));
+        assertTrue(expResult.equals(result, oom, rm));
         // Test 2
         // From Example 2: https://graphics.stanford.edu/courses/cs348a-17-winter/Papers/quaternion.pdf
         axisOfRotation = new V3D_Vector(1, 1, 1).getUnitVector(oomt, rm);
@@ -887,22 +887,22 @@ public class V3D_VectorTest extends V3D_Test {
         instance = new V3D_Vector(1, 0, 0);
         expResult = new V3D_Vector(0, 1, 0);
         result = instance.rotate(axisOfRotation, theta, oom, rm);
-        assertTrue(expResult.equals(result));
+        assertTrue(expResult.equals(result, oom, rm));
 //        // Test 3 Fails :(
 //        // From Example 1: https://www.imsc.res.in/~knr/131129workshop/writeup_knr.pdf
 //        axisOfRotation = new V3D_Vector(0, 1, 0);
 //        theta = Pi.divide(3);
 //        instance = new V3D_Vector(1, -1, 2);
-//        Math_BigRational root3 = new Math_BigRationalSqrt(3, oomt).getSqrt(oomt);
+//        Math_BigRational sqrt3 = new Math_BigRationalSqrt(3, oom, rm).getSqrt(oom, rm);
 //        Math_BigRational x = (Math_BigRational.TEN.add(
-//                Math_BigRational.valueOf(4).multiply(root3))).divide(8);
+//                Math_BigRational.valueOf(4).multiply(sqrt3))).divide(8);
 //        Math_BigRational y = (Math_BigRational.ONE.add(
-//                Math_BigRational.valueOf(2).multiply(root3))).divide(8);
+//                Math_BigRational.valueOf(2).multiply(sqrt3))).divide(8);
 //        Math_BigRational z = (Math_BigRational.valueOf(14).subtract(
-//                Math_BigRational.valueOf(3).multiply(root3))).divide(8);        
+//                Math_BigRational.valueOf(3).multiply(sqrt3))).divide(8);        
 //        expResult = new V3D_Vector(x, y, z);
-//        result = instance.rotate(axisOfRotation, theta, bI, oom);
-//        assertTrue(expResult.equals(result));
+//        result = instance.rotate(axisOfRotation, theta, oom, rm);
+//        assertTrue(expResult.equals(result, oom, rm));
         // Test 4
         // From Case 1 https://www.tobynorris.com/work/prog/csharp/quatview/help/case_1.htm
         oom = -4;
@@ -911,7 +911,7 @@ public class V3D_VectorTest extends V3D_Test {
         instance = new V3D_Vector(0.6, 0.8, 0);
         expResult = new V3D_Vector(0.9196, 0.3928, 0);
         result = instance.rotate(axisOfRotation, theta, oom, rm);
-        assertTrue(expResult.equals(result));
+        assertTrue(expResult.equals(result, oom, rm));
 //        // Test 5
 //        // From Case 2 https://www.tobynorris.com/work/prog/csharp/quatview/help/case_1.htm
 //        oom = -4;
@@ -920,8 +920,8 @@ public class V3D_VectorTest extends V3D_Test {
 //        theta = Math_BigRational.valueOf("53.13").divide(Math_BigRational.valueOf(180)).multiply(Pi).negate();
[…]
