+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Vector.java
@@ -637,9 +637,9 @@ public class V3D_Vector implements Serializable {
 //                                return true;
 //                            } else {
 //                                Math_BigRational sy = getDY(oom, rm)
-//                                        .divide(v.getDY(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDY(oom, rm));
 //                                Math_BigRational sz = getDZ(oom, rm)
-//                                        .divide(v.getDZ(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDZ(oom, rm));
 //                                return sy.compareTo(sz) == 0;
 //                            }
 //                        }
@@ -659,9 +659,9 @@ public class V3D_Vector implements Serializable {
 //                                return true;
 //                            } else {
 //                                Math_BigRational sy = getDY(oom, rm)
-//                                        .divide(v.getDY(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDY(oom, rm));
 //                                Math_BigRational sz = getDZ(oom, rm)
-//                                        .divide(v.getDZ(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDZ(oom, rm));
 //                                return sy.compareTo(sz) == 0;
 //                            }
 //                        }
@@ -681,9 +681,9 @@ public class V3D_Vector implements Serializable {
 //                                return false;
 //                            } else {
 //                                Math_BigRational sx = getDX(oom, rm)
-//                                        .divide(v.getDX(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDX(oom, rm));
 //                                Math_BigRational sz = getDZ(oom, rm)
-//                                        .divide(v.getDZ(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDZ(oom, rm));
 //                                return sx.compareTo(sz) == 0;
 //                            }
 //                        }
@@ -697,9 +697,9 @@ public class V3D_Vector implements Serializable {
 //                        if (dz.isZero()) {
 //                            if (v.dz.isZero()) {
 //                                Math_BigRational sx = getDX(oom, rm)
-//                                        .divide(v.getDX(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDX(oom, rm));
 //                                Math_BigRational sy = getDY(oom, rm)
-//                                        .divide(v.getDY(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDY(oom, rm));
 //                                return sx.compareTo(sy) == 0;
 //                            } else {
 //                                return false;
@@ -709,12 +709,12 @@ public class V3D_Vector implements Serializable {
 //                                return false;
 //                            } else {
 //                                Math_BigRational sx = getDX(oom, rm)
-//                                        .divide(v.getDX(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDX(oom, rm));
 //                                Math_BigRational sy = getDY(oom, rm)
-//                                        .divide(v.getDY(oom, rm)).round(oom, rm);
+//                                        .divide(v.getDY(oom, rm));
 //                                if (sx.compareTo(sy) == 0) {
 //                                    Math_BigRational sz = getDZ(oom, rm)
-//                                            .divide(v.getDZ(oom, rm)).round(oom, rm);
+//                                            .divide(v.getDZ(oom, rm));
 //                                    return sy.compareTo(sz) == 0;
 //                                } else {
 //                                    return false;
[…]
