+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_VectorDoubleTest.java
@@ -1,0 +1,901 @@
+/*
+ * Copyright 2020 Andy Turner, University of Leeds.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.ac.leeds.ccg.v3d.geometry.d;
+
+import uk.ac.leeds.ccg.v3d.geometry.*;
+import java.math.RoundingMode;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
+
+/**
+ * Test of V3D_VectorDouble class.
+ *
+ * @author Andy Turner
+ * @version 1.0
+ */
+public class V3D_VectorDoubleTest extends V3D_TestDouble {
+
+    public V3D_VectorDoubleTest() {
+    }
+
+    @BeforeAll
+    public static void setUpClass() {
+    }
+
+    @AfterAll
+    public static void tearDownClass() {
+    }
+
+    @BeforeEach
+    public void setUp() {
+    }
+
+    @AfterEach
+    public void tearDown() {
+    }
+
+    /**
+     * Test of getDotProduct method, of class V3D_VectorDouble.
+     */
+    @Test
+    public void testGetDotProduct() {
+        System.out.println("getDotProduct");
+        V3D_VectorDouble v = P0P1P0;
+        V3D_VectorDouble instance = P1P0P0;
+        double expResult = 0d;
+        double result = instance.getDotProduct(v);
+        assertTrue(expResult == result);
+        // Test 2
+        v = P0P1P0;
+        instance = P0P0N1;
+        result = instance.getDotProduct(v);
+        assertTrue(expResult == result);
+        // Test 3
+        v = P1P1P1;
+        instance = N1N1N1;
+        expResult = -3d;
+        result = instance.getDotProduct(v);
+        assertTrue(expResult == result);
+    }
+
+    /**
+     * Test of isOrthogonal method, of class V3D_VectorDouble.
+     */
+    @Test
+    public void testIsOrthogonal() {
+        System.out.println("isOrthogonal");
+        V3D_VectorDouble v;
+        V3D_VectorDouble instance;
+        // Test 1
+        v = P1P0P0;
+        instance = P0P1P0;
+        assertTrue(instance.isOrthogonal(v));
+        instance = P0P0P1;
+        assertTrue(instance.isOrthogonal(v));
+        instance = P0N1P0;
+        assertTrue(instance.isOrthogonal(v));
+        instance = N1P0P0;
+        assertFalse(instance.isOrthogonal(v));
+        instance = P1P1P0;
+        assertFalse(instance.isOrthogonal(v));
+        instance = P1N1P0;
+        assertFalse(instance.isOrthogonal(v));
+        instance = N1N1P0;
+        assertFalse(instance.isOrthogonal(v));
+        instance = N1P1P0;
+        assertFalse(instance.isOrthogonal(v));
+        instance = P0P1N1;
+        assertTrue(instance.isOrthogonal(v));
[…]
