+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/d/light/V3D_VDouble.java
@@ -1,0 +1,258 @@
+/*
+ * Copyright 2020 Andy Turner, University of Leeds.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.ac.leeds.ccg.v3d.geometry.d.light;
+
+import java.io.Serializable;
+import java.util.Objects;
+import uk.ac.leeds.ccg.v3d.geometry.d.V3D_PointDouble;
+import uk.ac.leeds.ccg.v3d.geometry.d.V3D_VectorDouble;
+
+/**
+ * Intended for use as a vector or a coordinate. The idea is that this is
+ * lightweight despite using heavyweight numbers. It is not as heavyweight as
+ * V3D_Vector which uses even more heavyweight Math_BigRationalSqrt numbers for
+ * coordinates.
+ *
+ * @author Andy Turner
+ * @version 1.0
+ */
+public class V3D_VDouble implements Serializable {
+
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * The x.
+     */
+    public double x;
+
+    /**
+     * The y.
+     */
+    public double y;
+
+    /**
+     * The z.
+     */
+    public double z;
+
+    /**
+     * The origin {@code <0,0,0>}.
+     */
+    public static final V3D_VDouble ZERO = new V3D_VDouble(0, 0, 0);
+
+    /**
+     * @param v Used to construct this.
+     */
+    public V3D_VDouble(V3D_VDouble v) {
+        this.x = v.x;
+        this.y = v.y;
+        this.z = v.z;
+    }
+
+    /**
+     * @param x What {@link #x} is set to.
+     * @param y What {@link #y} is set to.
+     * @param z What {@link #z} is set to.
+     */
+    public V3D_VDouble(double x, double y, double z) {
+        this.x = x;
+        this.y = y;
+        this.z = z;
+    }
+
+    /**
+     * @param p A point to construct from.
+     */
+    public V3D_VDouble(V3D_PointDouble p) {
+        this(p.getX(), p.getY(), p.getZ());
+    }
+
+    /**
+     * @param v A vector to construct from.
+     */
+    public V3D_VDouble(V3D_VectorDouble v) {
+        this(v.dx, v.dy, v.dz);
+    }
+
+    /**
+     * Effectively this becomes a vector of the difference. What needs to be
+     * applied to p to get q.
+     *
+     * @param p The start.
+     * @param q The end.
+     */
+    public V3D_VDouble(V3D_VDouble p, V3D_VDouble q) {
+        this(q.x - p.x, q.y - p.y, q.z - p.z);
+    }
+
+    @Override
+    public String toString() {
+        return toString("");
+    }
+
+    /**
+     * @param pad A padding of spaces.
+     * @return A description of this.
+     */
+    public String toString(String pad) {
+        return this.getClass().getSimpleName() + "\n"
+                + pad + "(\n"
+                + toStringFields(pad + " ") + "\n"
+                + pad + ")";
+    }
+
+    /**
+     * @param pad A padding of spaces.
+     * @return A description of the fields.
+     */
+    protected String toStringFields(String pad) {
+        return pad + "x=" + x + ",\n"
+                + pad + "y=" + y + ",\n"
+                + pad + "z=" + z;
+    }
+
[…]
