+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TetrahedronDouble.java
@@ -1,0 +1,1639 @@
+/*
+ * Copyright 2020 Andy Turner, University of Leeds.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.ac.leeds.ccg.v3d.geometry.d;
+
+import uk.ac.leeds.ccg.math.arithmetic.Math_Double;
+
+/**
+ *
+ * A V3D_TetrahedronDouble comprises 4 corner V3D_PointDouble points that are
+ * not coplanar or collinear or coincident. There are 4 V3D_TriangleDouble faces
+ * defined from the outside with the points arranged clockwise. Each face has a
+ * V3D_Line_SegmentDouble edge equal to that of another face, but these are
+ * stored independently in that the order of the points and the directions of
+ * the vectors might be opposite. The points are shared between all the
+ * triangles. {@code
+ * pl *- - - - - - - - - - - + - - - - - - - - - - -* qv
+ * \ ~                                         /|
+ * \     ~                                   / |
+ * \         ~                             /  |
+ * \             ~                       /   |
+ * \                 ~                 /    |
+ * \                      ~          /     |
+ * \                          ~    /      |
+ * \                             / ~     |
+ * \                           /       ~s
+ * \                         /        :
+ * \                       /
+ * \                     /      :
+ * \                   /
+ * \                 /     :
+ * \               /
+ * \             /    :
+ * \           /
+ * \         /   :
+ * \       /
+ * \     /  :
+ * \   /
+ * \ /:
+ *
+ * r
+ * }
+ *
+ * @author Andy Turner
+ * @version 1.0
+ */
+public class V3D_TetrahedronDouble extends V3D_FiniteGeometryDouble
+        implements V3D_VolumeDouble {
+
+    private static final long serialVersionUID = 1L;
+
+    /**
+     * For defining one of the points that defines the tetrahedron.
+     */
+    public V3D_VectorDouble p;
+
+    /**
+     * For defining one of the points that defines the tetrahedron.
+     */
+    public V3D_VectorDouble q;
+
+    /**
+     * For defining one of the points that defines the tetrahedron.
+     */
+    public V3D_VectorDouble r;
+
+    /**
+     * For defining one of the points that defines the tetrahedron.
+     */
+    public V3D_VectorDouble s;
+
+    /**
+     * For storing the pqr triangle of the tetrahedron.
+     */
+    public V3D_TriangleDouble pqr;
+
+    /**
+     * For storing the qsr triangle of the tetrahedron.
+     */
+    public V3D_TriangleDouble qsr;
+
+    /**
+     * For storing the spr triangle of the tetrahedron.
+     */
+    public V3D_TriangleDouble spr;
+
+    /**
+     * For storing the psq triangle of the tetrahedron.
+     */
+    public V3D_TriangleDouble psq;
+
+    /**
+     * Create a new instance. {@code pl}, {@code qv}, {@code r} and {@code s}
+     * must all be different, not the zero vector and collectively they must be
+     * three dimensional. This is generally the fastest way to construct a
+     * tetrahedron.
+     *
[…]
