+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_PointDoubleTest.java
@@ -1,0 +1,477 @@
+/*
+ * Copyright 2020 Andy Turner, University of Leeds.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.ac.leeds.ccg.v3d.geometry.d;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+//import org.junit.jupiter.api.Disabled;
+
+/**
+ * Test of V3D_PointDouble class.
+ *
+ * @author Andy Turner
+ * @version 1.0
+ */
+public class V3D_PointDoubleTest extends V3D_TestDouble {
+
+    public V3D_PointDoubleTest() {
+    }
+
+    @BeforeAll
+    public static void setUpClass() {
+    }
+
+    @AfterAll
+    public static void tearDownClass() {
+    }
+
+    @BeforeEach
+    public void setUp() {
+    }
+
+    @AfterEach
+    public void tearDown() {
+    }
+
+    /**
+     * Test of equals method, of class V3D_PointDouble.
+     */
+    @Test
+    public void testEquals_V3D_PointDouble() {
+        System.out.println("equals");
+        V3D_PointDouble instance = pP0P0P0;
+        double x = 0d;
+        double y = 0d;
+        double z = 0d;
+        V3D_PointDouble p = new V3D_PointDouble(x, y, z);
+        assertTrue(instance.equals(p));
+        // Test 2
+        x = 1d;
+        y = 10d;
+        z = 0d;
+        instance = new V3D_PointDouble(x, y, z);
+        p = new V3D_PointDouble(x, y, z);
+        assertTrue(instance.equals(p));
+        // Test 3
+        p = new V3D_PointDouble(0d, 10d, 1d);
+        assertFalse(instance.equals(p));
+    }
+
+    /**
+     * Test of getEnvelope method, of class V3D_PointDouble.
+     */
+    @Test
+    public void testGetEnvelope() {
+        System.out.println("getEnvelope");
+        V3D_EnvelopeDouble expResult = new V3D_EnvelopeDouble(pP1P2N2);
+        V3D_EnvelopeDouble result = pP1P2N2.getEnvelope();
+        assertTrue(expResult.equals(result));
+    }
+
+    /**
+     * Test of getDistance method, of class V3D_PointDouble.
+     */
+    @Test
+    public void testGetDistance_V3D_PointDouble() {
+        System.out.println("getDistance");
+        V3D_PointDouble p = V3D_PointDouble.ORIGIN;
+        V3D_PointDouble instance = V3D_PointDouble.ORIGIN;
+        double expResult = 0d;
+        double result = instance.getDistance(p);
+        assertEquals(expResult, result);
+        // Test 2
+        instance = pP1P0P0;
+        expResult = 1d;
+        result = instance.getDistance(p);
+        assertEquals(expResult, result);
+        // Test 3
+        instance = pP1P1P0;
+        expResult = Math.sqrt(2d);
[…]
