+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_EnvelopeTest.java
@@ -28,7 +28,6 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Disabled;
 import uk.ac.leeds.ccg.math.number.Math_BigRational;
 import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
-import uk.ac.leeds.ccg.v3d.V3D_Test;
 import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
 
 /**
