+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TetrahedronDouble.java
@@ -447,7 +447,7 @@ public class V3D_TetrahedronDouble extends V3D_FiniteGeometryDouble
                     if (spri == null) {
                         return V3D_LineSegmentDouble.getGeometry(qsrip, pqrip);
                     } else if (spri instanceof V3D_PointDouble sprip) {
-                        return V3D_TriangleDouble.getGeometry(qsrip, pqrip, sprip);
+                        return V3D_TriangleDouble.getGeometry(qsrip, pqrip, sprip, epsilon);
                     } else {
                         return spri;
                     }
@@ -461,7 +461,7 @@ public class V3D_TetrahedronDouble extends V3D_FiniteGeometryDouble
                     if (spri == null) {
                         return V3D_LineSegmentDouble.getGeometry(psqip, pqrip);
                     } else if (spri instanceof V3D_PointDouble sprip) {
-                        return V3D_TriangleDouble.getGeometry(psqip, pqrip, sprip);
+                        return V3D_TriangleDouble.getGeometry(psqip, pqrip, sprip, epsilon);
                     } else {
                         return spri;
                     }
@@ -504,20 +504,20 @@ public class V3D_TetrahedronDouble extends V3D_FiniteGeometryDouble
             V3D_PointDouble q, V3D_PointDouble r, V3D_PointDouble s,
             double epsilon) {
         if (p.equals(q)) {
-            return V3D_TriangleDouble.getGeometry(p, r, s);
+            return V3D_TriangleDouble.getGeometry(p, r, s, epsilon);
         } else if (p.equals(r)) {
-            return V3D_TriangleDouble.getGeometry(p, q, s);
+            return V3D_TriangleDouble.getGeometry(p, q, s, epsilon);
         } else if (p.equals(s)) {
-            return V3D_TriangleDouble.getGeometry(p, q, r);
+            return V3D_TriangleDouble.getGeometry(p, q, r, epsilon);
         } else if (q.equals(r)) {
-            return V3D_TriangleDouble.getGeometry(p, q, s);
+            return V3D_TriangleDouble.getGeometry(p, q, s, epsilon);
         } else if (q.equals(s)) {
-            return V3D_TriangleDouble.getGeometry(p, q, r);
+            return V3D_TriangleDouble.getGeometry(p, q, r, epsilon);
         } else if (r.equals(s)) {
-            return V3D_TriangleDouble.getGeometry(p, q, r);
+            return V3D_TriangleDouble.getGeometry(p, q, r, epsilon);
         } else {
-            if (V3D_PlaneDouble.isCoplanar(p, q, r, s)) {
-                if (V3D_LineDouble.isCollinear(p, q, r, s)) {
+            if (V3D_PlaneDouble.isCoplanar(epsilon, p, q, r, s)) {
+                if (V3D_LineDouble.isCollinear(epsilon, p, q, r, s)) {
                     //return V3D_Line(pl, qv, r, s);
                     throw new UnsupportedOperationException("Need code to construct a line segment from 4 points!");
                 }
@@ -627,7 +627,7 @@ public class V3D_TetrahedronDouble extends V3D_FiniteGeometryDouble
                                         if (spri == null) {
                                             return new V3D_LineSegmentDouble(lp, qsrip);
                                         } else {
-                                            return V3D_TriangleDouble.getGeometry(lp, qsrip, (V3D_PointDouble) spri);
+                                            return V3D_TriangleDouble.getGeometry(lp, qsrip, (V3D_PointDouble) spri, epsilon);
                                         }
                                     } else {
                                         return qsri;
@@ -642,7 +642,7 @@ public class V3D_TetrahedronDouble extends V3D_FiniteGeometryDouble
                                             return new V3D_LineSegmentDouble(lp, psqip);
                                         } else {
[…]
