+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_RayTest.java
@@ -26,7 +26,6 @@ import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 import uk.ac.leeds.ccg.math.number.Math_BigRational;
 import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
-import uk.ac.leeds.ccg.v3d.V3D_Test;
 import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
 
 /**
