+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TetrahedronDoubleTest.java
@@ -1,0 +1,626 @@
+/*
+ * Copyright 2022 Centre for Computational Geography.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.ac.leeds.ccg.v3d.geometry.d;
+
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import static org.junit.jupiter.api.Assertions.*;
+import uk.ac.leeds.ccg.math.arithmetic.Math_Double;
+
+/**
+ *
+ * @author Andy Turner
+ */
+public class V3D_TetrahedronDoubleTest extends V3D_DoubleTest {
+
+    public V3D_TetrahedronDoubleTest() {
+    }
+
+    @BeforeAll
+    public static void setUpClass() {
+    }
+
+    @AfterAll
+    public static void tearDownClass() {
+    }
+
+    @BeforeEach
+    public void setUp() {
+    }
+
+    @AfterEach
+    public void tearDown() {
+    }
+
+    /**
+     * Test of toString method, of class V3D_TetrahedronDouble.
+     */
+    @Test
+    public void testToString_0args() {
+        System.out.println("toString");
+        double epsilon = 1d / 10000000d;
+        V3D_TetrahedronDouble instance = new V3D_TetrahedronDouble(pP0P0P0, pP1P0P0,
+                pP0P1P0, pP0P1P1);
+        String expResult = "V3D_TetrahedronDouble\n"
+                + "(\n"
+                + " p= V3D_VectorDouble(dx=0.0, dy=0.0, dz=0.0),\n"
+                + " q= V3D_VectorDouble(dx=1.0, dy=0.0, dz=0.0),\n"
+                + " r= V3D_VectorDouble(dx=0.0, dy=1.0, dz=0.0),\n"
+                + " s= V3D_VectorDouble(dx=0.0, dy=1.0, dz=1.0)\n"
+                + ")";
+        String result = instance.toString();
+        //System.out.println(result);
+        assertTrue(expResult.equalsIgnoreCase(result));
+    }
+
+    /**
+     * Test of toString method, of class V3D_TetrahedronDouble.
+     */
+    @Test
+    public void testToString_String() {
+        System.out.println("toString");
+        double epsilon = 1d / 10000000d;
+        String pad = "";
+        V3D_TetrahedronDouble instance = new V3D_TetrahedronDouble(pP0P0P0, pP1P0P0,
+                pP0P1P0, pP0P1P1);
+        String expResult = "V3D_TetrahedronDouble\n"
+                + "(\n"
+                + " offset=V3D_VectorDouble\n"
+                + " (\n"
+                + "  dx=0.0,\n"
+                + "  dy=0.0,\n"
+                + "  dz=0.0\n"
+                + " ) ,\n"
+                + " p=V3D_VectorDouble\n"
+                + " (\n"
[…]
