+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TriangleDouble.java
@@ -350,6 +350,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
      */
     public V3D_PlaneDouble getPQPl() {
         if (pqpl == null) {
+            pq = getPQ();
             pqpl = new V3D_PlaneDouble(pq.getP(), pq.l.v.getCrossProduct(
                 getPl().n));
         }
@@ -363,6 +364,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
      */
     public V3D_PlaneDouble getQRPl() {
         if (qrpl == null) {
+            qr = getQR();
             qrpl = new V3D_PlaneDouble(qr.getP(), qr.l.v.getCrossProduct(
                 getPl().n));
         }
@@ -376,6 +378,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
      */
     public V3D_PlaneDouble getRPPl() {
         if (rppl == null) {
+            rp = getRP();
             rppl = new V3D_PlaneDouble(rp.getP(), rp.l.v.getCrossProduct(
                 getPl().n));
         }
@@ -977,7 +980,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
                                             if (rpplil instanceof V3D_LineSegmentDouble) {
                                                 return rpplil;
                                             } else {
-                                                return V3D_LineSegmentDouble.getGeometry((V3D_PointDouble) rpplil, (V3D_PointDouble) pqplil, lp);
+                                                return V3D_LineSegmentDouble.getGeometry((V3D_PointDouble) rpplil, (V3D_PointDouble) pqplil, lp, epsilon);
                                             }
                                         }
                                     } else {
@@ -985,7 +988,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
                                             return qrplil;
                                         } else {
                                             //if (rpplil == null) {
-                                            return V3D_LineSegmentDouble.getGeometry((V3D_PointDouble) pqplil, (V3D_PointDouble) qrplil, lp);
+                                            return V3D_LineSegmentDouble.getGeometry((V3D_PointDouble) pqplil, (V3D_PointDouble) qrplil, lp, epsilon);
                                             //} else {
                                             //    if (rpplil instanceof V3D_LineSegment) {
                                             //        return rpplil;
@@ -1036,7 +1039,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
                                             if (rpplil instanceof V3D_LineSegmentDouble) {
                                                 return rpplil;
                                             } else {
-                                                return V3D_LineSegmentDouble.getGeometry((V3D_PointDouble) rpplil, (V3D_PointDouble) pqplil, lq);
+                                                return V3D_LineSegmentDouble.getGeometry((V3D_PointDouble) rpplil, (V3D_PointDouble) pqplil, lq, epsilon);
                                             }
                                         }
                                     } else {
@@ -1044,7 +1047,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
                                             return qrplil;
                                         } else {
                                             //if (rpplil == null) {
-                                            return V3D_LineSegmentDouble.getGeometry((V3D_PointDouble) pqplil, (V3D_PointDouble) qrplil, lq);
[…]
