### 2022-11-20 ac03559
Fixed bug in triangle-triangle intersection.
