+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_PlaneDouble.java
@@ -263,14 +263,15 @@ public class V3D_PlaneDouble extends V3D_GeometryDouble {
         }
         this.p = p;
         this.n = pq.getCrossProduct(qr);
-
+        if (n.isZero()) {
+            throw new RuntimeException("Points do not define a plane");
+        }
         V3D_VectorDouble v;
         if (ptv.isZero()) {
             v = p.add(q).add(r).reverse();
         } else {
             v = ptv;
         }
-
         double direction = n.getDotProduct(v) / n.getDotProduct(n);
         if (direction < 0d) {
             n = n.reverse();
@@ -1061,7 +1062,16 @@ public class V3D_PlaneDouble extends V3D_GeometryDouble {
             // There is no intersection.
             return null;
         }
+        
+        try {
         V3D_PointDouble pi = pl.getPointOfProjectedIntersection(getP(), epsilon);
+        } catch (RuntimeException e) {
+            V3D_PointDouble pi = pl.getPointOfProjectedIntersection(getP(), epsilon);
+        }
+        
+        V3D_PointDouble pi = pl.getPointOfProjectedIntersection(getP(), epsilon);
+        
+        
         return new V3D_LineDouble(pi, v);
     }
 
