+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TriangleDoubleTest.java
@@ -22,6 +22,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 import uk.ac.leeds.ccg.math.arithmetic.Math_Double;
+import uk.ac.leeds.ccg.v3d.geometry.V3D_LineSegment;
 
 /**
  * Test of V3D_TriangleDouble class.
@@ -491,23 +492,23 @@ public class V3D_TriangleDoubleTest extends V3D_DoubleTest {
         V3D_PointDouble rtf = new V3D_PointDouble(offset, new V3D_VectorDouble(1 * multiplier, 1 * multiplier, -1 * multiplier));
         V3D_PointDouble rta = new V3D_PointDouble(offset, new V3D_VectorDouble(1 * multiplier, 1 * multiplier, 1 * multiplier));
         // BLUE
-        V3D_TriangleDouble b1 = new V3D_TriangleDouble(centroid, lbf, ltf, rtf);
-        V3D_TriangleDouble b2 = new V3D_TriangleDouble(centroid, lbf, rbf, rtf);
+        V3D_TriangleDouble b1 = new V3D_TriangleDouble(lbf, ltf, rtf);
+        V3D_TriangleDouble b2 = new V3D_TriangleDouble(lbf, rbf, rtf);
         // RED
-        V3D_TriangleDouble r1 = new V3D_TriangleDouble(centroid, lbf, ltf, lta);
-        V3D_TriangleDouble r2 = new V3D_TriangleDouble(centroid, lbf, lba, lta);
+        V3D_TriangleDouble r1 = new V3D_TriangleDouble(lbf, ltf, lta);
+        V3D_TriangleDouble r2 = new V3D_TriangleDouble(lbf, lba, lta);
         // YELLOW
-        V3D_TriangleDouble y1 = new V3D_TriangleDouble(centroid, lba, lta, rta);
-        V3D_TriangleDouble y2 = new V3D_TriangleDouble(centroid, lba, rba, rta);
+        V3D_TriangleDouble y1 = new V3D_TriangleDouble(lba, lta, rta);
+        V3D_TriangleDouble y2 = new V3D_TriangleDouble(lba, rba, rta);
         // GREEN
-        V3D_TriangleDouble g1 = new V3D_TriangleDouble(centroid, rbf, rtf, rta);
-        V3D_TriangleDouble g2 = new V3D_TriangleDouble(centroid, rbf, rta, rba);
+        V3D_TriangleDouble g1 = new V3D_TriangleDouble(rbf, rtf, rta);
+        V3D_TriangleDouble g2 = new V3D_TriangleDouble(rbf, rta, rba);
         // ORANGE
-        V3D_TriangleDouble o1 = new V3D_TriangleDouble(centroid, ltf, lta, rta);
-        V3D_TriangleDouble o2 = new V3D_TriangleDouble(centroid, rtf, ltf, rta);
+        V3D_TriangleDouble o1 = new V3D_TriangleDouble(ltf, lta, rta);
+        V3D_TriangleDouble o2 = new V3D_TriangleDouble(rtf, ltf, rta);
         // PINK
-        V3D_TriangleDouble p1 = new V3D_TriangleDouble(centroid, lbf, rbf, rba);
-        V3D_TriangleDouble p2 = new V3D_TriangleDouble(centroid, lbf, lba, rba);
+        V3D_TriangleDouble p1 = new V3D_TriangleDouble(lbf, rbf, rba);
+        V3D_TriangleDouble p2 = new V3D_TriangleDouble(lbf, lba, rba);
         points[0] = lbf;
         points[1] = lba;
[…]
