### 2022-11-20 b9ee919
Fixed bug in triangle-triangle intersection.
