### 2022-11-20 2a5426c
Fixed bug in triangle-triangle intersection.
