### 2022-11-28 #4283 @agdturner
Hi @arfon,
I have improved the rendering code and applied this to visualise a few other data sets going beyond the Hello World Utah Teapot. I added more images in the [ccg-r3d README](http://gihub.com/agdturner/ccg-r3d).
A colleague supplied me with some data for a mountain region that they are studying. It is just the surface triangles and not really volumes, but I am hopeful this will lead to the development of a science use case scenario.
I've extended the library so that calculations can be done optionally using IEEE double precision floating point coordinates. This should make it straightforward to compare results and provide examples of differences and where using more heavyweight arbitrary precision numbers for coordinates is useful. I think I might be able to find some examples where it is beneficial to have greater precision and range than is offered by double precision.
I've been looking at some open source software with 3D capabilities (Apache Commons Geometry, Blender,  MeshLab, three.js and QGIS), and have been thinking about what sort of comparisons to make with ccg-v3d.
