+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/V3D_LineTest.java
@@ -15,9 +15,8 @@
  */
 package uk.ac.leeds.ccg.v3d.geometry;
 
-import java.math.BigDecimal;
+import ch.obermuhlner.math.big.BigRational;
 import java.math.BigInteger;
-import java.math.MathContext;
 import java.math.RoundingMode;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.AfterAll;
@@ -25,12 +24,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.Disabled;
 import uk.ac.leeds.ccg.math.arithmetic.Math_BigDecimal;
-import uk.ac.leeds.ccg.math.number.Math_BigRational;
 import uk.ac.leeds.ccg.math.matrices.Math_Matrix_BR;
 import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
-import uk.ac.leeds.ccg.v3d.core.V3D_Environment;
 
 /**
  * Test class for V3D_Line.
@@ -151,7 +147,7 @@ public class V3D_LineTest extends V3D_Test {
         pt = new V3D_Point(N0_1E12, N0_1E12, N0_1E12);
         assertTrue(instance.isIntersectedBy(pt, oom, rm));
         // Test 5 works as the rounding puts pt on the line.
-        Math_BigRational a = P0_1E2.add(P1E12);
+        BigRational a = P0_1E2.add(P1E12);
         pt = new V3D_Point(a, a, a);
         assertTrue(instance.isIntersectedBy(pt, oom, rm));
         // Test 6 works as the rounding puts pt on the line.
@@ -250,10 +246,10 @@ public class V3D_LineTest extends V3D_Test {
         instance = new V3D_Line(pP1N1P1, pN1P1N1, oom, rm);
         assertFalse(instance.isParallel(l, oom, rm));
         // Test 13
-        Math_BigRational a = P0_1E12.add(P1E12);
-        Math_BigRational b = N0_1E12.add(N1E12);
-        Math_BigRational a1 = P0_1E12.add(P1E12).add(1);
-        Math_BigRational b1 = N0_1E12.add(N1E12).add(1);
+        BigRational a = P0_1E12.add(P1E12);
+        BigRational b = N0_1E12.add(N1E12);
+        BigRational a1 = P0_1E12.add(P1E12).add(1);
+        BigRational b1 = N0_1E12.add(N1E12).add(1);
         l = new V3D_Line(new V3D_Point(a, a, a), new V3D_Point(b, b, b), oom, rm);
         instance = new V3D_Line(new V3D_Point(a1, a, a), new V3D_Point(b1, b, b), oom, rm);
         assertTrue(instance.isParallel(l, oom, rm)); // Right answer, but for the wrong reason!
@@ -606,13 +602,13 @@ public class V3D_LineTest extends V3D_Test {
         int oom = -1;
         RoundingMode rm = RoundingMode.HALF_UP;
         V3D_Line instance = V3D_Line.X_AXIS;
-        Math_BigRational[][] m = new Math_BigRational[2][3];
-        m[0][0] = Math_BigRational.ZERO;
-        m[0][1] = Math_BigRational.ZERO;
-        m[0][2] = Math_BigRational.ZERO;
-        m[1][0] = Math_BigRational.ONE;
-        m[1][1] = Math_BigRational.ZERO;
-        m[1][2] = Math_BigRational.ZERO;
+        BigRational[][] m = new BigRational[2][3];
[…]
