+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_ConvexHullCoplanar.java
@@ -15,7 +15,7 @@
  */
 package uk.ac.leeds.ccg.v3d.geometry;
 
-import java.math.BigDecimal;
+import ch.obermuhlner.math.big.BigRational;
 import java.math.RoundingMode;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -24,9 +24,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 import java.util.TreeSet;
-import uk.ac.leeds.ccg.math.number.Math_BigRational;
 import uk.ac.leeds.ccg.math.number.Math_BigRationalSqrt;
-//import java.util.ArrayList;
 
 /**
  * A class for representing and using coplanar convex hulls. These are a special
@@ -160,8 +158,8 @@ public class V3D_ConvexHullCoplanar extends V3D_FiniteGeometry
         if (xminIndex == xmaxIndex) {
             V3D_LineSegment yd = new V3D_LineSegment(ymaxp, yminp, oom, rm);
             V3D_LineSegment zd = new V3D_LineSegment(zmaxp, zminp, oom, rm);
-            Math_BigRational ydl2 = yd.getLength2(oom, rm);
-            Math_BigRational zdl2 = zd.getLength2(oom, rm);
+            BigRational ydl2 = yd.getLength2(oom, rm);
+            BigRational zdl2 = zd.getLength2(oom, rm);
             if (ydl2.compareTo(zdl2) == 1) {
                 this.points.add(yminp);
                 this.points.add(ymaxp);
@@ -174,8 +172,8 @@ public class V3D_ConvexHullCoplanar extends V3D_FiniteGeometry
         } else if (yminIndex == ymaxIndex) {
             V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp, oom, rm);
             V3D_LineSegment zd = new V3D_LineSegment(zmaxp, zminp, oom, rm);
-            Math_BigRational xdl2 = xd.getLength2(oom, rm);
-            Math_BigRational zdl2 = zd.getLength2(oom, rm);
+            BigRational xdl2 = xd.getLength2(oom, rm);
+            BigRational zdl2 = zd.getLength2(oom, rm);
             if (xdl2.compareTo(zdl2) == 1) {
                 this.points.add(xminp);
                 this.points.add(xmaxp);
@@ -188,8 +186,8 @@ public class V3D_ConvexHullCoplanar extends V3D_FiniteGeometry
         } else if (zminIndex == zmaxIndex) {
             V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp, oom, rm);
             V3D_LineSegment yd = new V3D_LineSegment(ymaxp, yminp, oom, rm);
-            Math_BigRational xdl2 = xd.getLength2(oom, rm);
-            Math_BigRational ydl2 = yd.getLength2(oom, rm);
+            BigRational xdl2 = xd.getLength2(oom, rm);
+            BigRational ydl2 = yd.getLength2(oom, rm);
             if (xdl2.compareTo(ydl2) == 1) {
                 this.points.add(xminp);
                 this.points.add(xmaxp);
@@ -203,9 +201,9 @@ public class V3D_ConvexHullCoplanar extends V3D_FiniteGeometry
             V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp, oom, rm);
             V3D_LineSegment yd = new V3D_LineSegment(ymaxp, yminp, oom, rm);
             V3D_LineSegment zd = new V3D_LineSegment(zmaxp, zminp, oom, rm);
-            Math_BigRational xdl2 = xd.getLength2(oom, rm);
[…]
