### 2022-12-13 dcc8fdb
Updates for ccg-math 3.0
