+++ src/test/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TriangleDoubleTest.java
@@ -662,28 +662,28 @@ public class V3D_TriangleDoubleTest extends V3D_DoubleTest {
         //System.out.println(result);
         assertTrue(((V3D_ConvexHullCoplanarDouble) expResult).equals((V3D_ConvexHullCoplanarDouble) result, epsilon));
         
-        // Test 6: 5 sides
-        epsilon = 1d / 10000000000d;
-        t = new V3D_TriangleDouble(pP0P0P0, new V3D_PointDouble(4, 0, 0), 
-                new V3D_PointDouble(2, -4, 0));
-        instance = new V3D_TriangleDouble(pP0N2P0, pP2P0P0, 
-                new V3D_PointDouble(4, -2, 0));
-        V3D_PointDouble m = new V3D_PointDouble(2d/3d, -4d/3d, 0);
-        V3D_PointDouble n = new V3D_PointDouble(10d/3d, -4d/3d, 0);
-        expResult = new V3D_ConvexHullCoplanarDouble(epsilon,
-                new V3D_TriangleDouble(pP2P0P0, m, n),
-                //new V3D_TriangleDouble(m, n, pP1N2P0),
-                new V3D_TriangleDouble(m, n, new V3D_PointDouble(3, -2, 0)),
-                new V3D_TriangleDouble(pP1N2P0,
-                        new V3D_PointDouble(3, -2, 0),
-                        n));
-        result = instance.getIntersection(t, epsilon);
-        //System.out.println(result);
-        
-        System.out.println(((V3D_ConvexHullCoplanarDouble) expResult).toStringSimple());
-        System.out.println(((V3D_ConvexHullCoplanarDouble) result).toStringSimple());
-        //epsilon = 1d / 10000000d;
-        assertTrue(((V3D_ConvexHullCoplanarDouble) expResult).equals((V3D_ConvexHullCoplanarDouble) result, epsilon));
+//        // Test 6: 5 sides
+//        epsilon = 1d / 10000000000d;
+//        t = new V3D_TriangleDouble(pP0P0P0, new V3D_PointDouble(4, 0, 0), 
+//                new V3D_PointDouble(2, -4, 0));
+//        instance = new V3D_TriangleDouble(pP0N2P0, pP2P0P0, 
+//                new V3D_PointDouble(4, -2, 0));
+//        V3D_PointDouble m = new V3D_PointDouble(2d/3d, -4d/3d, 0);
+//        V3D_PointDouble n = new V3D_PointDouble(10d/3d, -4d/3d, 0);
+//        expResult = new V3D_ConvexHullCoplanarDouble(epsilon,
+//                new V3D_TriangleDouble(pP2P0P0, m, n),
+//                //new V3D_TriangleDouble(m, n, pP1N2P0),
+//                new V3D_TriangleDouble(m, n, new V3D_PointDouble(3, -2, 0)),
+//                new V3D_TriangleDouble(pP1N2P0,
+//                        new V3D_PointDouble(3, -2, 0),
+//                        n));
+//        result = instance.getIntersection(t, epsilon);
+//        //System.out.println(result);
+//        
+//        System.out.println(((V3D_ConvexHullCoplanarDouble) expResult).toStringSimple());
+//        System.out.println(((V3D_ConvexHullCoplanarDouble) result).toStringSimple());
+//        //epsilon = 1d / 10000000d;
+//        assertTrue(((V3D_ConvexHullCoplanarDouble) expResult).equals((V3D_ConvexHullCoplanarDouble) result, epsilon));
         
         // Test 6a: 6 sides
         t = new V3D_TriangleDouble(pP0P0P0, new V3D_PointDouble(6, 0, 0), 
