### 2022-12-14 e69348f
Updates for ccg-math 3.0
