+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TetrahedronDouble.java
@@ -25,30 +25,30 @@ package uk.ac.leeds.ccg.v3d.geometry.d;
  * the vectors might be opposite. The points are shared between all the
  * triangles. {@code
  * pl *- - - - - - - - - - - + - - - - - - - - - - -* qv
- * \ ~                                         /|
- * \     ~                                   / |
- * \         ~                             /  |
- * \             ~                       /   |
- * \                 ~                 /    |
- * \                      ~          /     |
- * \                          ~    /      |
- * \                             / ~     |
- * \                           /       ~s
- * \                         /        :
- * \                       /
- * \                     /      :
- * \                   /
- * \                 /     :
- * \               /
- * \             /    :
- * \           /
- * \         /   :
- * \       /
- * \     /  :
- * \   /
- * \ /:
- *
- * r
+ *     \ ~                                         /|
+ *      \     ~                                   / |
+ *       \         ~                             /  |
+ *        \             ~                       /   |
+ *         \                 ~                 /    |
+ *          \                      ~          /     |
+ *           \                          ~    /      |
+ *            \                             / ~     |
+ *             \                           /       ~s
+ *              \                         /        :
+ *               \                       /
+ *                \                     /      :
+ *                 \                   /
+ *                  \                 /     :
+ *                   \               /
+ *                    \             /    :
+ *                     \           /
+ *                      \         /   :
+ *                       \       /
+ *                        \     /  :
+ *                         \   /
+ *                          \ /:
+ *                           *
+ *                           r
  * }
  *
  * @author Andy Turner
