+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TriangleDouble.java
@@ -26,30 +26,30 @@ import uk.ac.leeds.ccg.v3d.geometry.d.light.V3D_VTriangleDouble;
  * following depicts a generic triangle {@code
  * pq
  * pv *- - - - - - - - - - - + - - - - - - - - - - -* qv
- * \~                   mpq                   ~/
- * \  ~                 |                 ~  /
- * \    ~              |              ~    /
- * \      ~           |           ~      /
- * \        ~        |        ~        /
- * \          ~     |     ~          /
- * \            ~  |  ~            /
- * -n  -n  -n  -n  -n  c  +n  +n  +n  +n  +n  normal heading out from the page.
- * \          ~  |  ~          /
- * \      ~     |     ~      /
- * \  ~        |        ~  /
- * + mrp      |      mqr +
- * rp  \         |         /  qr
- * \        |        /
- * \       |       /
- * \      |      /
- * \     |     /
- * \    |    /
- * \   |   /
- * \  |  /
- * \ | /
- * \|/
- *
- * r
+ *     \~                   mpq                   ~/
+ *      \  ~                 |                 ~  /
+ *       \    ~              |              ~    /
+ *        \      ~           |           ~      /
+ *         \        ~        |        ~        /
+ *          \          ~     |     ~          /
+ *           \            ~  |  ~            /
+ *       -n  -n  -n  -n  -n  c  +n  +n  +n  +n  +n  normal heading out from the page.
+ *             \          ~  |  ~          /
+ *              \      ~     |     ~      /
+ *               \  ~        |        ~  /
+ *                + mrp      |      mqr +
+ *             rp  \         |         /  qr
+ *                  \        |        /
+ *                   \       |       /
+ *                    \      |      /
+ *                     \     |     /
+ *                      \    |    /
+ *                       \   |   /
+ *                        \  |  /
+ *                         \ | /
+ *                          \|/
+ *                           *
+ *                           r
  * }
  *
  * @author Andy Turner
@@ -68,7 +68,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
      * The tolerance of {@link #pl}.
      */
     protected double plEpsilon;
-    
+
     /**
      * Defines one of the corners of the triangle.
      */
@@ -107,20 +107,32 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
      * plane normal and with a normal orthogonal to the plane normal.
      */
     private V3D_PlaneDouble pqpl;
-    private double pqplEpsilon;
     
+    /**
+     * The epsilon used in generating {@link #pqpl}
+     */
+    private double pqplEpsilon;
+
     /**
      * For storing the plane aligning with {@link #qr} in the direction of the
      * plane normal and with a normal orthogonal to the plane normal.
      */
     private V3D_PlaneDouble qrpl;
+
+    /**
+     * The epsilon used in generating {@link #qrpl}
+     */
     private double qrplEpsilon;
-    
+
     /**
      * For storing the plane aligning with {@link #rp} in the direction of the
      * plane normal and with a normal orthogonal to the plane normal.
      */
     private V3D_PlaneDouble rppl;
+    
+    /**
+     * The epsilon used in generating {@link #rppl}
+     */
     private double rpplEpsilon;
 
 //    /**
@@ -251,7 +263,8 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
 //        this.r = r.getVector().subtract(p.offset);
 //    }
     /**
-     * @return {@link #pl} if {@link pl}this has at least as mto at least the oom precision using
[…]
