+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_ConvexHullCoplanar.java
@@ -163,11 +163,11 @@ public class V3D_ConvexHullCoplanar extends V3D_FiniteGeometry
             if (ydl2.compareTo(zdl2) == 1) {
                 this.points.add(yminp);
                 this.points.add(ymaxp);
-                getConvexHull0(pts, yminp, ymaxp, n, 1, oom, rm);
+                compute(pts, yminp, ymaxp, n, 1, oom, rm);
             } else {
                 this.points.add(zminp);
                 this.points.add(zmaxp);
-                getConvexHull0(pts, zminp, zmaxp, n, 1, oom, rm);
+                compute(pts, zminp, zmaxp, n, 1, oom, rm);
             }
         } else if (yminIndex == ymaxIndex) {
             V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp, oom, rm);
@@ -177,11 +177,11 @@ public class V3D_ConvexHullCoplanar extends V3D_FiniteGeometry
             if (xdl2.compareTo(zdl2) == 1) {
                 this.points.add(xminp);
                 this.points.add(xmaxp);
-                getConvexHull0(pts, xminp, xmaxp, n, 1, oom, rm);
+                compute(pts, xminp, xmaxp, n, 1, oom, rm);
             } else {
                 this.points.add(zminp);
                 this.points.add(zmaxp);
-                getConvexHull0(pts, zminp, zmaxp, n, 1, oom, rm);
+                compute(pts, zminp, zmaxp, n, 1, oom, rm);
             }
         } else if (zminIndex == zmaxIndex) {
             V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp, oom, rm);
@@ -191,11 +191,11 @@ public class V3D_ConvexHullCoplanar extends V3D_FiniteGeometry
             if (xdl2.compareTo(ydl2) == 1) {
                 this.points.add(xminp);
                 this.points.add(xmaxp);
-                getConvexHull0(pts, xminp, xmaxp, n, 1, oom, rm);
+                compute(pts, xminp, xmaxp, n, 1, oom, rm);
             } else {
                 this.points.add(yminp);
                 this.points.add(ymaxp);
-                getConvexHull0(pts, yminp, ymaxp, n, 1, oom, rm);
+                compute(pts, yminp, ymaxp, n, 1, oom, rm);
             }
         } else {
             V3D_LineSegment xd = new V3D_LineSegment(xmaxp, xminp, oom, rm);
@@ -208,21 +208,21 @@ public class V3D_ConvexHullCoplanar extends V3D_FiniteGeometry
                 if (xdl2.compareTo(zdl2) == 1) {
                     this.points.add(xminp);
                     this.points.add(xmaxp);
-                    getConvexHull0(pts, xminp, xmaxp, n, 1, oom, rm);
+                    compute(pts, xminp, xmaxp, n, 1, oom, rm);
                 } else {
                     this.points.add(zminp);
                     this.points.add(zmaxp);
-                    getConvexHull0(pts, zminp, zmaxp, n, 1, oom, rm);
+                    compute(pts, zminp, zmaxp, n, 1, oom, rm);
                 }
             } else {
                 if (ydl2.compareTo(zdl2) == 1) {
                     this.points.add(yminp);
                     this.points.add(ymaxp);
-                    getConvexHull0(pts, yminp, ymaxp, n, 1, oom, rm);
+                    compute(pts, yminp, ymaxp, n, 1, oom, rm);
                 } else {
                     this.points.add(zminp);
                     this.points.add(zmaxp);
-                    getConvexHull0(pts, zminp, zmaxp, n, 1, oom, rm);
+                    compute(pts, zminp, zmaxp, n, 1, oom, rm);
[…]
