+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/V3D_Triangle.java
@@ -27,33 +27,33 @@ import uk.ac.leeds.ccg.v3d.geometry.light.V3D_VTriangle;
  * For representing and processing triangles in 3D. A triangle has a non-zero
  * area. The corner points are {@link #pl}, {@link #q} and {@link #r}. The
  * following depicts a generic triangle {@code
-                           pq
-  pv *- - - - - - - - - - - + - - - - - - - - - - -* qv
-     \~                   mpq                   ~/
-      \  ~                 |                 ~  /
-       \    ~              |              ~    /
-        \      ~           |           ~      /
-         \        ~        |        ~        /
-          \          ~     |     ~          /
-           \            ~  |  ~            /
-       -n  -n  -n  -n  -n  c  +n  +n  +n  +n  +n  normal heading out from the page.
-             \          ~  |  ~          /
-              \      ~     |     ~      /
-               \  ~        |        ~  /
-                + mrp      |      mqr +
-             rp  \         |         /  qr
-                  \        |        /
-                   \       |       /
-                    \      |      /
-                     \     |     /
-                      \    |    /
-                       \   |   /
-                        \  |  /
-                         \ | /
-                          \|/
-
-                           r
- }
+ *                          pq
+ * pv *- - - - - - - - - - - + - - - - - - - - - - -* qv
+ *     \~                   mpq                   ~/
+ *      \  ~                 |                 ~  /
+ *       \    ~              |              ~    /
+ *        \      ~           |           ~      /
+ *         \        ~        |        ~        /
+ *          \          ~     |     ~          /
+ *           \            ~  |  ~            /
+ *            \  -n  -n  -n  c  +n  +n  +n  +n  normal going out of the page.
+ *             \          ~  |  ~          /
+ *              \      ~     |     ~      /
+ *               \  ~        |        ~  /
+ *                + mrp      |      mqr +
+ *             rp  \         |         /  qr
+ *                  \        |        /
+ *                   \       |       /
+ *                    \      |      /
+ *                     \     |     /
+ *                      \    |    /
+ *                       \   |   /
+ *                        \  |  /
+ *                         \ | /
+ *                          \|/
+ *                           *
+ *                           r
+ * }
  *
  * @author Andy Turner
  * @version 1.0
@@ -73,7 +73,6 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
 //     * hand rule.
 //     */
 //    public V3D_Vector normal;
-
     /**
      * Defines one of the corners of the triangle.
      */
@@ -93,12 +92,12 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
      * The order of magnitude used for the calculation of {@link #pl}.
      */
     public int oom;
-    
+
     /**
      * The RoundingMode used for the calculation of {@link #pl}.
      */
     public RoundingMode rm;
-    
+
     /**
      * For storing the line segment from {@link #getP()} to {@link #getQ()} for
      * a given Order of Magnitude and RoundingMode precision.
@@ -200,7 +199,7 @@ public class V3D_Triangle extends V3D_FiniteGeometry implements V3D_Face {
         q = new V3D_Vector(t.q);
[…]
