### 2022-12-21 c6c63dd
Fixes for double geometry
