+++ src/main/java/uk/ac/leeds/ccg/v3d/geometry/d/V3D_TriangleDouble.java
@@ -680,11 +680,11 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
             if (r.isAligned(lsq)) {
                 return ls;
             } else {
-                return V3D_LineSegmentDouble.getGeometry(r.l.getP(), lsp);
+                return V3D_LineSegmentDouble.getGeometry(r.l.getP(), lsp, epsilon);
             }
         } else {
             if (r.isAligned(lsq)) {
-                return V3D_LineSegmentDouble.getGeometry(r.l.getP(), lsq);
+                return V3D_LineSegmentDouble.getGeometry(r.l.getP(), lsq, epsilon);
             } else {
                 throw new RuntimeException("Exception in triangle-linesegment intersection.");
             }
@@ -722,9 +722,9 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
                 V3D_PlaneDouble lippl = ls.getPPL();
                 V3D_PointDouble lp = l.getP();
                 if (lippl.isOnSameSide(lp, lsq)) {
-                    return V3D_LineSegmentDouble.getGeometry(lsp, lp);
+                    return V3D_LineSegmentDouble.getGeometry(lsp, lp, epsilon);
                 } else {
-                    return V3D_LineSegmentDouble.getGeometry(lsp, l.getQ());
+                    return V3D_LineSegmentDouble.getGeometry(lsp, l.getQ(), epsilon);
                 }
             }
         } else {
@@ -732,9 +732,9 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
                 V3D_PlaneDouble liqpl = ls.getQPL();
                 V3D_PointDouble lp = l.getP();
                 if (liqpl.isOnSameSide(lp, lsp)) {
-                    return V3D_LineSegmentDouble.getGeometry(lsq, lp);
+                    return V3D_LineSegmentDouble.getGeometry(lsq, lp, epsilon);
                 } else {
-                    return V3D_LineSegmentDouble.getGeometry(lsq, l.getQ());
+                    return V3D_LineSegmentDouble.getGeometry(lsq, l.getQ(), epsilon);
                 }
             } else {
                 return l;
@@ -764,7 +764,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
             if (liqr == null) {
                 if (lirp instanceof V3D_PointDouble lirpp) {
                     if (ptpl.isOnSameSide(lirpp, opt)) {
-                        return V3D_LineSegmentDouble.getGeometry(pt, lirpp);
+                        return V3D_LineSegmentDouble.getGeometry(pt, lirpp, epsilon);
                     } else {
                         return lirpp;
                     }
@@ -775,7 +775,7 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
                 if (lirp == null) {
                     if (liqr instanceof V3D_PointDouble liqrp) {
                         if (ptpl.isOnSameSide(liqrp, opt)) {
-                            return V3D_LineSegmentDouble.getGeometry(pt, liqrp);
+                            return V3D_LineSegmentDouble.getGeometry(pt, liqrp, epsilon);
                         } else {
                             return liqrp;
                         }
@@ -794,13 +794,13 @@ public class V3D_TriangleDouble extends V3D_FiniteGeometryDouble implements V3D_
                     if (ptpl.isOnSameSide(lirpp, opt)) {
                         if (ptpl.isOnSameSide(liqrp, opt)) {
                             //return getGeometry(pt, lirpp, liqrp);
-                            return V3D_LineSegmentDouble.getGeometry(lirpp, liqrp);
[…]
