### 2023-01-10 #4283 @arfon
@agdturner – keeping submissions open pending major updates is discouraged in the long run at JOSS, primarily because of the challenge of keeping reviewers 'on the hook'. Given that this submission has been in a hold since October, I'm willing to give this another month but if this isn't ready for reviewers again by then we'll reject and encourage resubmission as and when you're able to make these updates.
